Move bogaert and data into its own private submodule to reduce code clutter
