+++ src/legendre/mod.rs
@@ -21,6 +21,8 @@
 //! assert_abs_diff_eq!(integral, 0.0);
 //! ```
 
+mod bogaert;
+
 pub mod iterators;
 use iterators::{
     GaussLegendreIntoIter, GaussLegendreIter, GaussLegendreNodes, GaussLegendreWeights,
@@ -107,189 +109,6 @@ impl GaussLegendre {
 
 impl_node_weight_rule! {GaussLegendre, GaussLegendreNodes, GaussLegendreWeights, GaussLegendreIter, GaussLegendreIntoIter}
 
-/// This algorithm is based on an expansion of Legendre polynomials in terms of Bessel functions
-/// where for large degrees only the first terms in the expansion matter. This means that
-/// the zeros of the Legendre polynomials (the nodes in the quadrature rules) can be approximated
-/// as the zeros of an expression of Bessel functions. The weights can be computed from the nodes.
-/// The zeros of the Bessel functions are computed in terms of an expression involving only the
-/// zeros of the zero:th order Bessel function and the first order Bessel function at those zeros.
-/// For large enough degrees these expansions are accurate to within machine precision,
-/// and lookup tables with exact values are used below those degrees.
-/// For an exact derivation see the paper: `<https://doi.org/10.1137/140954969>`.
-mod bogaert {
-    use super::data::{CL, EVEN_THETA_ZEROS, EVEN_WEIGHTS, J1, JZ, ODD_THETA_ZEROS, ODD_WEIGHTS};
-    use core::{cmp::Ordering, f64::consts::PI};
-
-    /// This function computes the `k`th zero of Bessel function j_0.
-    /// # Panic
-    /// Panics if `k == 0`.
-    #[rustfmt::skip]
-    // Inlined because the function is only used in a single location
-    #[inline]
-    #[must_use]
-    fn bessel_j0_zero(k: usize) -> f64 {
-        if k > 20 {
-            let z: f64 = PI * (k as f64 - 0.25);
-            let r = 1.0 / z;
-            let r2 = r * r;
-            z + r * (0.125 + r2 * (-8.072_916_666_666_667e-2 + r2 * (0.246_028_645_833_333_34 + r2 * (-1.824_438_767_206_101 + r2 * (25.336_414_797_343_906 + r2 * (-567.644_412_135_183_4 + r2 * (18_690.476_528_232_066  + r2 * (-8.493_535_802_991_488e5 + 5.092_254_624_022_268e7 * r2))))))))
-        } else {
-            JZ[k - 1]
-        }
-    }
-
-    /// This function computes the square of Bessel function j_1
-    /// evaluated at the `k`th zero of Bessel function j_0.
-    /// # Panic
-    /// Panics if `k == 0`.
-    #[rustfmt::skip]
-    // Inlined because the function is only used in a single location
-    #[inline]
-    #[must_use]
-    fn bessel_j1_squared(k: usize) -> f64 {
-        if k > 21 {
-            let x: f64 = 1.0 / (k as f64 - 0.25);
-            let x2 = x * x;
-            x * (0.202_642_367_284_675_55 + x2 * x2 * (-3.033_804_297_112_902_7e-4 + x2 * (1.989_243_642_459_693e-4 + x2 * (-2.289_699_027_721_116_6e-4 + x2 * (4.337_107_191_307_463e-4 + x2 * (-1.236_323_497_271_754e-3 + x2 * (4.961_014_232_688_831_4e-3 + x2 * (-2.668_373_937_023_237_7e-2 + 0.185_395_398_206_345_62 * x2))))))))
-        } else {
-            J1[k - 1]
-        }
-    }
-
-    /// A struct containing a Gauss-Legendre node and its associated weight.
-    #[derive(Debug, Clone, Copy, PartialEq)]
-    pub struct NodeWeightPair {
-        node: f64,
-        weight: f64,
-    }
-
-    impl NodeWeightPair {
-        /// Compute a new `NodeWeightPair`.
-        pub fn new(n: usize, k: usize) -> Self {
-            ThetaWeightPair::new(n, k).into()
-        }
-
-        /// Destructure `self` into a tuple containing the node and weight.  
-        /// The first number is the node, and the second is the weight.
-        // Inlined because the function is trivial
-        #[inline]
-        pub const fn into_tuple(self) -> (f64, f64) {
-            (self.node, self.weight)
-        }
-    }
-
-    impl From<ThetaWeightPair> for NodeWeightPair {
-        // Inlined because the function is only used in a single location
-        #[inline]
-        fn from(value: ThetaWeightPair) -> Self {
-            Self {
-                node: value.theta.cos(),
-                weight: value.weight,
-            }
-        }
-    }
-
-    /// A Gauss-Legendre node-weight pair in theta-space.
-    #[derive(Debug, Clone, Copy, PartialEq)]
-    struct ThetaWeightPair {
-        theta: f64,
-        weight: f64,
-    }
-
-    impl ThetaWeightPair {
-        /// Compute a new ThetaWeightPair.
-        /// # Panic
-        /// Panics if `k == 0` or `k > n`.
-        #[must_use]
-        fn new(n: usize, k: usize) -> Self {
-            if n <= 100 {
-                // If n is small enough we can just use tabulated values.
-                Self::tabulated_pair(n, k - 1)
-            } else if 2 * k - 1 > n {
-                let mut p = Self::compute_pair(n, n - k + 1);
-                p.theta = PI - p.theta;
-                p
-            } else {
-                Self::compute_pair(n, k)
-            }
-        }
-
-        /// Compute a node-weight pair, with k limited to half the range
-        /// # Panic
-        /// Panics if `k == 0`.
-        #[rustfmt::skip]
-        #[must_use]
-        fn compute_pair(n: usize, k: usize) -> Self {
-            // First get the j_0 zero
-            let w: f64 = 1.0 / (n as f64 + 0.5);
-            let nu = bessel_j0_zero(k);
-            let theta = w * nu;
-            let x = theta * theta;
-
-            // Get the asymptotic j_1(nu) squared
-            let b = bessel_j1_squared(k);
-
-            // Get the Chebyshev interpolants for the nodes...
-            let sf1t = (((((-1.290_529_962_742_805_1e-12 * x + 2.407_246_858_643_301_3e-10) * x - 3.131_486_546_359_920_4e-8) * x + 2.755_731_689_620_612_4e-6) * x - 1.488_095_237_139_091_4e-4) * x + 4.166_666_666_651_934e-3) * x - 4.166_666_666_666_63e-2;
-            let sf2t = (((((2.206_394_217_818_71e-9 * x - 7.530_367_713_737_693e-8) * x + 1.619_692_594_538_362_7e-6) * x - 2.533_003_260_082_32e-5) * x + 2.821_168_860_575_604_5e-4) * x - 2.090_222_483_878_529e-3) * x + 8.159_722_217_729_322e-3;
-            let sf3t = (((((-2.970_582_253_755_262_3e-8 * x + 5.558_453_302_237_962e-7) * x - 5.677_978_413_568_331e-6) * x + 4.184_981_003_295_046e-5) * x - 2.513_952_932_839_659e-4) * x + 1.286_541_985_428_451_3e-3) * x - 4.160_121_656_202_043e-3;
-
-            // ...and weights
-            let wsf1t = ((((((((-2.209_028_610_446_166_4e-14 * x + 2.303_657_268_603_773_8e-12) * x - 1.752_577_007_354_238e-10) * x + 1.037_560_669_279_168e-8) * x - 4.639_686_475_532_213e-7) * x + 1.496_445_936_250_286_4e-5) * x - 3.262_786_595_944_122e-4) * x + 4.365_079_365_075_981e-3) * x - 3.055_555_555_555_53e-2) * x + 8.333_333_333_333_333e-2;
-            let wsf2t = (((((((3.631_174_121_526_548e-12 * x + 7.676_435_450_698_932e-11) * x - 7.129_128_572_336_422e-9) * x + 2.114_838_806_859_471_6e-7) * x - 3.818_179_186_800_454e-6) * x + 4.659_695_306_949_684e-5) * x - 4.072_971_856_113_357_5e-4) * x + 2.689_594_356_947_297e-3) * x - 1.111_111_111_112_149_2e-2;
-            let wsf3t = (((((((2.018_267_912_567_033e-9 * x - 4.386_471_225_202_067e-8) * x + 5.088_983_472_886_716e-7) * x - 3.979_333_165_191_352_5e-6) * x + 2.005_593_263_964_583_4e-5) * x - 4.228_880_592_829_212e-5) * x - 1.056_460_502_540_761_4e-4) * x - 9.479_693_089_585_773e-5) * x + 6.569_664_899_264_848e-3;
-
-            // Then refine with the expansions from the paper
-            let nu_o_sin = nu / theta.sin();
-            let b_nu_o_sin = b * nu_o_sin;
-            let w_inv_sinc = w * w * nu_o_sin;
-            let wis2 = w_inv_sinc * w_inv_sinc;
-
-            // Compute the theta-node and the weight
-            let theta = w * (nu + theta * w_inv_sinc * (sf1t + wis2 * (sf2t + wis2 * sf3t)));
-            let weight = 2.0 * w / (b_nu_o_sin + b_nu_o_sin * wis2 * (wsf1t + wis2 * (wsf2t + wis2 * wsf3t)));
-
-            Self { theta, weight }
-        }
-
-        /// Returns tabulated theta and weight values, valid for n <= 100
-        /// # Panic
-        /// Panics if `n > 100`.
-        #[must_use]
-        fn tabulated_pair(n: usize, k: usize) -> Self {
-            // Odd Legendre degree
-            let (theta, weight) = if n % 2 == 1 {
-                let n2 = (n - 1) / 2;
-                match k.cmp(&n2) {
-                    Ordering::Equal => (PI / 2.0, 2.0 / (CL[n] * CL[n])),
-                    Ordering::Less => (
-                        ODD_THETA_ZEROS[n2 - 1][n2 - k - 1],
-                        ODD_WEIGHTS[n2 - 1][n2 - k - 1],
-                    ),
-                    Ordering::Greater => (
-                        PI - ODD_THETA_ZEROS[n2 - 1][k - n2 - 1],
-                        ODD_WEIGHTS[n2 - 1][k - n2 - 1],
-                    ),
-                }
-            // Even Legendre degree
-            } else {
-                let n2 = n / 2;
-                match k.cmp(&n2) {
-                    Ordering::Less => (
-                        EVEN_THETA_ZEROS[n2 - 1][n2 - k - 1],
-                        EVEN_WEIGHTS[n2 - 1][n2 - k - 1],
-                    ),
-                    Ordering::Equal | Ordering::Greater => (
-                        PI - EVEN_THETA_ZEROS[n2 - 1][k - n2],
-                        EVEN_WEIGHTS[n2 - 1][k - n2],
-                    ),
-                }
-            };
-            Self { theta, weight }
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -316,230 +135,4 @@ mod tests {
         let other_quad = GaussLegendre::new(3);
         assert_ne!(quad, other_quad);
     }
-}
-
-#[rustfmt::skip]
-mod data {
-    /// The 20 first zeros of Bessel function j_0(x).
-    pub static JZ: [f64; 20] = [2.404_825_557_695_773, 5.520_078_110_286_311, 8.653_727_912_911_013, 11.791_534_439_014_281, 14.930_917_708_487_787, 18.071_063_967_910_924, 21.211_636_629_879_26, 24.352_471_530_749_302, 27.493_479_132_040_253, 30.634_606_468_431_976, 33.775_820_213_573_57, 36.917_098_353_664_045, 40.058_425_764_628_24, 43.199_791_713_176_73, 46.341_188_371_661_815, 49.482_609_897_397_815, 52.624_051_841_115, 55.765_510_755_019_98, 58.906_983_926_080_94, 62.048_469_190_227_166];
-    /// The 21 first values of Bessel function j_1(x)^2 where x are the zeros of Bessel function j_0(x).
-    pub static J1: [f64; 21] = [0.269_514_123_941_916_9, 0.115_780_138_582_203_69, 0.073_686_351_136_408_22, 0.054_037_573_198_116_28, 0.042_661_429_017_243_09, 0.035_242_103_490_996_1, 0.030_021_070_103_054_673, 0.026_147_391_495_308_09, 0.023_159_121_824_691_393, 0.020_783_829_122_267_86, 0.018_850_450_669_317_67, 0.017_246_157_569_665_008, 0.015_893_518_105_923_6, 0.014_737_626_096_472_19, 0.013_738_465_145_387_117, 0.012_866_181_737_615_133, 0.012_098_051_548_626_797, 0.011_416_471_224_491_609, 0.010_807_592_791_180_204, 0.010_260_372_926_280_762, 0.009_765_897_139_791_051];
-
-    // Tabulated nodes and weights
-    // The required theta values for the Legendre nodes for n <= 100
-    static EVEN_THETA_ZERO_1: [f64; 1] = [9.553_166_181_245_093e-1];
-    static EVEN_THETA_ZERO_2: [f64; 2] = [1.223_899_586_470_372_6, 5.332_956_802_491_27e-1];
-    static EVEN_THETA_ZERO_3: [f64; 3] = [1.329_852_612_388_110_3, 8.483_666_264_874_876e-1, 3.696_066_519_448_289_5e-1];
-    static EVEN_THETA_ZERO_4: [f64; 4] = [1.386_317_078_892_131_3, 1.017_455_539_490_153_4, 6.490_365_804_607_796e-1, 2.827_570_635_937_967_6e-1];
-    static EVEN_THETA_ZERO_5: [f64; 5] = [1.421_366_498_439_525, 1.122_539_327_631_709_5, 8.238_386_589_997_556e-1, 5.255_196_555_285_001e-1, 2.289_442_988_470_260_2e-1];
-    static EVEN_THETA_ZERO_6: [f64; 6] = [1.445_233_238_471_440_1, 1.194_120_375_947_706_6, 9.430_552_870_605_736e-1, 6.921_076_988_818_41e-1, 4.414_870_814_893_317_5e-1, 1.923_346_793_046_672_2e-1];
-    static EVEN_THETA_ZERO_7: [f64; 7] = [1.462_529_992_921_482, 1.246_003_586_776_677_7, 1.029_498_592_525_136_8, 8.130_407_055_389_455e-1, 5.966_877_608_172_734e-1, 3.806_189_306_666_775e-1, 1.658_171_411_523_664e-1];
-    static EVEN_THETA_ZERO_8: [f64; 8] = [1.475_640_280_808_194_3, 1.285_331_444_322_965_3, 1.095_033_401_803_444_4, 9.047_575_323_895_165e-1, 7.145_252_532_340_252e-1, 5.243_866_409_035_941e-1, 3.344_986_386_876_292e-1, 1.457_246_820_036_738_3e-1];
-    static EVEN_THETA_ZERO_9: [f64; 9] = [1.485_919_440_392_653, 1.316_167_494_718_022_6, 1.146_421_481_056_642_2, 9.766_871_104_439_833e-1, 8.069_738_930_788_195e-1, 6.373_005_058_706_191e-1, 4.677_113_145_328_286e-1, 2.983_460_782_092_325e-1, 1.299_747_364_196_768_3e-1];
-    static EVEN_THETA_ZERO_10: [f64; 10] = [1.494_194_914_310_399_6, 1.340_993_178_589_955_1, 1.187_794_926_634_098_8, 1.034_603_297_590_104_3, 8.814_230_742_890_136e-1, 7.282_625_848_696_073e-1, 5.751_385_026_314_285e-1, 4.220_907_301_111_166e-1, 2.692_452_880_289_302e-1, 1.172_969_277_059_561_4e-1];
-    static EVEN_THETA_ZERO_11: [f64; 11] = [1.501_000_399_130_816, 1.361_409_225_664_372_2, 1.221_820_208_990_359_8, 1.082_235_198_111_836_8, 9.426_568_273_796_608e-1, 8.030_892_957_063_359e-1, 6.635_400_754_448_063e-1, 5.240_242_709_487_282e-1, 3.845_781_703_583_911e-1, 2.453_165_389_983_612_8e-1, 1.068_723_357_985_26e-1];
-    static EVEN_THETA_ZERO_12: [f64; 12] = [1.506_695_545_558_101, 1.378_494_427_506_219_2, 1.250_294_703_417_273, 1.122_097_523_267_250_8, 9.939_044_422_989_455e-1, 8.657_177_770_401_081e-1, 7.375_413_075_437_536e-1, 6.093_818_382_449_566e-1, 4.812_531_951_313_686_6e-1, 3.531_886_675_690_781e-1, 2.252_936_226_353_075_8e-1, 9.814_932_949_793_685e-2];
-    static EVEN_THETA_ZERO_13: [f64; 13] = [1.511_531_546_703_289_3, 1.393_002_286_179_808, 1.274_473_959_424_494, 1.155_947_313_793_812, 1.037_423_319_077_439_1, 9.189_033_445_598_993e-1, 8.003_894_803_353_296e-1, 6.818_851_814_129_299e-1, 5.633_967_073_169_294e-1, 4.449_368_152_119_130_3e-1, 3.265_362_611_165_358_3e-1, 2.082_924_425_598_466_4e-1, 9.074_274_842_993_199e-2];
-    static EVEN_THETA_ZERO_14: [f64; 14] = [1.515_689_149_557_281, 1.405_475_003_062_348_6, 1.295_261_501_292_316, 1.185_049_147_889_021_5, 1.074_838_574_917_869_3, 9.646_306_371_285_441e-1, 8.544_265_718_392_254e-1, 7.442_282_945_111_358e-1, 6.340_389_954_584_301e-1, 5.238_644_768_825_68e-1, 4.137_165_857_369_638e-1, 3.036_239_070_914_334e-1, 1.936_769_929_947_376_2e-1, 8.437_551_461_511_597e-2];
-    static EVEN_THETA_ZERO_15: [f64; 15] = [1.519_301_729_274_526_6, 1.416_312_682_230_741_7, 1.313_324_092_045_794_7, 1.210_336_308_624_476_5, 1.107_349_759_228_459_1, 1.004_365_001_539_081, 9.013_828_087_667_156e-1, 7.984_043_170_121_236e-1, 6.954_313_000_299_367e-1, 5.924_667_257_887_386e-1, 4.895_160_050_896_97e-1, 3.865_901_987_860_505e-1, 2.837_160_095_793_467e-1, 1.809_780_449_917_272e-1, 7.884_320_726_554_946e-2];
-    static EVEN_THETA_ZERO_16: [f64; 16] = [1.522_469_852_641_529_2, 1.425_817_011_963_825_4, 1.329_164_502_391_080_7, 1.232_512_573_416_362_9, 1.135_861_522_840_293_7, 1.039_211_728_068_951_7, 9.425_636_940_046_778e-1, 8.459_181_315_837_994e-1, 7.492_760_951_181_414e-1, 6.526_392_394_594_561e-1, 5.560_103_418_005_303e-1, 4.593_944_730_762_095_6e-1, 3.628_020_075_350_028e-1, 2.662_579_994_723_86e-1, 1.698_418_454_282_15e-1, 7.399_171_309_970_96e-2];
-    static EVEN_THETA_ZERO_17: [f64; 17] = [1.525_270_780_617_194_3, 1.434_219_768_045_409_6, 1.343_169_000_217_436, 1.252_118_659_062_444_5, 1.161_068_957_629_157_7, 1.070_020_159_291_475, 9.789_726_059_789_103e-1, 8.879_267_623_988_12e-1, 7.968_832_893_748_415e-1, 7.058_431_727_509_84e-1, 6.148_079_652_926_101e-1, 5.237_802_779_694_73e-1, 4.327_648_832_448_234_3e-1, 3.417_715_500_266_717_6e-1, 2.508_238_767_288_224e-1, 1.599_966_542_668_327_7e-1, 6.970_264_809_814_095e-2];
-    static EVEN_THETA_ZERO_18: [f64; 18] = [1.527_764_849_261_740_4, 1.441_701_954_349_064_8, 1.355_639_243_522_655, 1.269_576_852_063_768_5, 1.183_514_935_851_550_7, 1.097_453_683_555_812_6, 1.011_393_333_949_027, 9.253_342_019_812_867e-1, 8.392_767_201_322_476e-1, 7.532_215_073_977_623e-1, 6.671_694_908_788_198e-1, 5.811_221_342_350_705e-1, 4.950_819_018_993_074_5e-1, 4.090_533_017_972_007_3e-1, 3.230_455_648_729_988e-1, 2.370_809_940_997_937e-1, 1.512_302_802_537_625e-1, 6.588_357_082_399_222e-2];
-    static EVEN_THETA_ZERO_19: [f64; 19] = [1.529_999_863_223_206_6, 1.448_406_982_124_841_8, 1.366_814_241_651_488_8, 1.285_221_744_143_731_7, 1.203_629_605_904_952_7, 1.122_037_965_173_752, 1.040_446_993_107_623_5, 9.588_569_097_730_896e-1, 8.772_680_085_516_152e-1, 7.956_806_951_062_013e-1, 7.140_955_526_031_66e-1, 6.325_134_568_448_222e-1, 5.509_357_927_460_005e-1, 4.693_648_943_475_422_5e-1, 3.878_050_333_015_201_6e-1, 3.062_649_591_511_897e-1, 2.247_658_146_033_686_4e-1, 1.433_746_167_818_849_6e-1, 6.246_124_541_276_674e-2];
-    static EVEN_THETA_ZERO_20: [f64; 20] = [1.532_014_188_279_762_9, 1.454_449_946_977_268_5, 1.376_885_814_601_482_6, 1.299_321_869_764_876_5, 1.221_758_200_747_475_2, 1.144_194_910_846_247_5, 1.066_632_125_552_939_8, 9.890_700_026_972_187e-1, 9.115_087_474_225_932e-1, 8.339_486_352_158_799e-1, 7.563_900_488_174_808e-1, 6.788_335_401_193_977e-1, 6.012_799_395_312_685e-1, 5.237_305_617_022_756e-1, 4.461_876_237_541_810_3e-1, 3.686_551_849_119_556_5e-1, 2.911_415_613_085_159e-1, 2.136_668_503_694_680_6e-1, 1.362_947_587_312_225e-1, 5.937_690_028_966_412e-2];
-    static EVEN_THETA_ZERO_21: [f64; 21] = [1.533_838_971_193_864_3, 1.459_924_288_056_445, 1.386_009_690_354_997, 1.312_095_239_305_276_7, 1.238_181_002_944_536, 1.164_267_059_803_796_7, 1.090_353_503_721_897_7, 1.016_440_450_472_067_4, 9.425_280_472_651_469e-1, 8.686_164_868_955_468e-1, 7.947_060_295_895_204e-1, 7.207_970_381_018_823e-1, 6.468_900_366_403_721e-1, 5.729_858_150_363_659e-1, 4.990_856_247_464_946e-1, 4.251_915_773_724_379e-1, 3.513_075_400_485_981_5e-1, 2.774_414_365_914_336e-1, 2.036_124_177_925_793_5e-1, 1.298_811_916_061_515_8e-1, 5.658_282_534_660_210_4e-2];
-    static EVEN_THETA_ZERO_22: [f64; 22] = [1.535_499_761_264_077_3, 1.464_906_652_494_521_6, 1.394_313_611_500_109_4, 1.323_720_686_538_524_3, 1.253_127_930_763_391, 1.182_535_404_796_98, 1.111_943_180_033_868_6, 1.041_351_343_083_674_3, 9.707_600_019_805_773e-1, 9.001_692_951_667_51e-1, 8.295_794_049_297_956e-1, 7.589_905_782_114_329e-1, 6.884_031_600_807_736e-1, 6.178_176_499_732_537e-1, 5.472_348_011_493_452e-1, 4.766_558_078_624_76e-1, 4.060_826_859_477_620_3e-1, 3.355_191_279_517_094e-1, 2.649_727_008_485_465_7e-1, 1.944_616_940_738_156_4e-1, 1.240_440_866_043_499_3e-1, 5.403_988_657_613_871_5e-2];
-    static EVEN_THETA_ZERO_23: [f64; 23] = [1.537_017_713_608_81, 1.469_460_505_124_226_7, 1.401_903_350_962_364_6, 1.334_346_289_590_505_4, 1.266_789_363_044_4, 1.199_232_618_763_735, 1.131_676_111_906_105_6, 1.064_119_908_394_702_6, 9.965_640_890_815_035e-1, 9.290_087_556_203_499e-1, 8.614_540_390_091_103e-1, 7.939_001_124_053_586e-1, 7.263_472_110_048_245e-1, 6.587_956_640_463_587e-1, 5.912_459_486_086_227e-1, 5.236_987_847_717_838e-1, 4.561_553_147_193_392e-1, 3.886_174_669_444_433e-1, 3.210_887_745_896_478_5e-1, 2.535_764_786_314_617e-1, 1.860_980_813_776_342_6e-1, 1.187_090_676_924_131_3e-1, 5.171_568_198_966_902e-2];
-    static EVEN_THETA_ZERO_24: [f64; 24] = [1.538_410_494_858_190_5, 1.473_638_845_472_166, 1.408_867_240_039_222_8, 1.344_095_709_533_508_7, 1.279_324_287_566_779_6, 1.214_553_011_719_529, 1.149_781_925_191_718_5, 1.085_011_078_936_665_8, 1.020_240_534_516_704_3, 9.554_703_680_422_405e-1, 8.907_006_757_608_307e-1, 8.259_315_822_134_856e-1, 7.611_632_524_946_588e-1, 6.963_959_112_887_658e-1, 6.316_298_735_371_144e-1, 5.668_655_960_010_827e-1, 5.021_037_684_870_694e-1, 4.373_454_855_522_296_3e-1, 3.725_925_956_833_896_5e-1, 3.078_484_858_841_617e-1, 2.431_200_981_264_999_3e-1, 1.784_242_126_043_536_8e-1, 1.138_140_258_514_833_1e-1, 4.958_315_373_802_413e-2];
-    static EVEN_THETA_ZERO_25: [f64; 25] = [1.539_692_973_716_708_4, 1.477_486_279_394_502_3, 1.415_279_620_944_410_3, 1.353_073_023_537_942_7, 1.290_866_514_321_281, 1.228_660_123_395_079_6, 1.166_453_885_011_658_6, 1.104_247_839_096_738, 1.042_042_033_248_543_2, 9.798_365_254_403_235e-1, 9.176_313_877_712_592e-1, 8.554_267_118_081_827e-1, 7.932_226_163_976_801e-1, 7.310_192_594_231_56e-1, 6.688_168_560_730_805e-1, 6.066_157_082_814_543e-1, 5.444_162_542_389_049e-1, 4.822_191_559_963_931e-1, 4.200_254_643_636_986_4e-1, 3.578_369_542_536_859_5e-1, 2.956_568_781_922_605_5e-1, 2.334_919_029_083_292_7e-1, 1.713_581_437_497_397_3e-1, 1.093_066_902_335_822_9e-1, 4.761_952_998_197_036e-2];
-    static EVEN_THETA_ZERO_26: [f64; 26] = [1.540_877_753_740_080_4, 1.481_040_617_373_741_4, 1.421_203_510_518_656_7, 1.361_366_453_804_322_9, 1.301_529_469_356_044_4, 1.241_692_581_525_935_6, 1.181_855_817_774_264_6, 1.122_019_209_772_750_4, 1.062_182_794_829_879_7, 1.002_346_617_783_007_4, 9.425_107_335_729_934e-1, 8.826_752_108_319_278e-1, 8.228_401_370_047_382e-1, 7.630_056_258_499_811e-1, 7.031_718_287_376_428e-1, 6.433_389_522_119_554e-1, 5.835_072_863_023_427e-1, 5.236_772_521_416_453e-1, 4.638_494_862_268_433e-1, 4.040_249_990_308_91e-1, 3.442_054_975_680_11e-1, 2.843_941_101_955_779e-1, 2.245_972_494_281_052e-1, 1.648_304_164_747_05e-1, 1.051_427_544_146_6e-1, 4.580_550_859_172_368e-2];
-    static EVEN_THETA_ZERO_27: [f64; 27] = [1.541_975_588_842_621_9, 1.484_334_121_018_556_6, 1.426_692_677_652_359, 1.369_051_275_783_071_5, 1.311_409_933_595_115, 1.253_768_670_970_438, 1.196_127_510_146_226_3, 1.138_486_476_526_912_5, 1.080_845_599_717_322, 1.023_204_914_871_722_8, 9.655_644_644_970_043e-1, 9.079_243_009_168_823e-1, 8.502_844_897_148_264e-1, 7.926_451_146_568_313e-1, 7.350_062_849_078_711e-1, 6.773_681_459_074_923e-1, 6.197_308_962_817_025e-1, 5.620_948_151_095_423e-1, 5.044_603_077_892_199e-1, 4.468_279_872_027_509e-1, 3.891_988_265_038_339e-1, 3.315_744_698_431_505_3e-1, 2.739_579_305_700_526e-1, 2.163_553_856_859_194e-1, 1.587_817_673_749_480_4e-1, 1.012_844_151_694_839_4e-1, 4.412_462_056_235_422e-2];
-    static EVEN_THETA_ZERO_28: [f64; 28] = [1.542_995_710_582_549, 1.487_394_484_904_746_7, 1.431_793_279_635_669_4, 1.376_192_108_950_239_4, 1.320_590_987_909_222_6, 1.264_989_932_881_031_5, 1.209_388_962_038_684, 1.153_788_095_965_648_1, 1.098_187_358_416_032_9, 1.042_586_777_292_403, 9.869_863_859_317_283e-1, 9.313_862_248_321_055e-1, 8.757_863_440_192_766e-1, 8.201_868_063_589_761e-1, 7.645_876_922_981_546e-1, 7.089_891_068_198_45e-1, 6.533_911_899_285_832e-1, 5.977_941_329_592_258e-1, 5.421_982_048_745_54e-1, 4.866_037_965_045_89e-1, 4.310_114_988_353_694e-1, 3.754_222_503_860_499e-1, 3.198_376_369_331_602e-1, 2.642_605_649_958_747_5e-1, 2.086_969_927_688_101e-1, 1.531_613_237_261_629e-1, 9.769_922_156_300_583e-2, 4.256_272_861_907_242_6e-2];
-    static EVEN_THETA_ZERO_29: [f64; 29] = [1.543_946_088_331_101_6, 1.490_245_617_072_432_7, 1.436_545_162_952_171_3, 1.382_844_737_841_275_6, 1.329_144_354_302_189_5, 1.275_444_025_914_442_8, 1.221_743_767_654_456_5, 1.168_043_596_353_244_5, 1.114_343_531_263_457_3, 1.060_643_594_778_787_1, 1.006_943_813_366_184_8, 9.532_442_187_977_768e-1, 8.995_448_498_101_764e-1, 8.458_457_543_830_885e-1, 7.921_469_929_325_243e-1, 7.384_486_428_849_507e-1, 6.847_508_053_901_545e-1, 6.310_536_154_445_759e-1, 5.773_572_576_394_624e-1, 5.236_619_915_567_429e-1, 4.699_681_944_935_857_3e-1, 4.162_764_370_726_534e-1, 3.625_876_255_789_86e-1, 3.089_032_914_359_211e-1, 2.552_262_416_643_531_5e-1, 2.015_622_306_384_971_8e-1, 1.479_251_692_966_708e-1, 9.435_916_010_280_74e-2, 4.110_762_866_287_674e-2];
-    static EVEN_THETA_ZERO_30: [f64; 30] = [1.544_833_637_851_665_4, 1.492_908_264_756_388_3, 1.440_982_906_138_650_8, 1.389_057_572_001_580_3, 1.337_132_272_892_735, 1.285_207_020_157_876_7, 1.233_281_826_234_298_4, 1.181_356_705_000_596_6, 1.129_431_672_204_958_9, 1.077_506_746_001_711_4, 1.025_581_947_637_229_2, 9.736_573_023_432_582e-1, 9.217_328_405_213_548e-1, 8.698_085_993_416_728e-1, 8.178_846_249_414_538e-1, 7.659_609_755_086_193e-1, 7.140_377_257_012_462e-1, 6.621_149_731_355_526e-1, 6.101_928_481_720_243e-1, 5.582_715_291_407_655e-1, 5.063_512_668_959_282e-1, 4.544_324_261_262_307e-1, 4.025_155_584_642_65e-1, 3.506_015_401_168_134e-1, 2.986_918_517_703_509_6e-1, 2.467_892_075_469_457_2e-1, 1.948_991_714_956_708e-1, 1.430_351_946_011_564e-1, 9.123_992_133_264_713e-2, 3.974_873_026_126_591_5e-2];
-    static EVEN_THETA_ZERO_31: [f64; 31] = [1.545_664_389_841_685_7, 1.495_400_520_006_868_5, 1.445_136_662_469_633_3, 1.394_872_825_707_861, 1.344_609_018_631_531_6, 1.294_345_250_782_284_1, 1.244_081_532_562_166_4, 1.193_817_875_503_760_3, 1.143_554_292_597_403, 1.093_290_798_696_378, 1.043_027_411_028_491_8, 9.927_641_498_535_134e-1, 9.425_010_393_224_361e-1, 8.922_381_086_194_002e-1, 8.419_753_935_054_036e-1, 7.917_129_384_431_113e-1, 7.414_507_995_789_215e-1, 6.911_890_490_185_72e-1, 6.409_277_811_053_988e-1, 5.906_671_218_914_769e-1, 5.404_072_438_741_682e-1, 4.901_483_897_634_233e-1, 4.398_909_124_691_514e-1, 3.896_353_458_699_818_3e-1, 3.393_825_380_385_225e-1, 2.891_339_221_891_95e-1, 2.388_921_255_071_779_8e-1, 1.886_625_339_124_777_5e-1, 1.384_581_678_870_181_7e-1, 8.832_030_722_827_103e-2, 3.847_679_847_963_677e-2];
-    static EVEN_THETA_ZERO_32: [f64; 32] = [1.546_443_627_125_265_5, 1.497_738_231_263_909_4, 1.449_032_845_902_631_4, 1.400_327_478_265_391_2, 1.351_622_135_921_669, 1.302_916_826_944_702_5, 1.254_211_560_091_483_6, 1.205_506_345_013_417, 1.156_801_192_508_980_7, 1.108_096_114_833_249_5, 1.059_391_126_084_216_6, 1.010_686_242_693_214, 9.619_814_840_575_053e-1, 9.132_768_733_691_264e-1, 8.645_724_387_181_842e-1, 8.158_682_145_859_558e-1, 7.671_642_439_014_559e-1, 7.184_605_809_290_07e-1, 6.697_572_954_095_121e-1, 6.210_544_786_425_143e-1, 5.723_522_526_623_284e-1, 5.236_507_845_164_78e-1, 4.749_503_092_950_064e-1, 4.262_511_688_770_346e-1, 3.775_538_805_043_669e-1, 3.288_592_658_750_794e-1, 2.801_687_136_893_754e-1, 2.314_847_695_998_852_6e-1, 1.828_126_524_563_463_4e-1, 1.341_649_789_468_091_2e-1, 8.558_174_883_654_483e-2, 3.728_374_374_031_613e-2];
-    static EVEN_THETA_ZERO_33: [f64; 33] = [1.547_175_997_094_614_8, 1.499_935_340_679_181_6, 1.452_694_693_272_706_3, 1.405_454_061_061_769, 1.358_213_450_511_184_3, 1.310_972_868_490_444_3, 1.263_732_322_416_537_8, 1.216_491_820_419_724_1, 1.169_251_371_540_540_2, 1.122_010_985_968_754, 1.074_770_675_338_453_4, 1.027_530_453_098_431_5, 9.802_903_349_842_006e-1, 9.330_503_396_284_544e-1, 8.858_104_893_623_263e-1, 8.385_708_112_832_335e-1, 7.913_313_387_011_139e-1, 7.440_921_131_314_511e-1, 6.968_531_870_945_337e-1, 6.496_146_281_309_019e-1, 6.023_765_246_993_705e-1, 5.551_389_950_762_09e-1, 5.079_022_012_032_895e-1, 4.606_663_710_240_283e-1, 4.134_318_360_639_670_5e-1, 3.661_990_979_414_349e-1, 3.189_689_535_781_378_7e-1, 2.717_427_498_485_402e-1, 2.245_229_557_871_702_7e-1, 1.773_146_332_323_969_3e-1, 1.301_300_193_754_780_8e-1, 8.300_791_095_077_07e-2, 3.616_244_959_900_389_4e-2];
-    static EVEN_THETA_ZERO_34: [f64; 34] = [1.547_865_604_457_777_7, 1.502_004_162_357_357_2, 1.456_142_728_021_903_7, 1.410_281_306_774_684_7, 1.364_419_904_164_498_2, 1.318_558_526_067_441, 1.272_697_178_801_115_2, 1.226_835_869_256_177_6, 1.180_974_605_051_351, 1.135_113_394_719_709, 1.089_252_247_936_466_6, 1.043_391_175_801_911_2, 9.975_301_911_979_639e-1, 9.516_693_092_438_447e-1, 9.058_085_478_865_098e-1, 8.599_479_286_766_251e-1, 8.140_874_778_035_997e-1, 7.682_272_274_981_821e-1, 7.223_672_179_660_644e-1, 6.765_075_001_043_38e-1, 6.306_481_393_987_597e-1, 5.847_892_216_487_433e-1, 5.389_308_616_059_791e-1, 4.930_732_164_176_132_6e-1, 4.472_165_073_094_733e-1, 4.013_610_560_689_044e-1, 3.555_073_496_130_768_4e-1, 3.096_561_615_434_305e-1, 2.638_087_993_597_793_6e-1, 2.179_676_599_607_749e-1, 1.721_376_573_496_165_9e-1, 1.263_306_713_881_449_7e-1, 8.058_436_603_519_718e-2, 3.510_663_068_970_053e-2];
-    static EVEN_THETA_ZERO_35: [f64; 35] = [1.548_516_088_202_564_1, 1.503_955_613_246_578, 1.459_395_145_012_190_3, 1.414_834_688_100_222_7, 1.370_274_247_295_441_4, 1.325_713_827_649_021_5, 1.281_153_434_570_536_2, 1.236_593_073_933_169, 1.192_032_752_196_711, 1.147_472_476_554_108_4, 1.102_912_255_109_027_6, 1.058_352_097_094_263_3, 1.013_792_013_144_153_1, 9.692_320_156_388_929e-1, 9.246_721_191_454_418e-1, 8.801_123_409_896_301e-1, 8.355_527_020_087_518e-1, 7.909_932_275_560_464e-1, 7.464_339_488_624_694e-1, 7.018_749_049_145_359e-1, 6.573_161_450_929_18e-1, 6.127_577_329_584_495e-1, 5.681_997_518_140_861e-1, 5.236_423_130_979_095e-1, 4.790_855_694_444_513e-1, 4.345_297_357_523_596e-1, 3.899_751_246_318_783e-1, 3.454_222_091_410_985e-1, 3.008_717_408_917_774e-1, 2.563_249_902_500_919e-1, 2.117_842_860_782_108e-1, 1.672_544_029_381_416e-1, 1.227_468_836_419_337_3e-1, 7.829_832_364_814_668e-2, 3.411_071_484_766_34e-2];
-    static EVEN_THETA_ZERO_36: [f64; 36] = [1.549_130_685_823_946, 1.505_799_405_819_664_2, 1.462_468_131_657_470_4, 1.419_136_867_330_461_4, 1.375_805_616_982_639, 1.332_474_384_976_155_3, 1.289_143_175_965_912_9, 1.245_811_994_984_327_2, 1.202_480_847_539_690_4, 1.159_149_739_732_435_7, 1.115_818_678_394_807_9, 1.072_487_671_261_111_6, 1.029_156_727_178_025_5, 9.858_258_563_677_261e-1, 9.424_950_707_611_702e-1, 8.991_643_844_255_134e-1, 8.558_338_141_192_846e-1, 8.125_033_800_232_146e-1, 7.691_731_067_161_328e-1, 7.258_430_244_984_031e-1, 6.825_131_712_172_896e-1, 6.391_835_948_321_686e-1, 5.958_543_570_955_633e-1, 5.525_255_389_612_024e-1, 5.091_972_487_450_747e-1, 4.658_696_348_260_689e-1, 4.225_429_061_321_313_6e-1, 3.792_173_666_095_906_6e-1, 3.358_934_762_285_008_5e-1, 2.925_719_658_301_625_4e-1, 2.492_540_707_015_179_5e-1, 2.059_420_554_273_186_4e-1, 1.626_405_628_266_887e-1, 1.193_608_172_622_853_8e-1, 7.613_840_464_754_681e-2, 3.316_974_474_186_059e-2];
-    static EVEN_THETA_ZERO_37: [f64; 37] = [1.549_712_287_207_882_8, 1.507_544_209_724_862_6, 1.465_376_137_339_015_7, 1.423_208_073_529_702_8, 1.381_040_021_900_765_2, 1.338_871_986_235_691_3, 1.296_703_970_558_498_6, 1.254_535_979_202_491_3, 1.212_368_016_889_501, 1.170_200_088_822_853_4, 1.128_032_200_798_161_8, 1.085_864_359_337_236_6, 1.043_696_571_852_037_5, 1.001_528_846_847_853_8, 9.593_611_941_780_779e-1, 9.171_936_253_674_232e-1, 8.750_261_540_268_988e-1, 8.328_587_963_932_301e-1, 7.906_915_720_393_252e-1, 7.485_245_048_233_193e-1, 7.063_576_241_759_075e-1, 6.641_909_668_761_97e-1, 6.220_245_795_476_037e-1, 5.798_585_222_396_646e-1, 5.376_928_736_905_555e-1, 4.955_277_392_687_366_7e-1, 4.533_632_633_323_484e-1, 4.111_996_491_651_494e-1, 3.690_371_925_202_636e-1, 3.268_763_409_876_008e-1, 2.847_178_057_580_674_3e-1, 2.425_627_889_274_157e-1, 2.004_134_942_584_602e-1, 1.582_744_399_049_656_7e-1, 1.161_565_488_818_554_6e-1, 7.409_445_176_394_482e-2, 3.227_929_535_095_246e-2];
-    static EVEN_THETA_ZERO_38: [f64; 38] = [1.550_263_480_064_160_4, 1.509_197_788_083_808_1, 1.468_132_100_566_875_7, 1.427_066_420_556_418_5, 1.386_000_751_198_713, 1.344_935_095_788_765_1, 1.303_869_457_820_298_5, 1.262_803_841_041_883, 1.221_738_249_521_212_8, 1.180_672_687_719_991_1, 1.139_607_160_582_508, 1.098_541_673_641_859, 1.057_476_233_148_907_8, 1.016_410_846_230_700_9, 9.753_455_210_872_528e-1, 9.342_802_672_387_127e-1, 8.932_150_958_393_124e-1, 8.521_500_200_807_685e-1, 8.110_850_557_169_691e-1, 7.700_202_217_553_082e-1, 7.289_555_413_804_263e-1, 6.878_910_432_074_51e-1, 6.468_267_630_110_35e-1, 6.057_627_461_556_542e-1, 5.646_990_510_834_698e-1, 5.236_357_544_389_876e-1, 4.825_729_588_028_297_4e-1, 4.415_108_047_277_878e-1, 4.004_494_901_533_595e-1, 3.593_893_030_723_592e-1, 3.183_306_793_460_978e-1, 2.772_743_115_465_352_6e-1, 2.362_213_703_174_824e-1, 1.951_740_017_836_102_2e-1, 1.541_366_059_551_230_7e-1, 1.131_198_202_589_879e-1, 7.215_736_988_593_89e-2, 3.143_540_438_351_455e-2];
-    static EVEN_THETA_ZERO_39: [f64; 39] = [1.550_786_588_415_152_4, 1.510_767_112_957_397_4, 1.470_747_641_421_583, 1.430_728_176_478_592_9, 1.390_708_720_885_325, 1.350_689_277_522_434_6, 1.310_669_849_435_604_6, 1.270_650_439_881_648_3, 1.230_631_052_380_981_6, 1.190_611_690_778_359, 1.150_592_359_314_214_4, 1.110_573_062_709_577, 1.070_553_806_268_363_3, 1.030_534_596_002_003_4, 9.905_154_387_828_985e-1, 9.504_963_425_353_942e-1, 9.104_773_164_759_499e-1, 8.704_583_714_184_727e-1, 8.304_395_201_669_024e-1, 7.904_207_780_260_52e-1, 7.504_021_634_749_075e-1, 7.103_836_990_664_584e-1, 6.703_654_126_486_746e-1, 6.303_473_390_491_956e-1, 5.903_295_224_434_432e-1, 5.503_120_197_533_818e-1, 5.102_949_056_413_983e-1, 4.702_782_800_468_414_7e-1, 4.302_622_799_152_492e-1, 3.902_470_981_180_917e-1, 3.502_330_152_869_736_5e-1, 3.102_204_561_556_976_5e-1, 2.702_100_956_292_792e-1, 2.302_030_745_053_307_3e-1, 1.902_014_842_102_915e-1, 1.502_096_126_336_221_2e-1, 1.102_378_261_690_820_9e-1, 7.031_899_075_931_525e-2, 3.063_451_333_411_226_4e-2];
-    static EVEN_THETA_ZERO_40: [f64; 40] = [1.551_283_705_347_968_2, 1.512_258_463_601_911, 1.473_233_225_313_284_7, 1.434_207_992_834_186_2, 1.395_182_768_588_723_3, 1.356_157_555_104_474_2, 1.317_132_355_046_745_8, 1.278_107_171_256_65, 1.239_082_006_794_203_2, 1.200_056_864_987_904_4, 1.161_031_749_492_588_6, 1.122_006_664_357_811_7, 1.082_981_614_109_627_3, 1.043_956_603_849_447_7, 1.004_931_639_374_79, 9.659_067_273_282_46e-1, 9.268_818_753_831_083e-1, 8.878_570_924_770_502e-1, 8.488_323_891_094_103e-1, 8.098_077_777_236_123e-1, 7.707_832_732_049_53e-1, 7.317_588_935_368_493e-1, 6.927_346_606_780_251e-1, 6.537_106_017_528_971e-1, 6.146_867_506_941_757e-1, 5.756_631_505_519_365e-1, 5.366_398_568_077_528e-1, 4.976_169_422_443_344_5e-1, 4.585_945_042_946_725e-1, 4.195_726_764_797_194e-1, 3.805_516_468_579_533_4e-1, 3.415_316_890_685_594e-1, 3.025_132_172_735_989e-1, 2.634_968_895_917_008_7e-1, 2.244_838_184_598_823_6e-1, 1.854_760_433_267_094_7e-1, 1.464_777_455_344_068_4e-1, 1.074_990_339_130_794_8e-1, 6.857_195_785_426_973e-2, 2.987_341_732_561_906_6e-2];
-    static EVEN_THETA_ZERO_41: [f64; 41] = [1.551_756_721_003_315_4, 1.513_677_510_435_354_8, 1.475_598_302_924_815, 1.437_519_100_549_654_1, 1.399_439_905_448_387, 1.361_360_719_846_430_4, 1.323_281_546_084_682_5, 1.285_202_386_651_141_5, 1.247_123_244_216_506_8, 1.209_044_121_674_894_4, 1.170_965_022_191_058_4, 1.132_885_949_255_841_4, 1.094_806_906_752_030_7, 1.056_727_899_033_393_5, 1.018_648_931_020_478_7, 9.805_700_083_178_549e-1, 9.424_911_373_589_552e-1, 9.044_123_255_867_554e-1, 8.663_335_816_813_894e-1, 8.282_549_158_498_738e-1, 7.901_763_401_989_443e-1, 7.520_978_692_204_963e-1, 7.140_195_204_316_73e-1, 6.759_413_152_305_657e-1, 6.378_632_800_575_392e-1, 5.997_854_479_978_337e-1, 5.617_078_610_344_953e-1, 5.236_305_732_820_187e-1, 4.855_536_557_378_013e-1, 4.474_772_034_530_068_4e-1, 4.094_013_466_928_585e-1, 3.713_262_689_388_439e-1, 3.332_522_371_792_479e-1, 2.951_796_555_193_184e-1, 2.571_091_661_074_227_4e-1, 2.190_418_543_971_735_5e-1, 1.809_797_103_814_301_8e-1, 1.429_268_140_230_164e-1, 1.048_930_290_780_323_4e-1, 6.690_962_797_843_65e-2, 2.914_922_224_685_900_8e-2];
-    static EVEN_THETA_ZERO_42: [f64; 42] = [1.552_207_346_590_136_3, 1.515_029_387_081_184_2, 1.477_851_430_283_928, 1.440_673_478_039_699_7, 1.403_495_532_240_969_2, 1.366_317_594_853_509, 1.329_139_667_940_348, 1.291_961_753_688_162_7, 1.254_783_854_436_838_4, 1.217_605_972_713_103, 1.180_428_111_269_301, 1.143_250_273_128_649_1, 1.106_072_461_638_634_4, 1.068_894_680_534_664, 1.031_716_934_016_664_8, 9.945_392_268_421_176e-1, 9.573_615_644_400_829e-1, 9.201_839_530_522_289e-1, 8.830_063_999_088_903e-1, 8.458_289_134_509_915e-1, 8.086_515_036_126_425e-1, 7.714_741_821_849_086e-1, 7.342_969_632_895_449e-1, 6.971_198_640_037_407e-1, 6.599_429_051_953_913e-1, 6.227_661_126_567_801e-1, 5.855_895_186_691_062e-1, 5.484_131_642_019_636e-1, 5.112_371_020_703_31e-1, 4.740_614_015_734_593e-1, 4.368_861_554_959_151_4e-1, 3.997_114_910_036_376e-1, 3.625_375_872_199_777_5e-1, 3.253_647_047_992_267e-1, 2.881_932_382_678_453e-1, 2.510_238_145_617_969e-1, 2.138_574_934_303_92e-1, 1.766_962_177_535_783_3e-1, 1.395_439_709_154_010_2e-1, 1.024_103_832_005_221_9e-1, 6.532_598_686_141_261e-2, 2.845_930_797_694_291_5e-2];
-    static EVEN_THETA_ZERO_43: [f64; 43] = [1.552_637_135_069_156, 1.516_318_752_418_798_2, 1.480_000_372_180_291_7, 1.443_681_995_989_991_8, 1.407_363_625_527_612_7, 1.371_045_262_534_953, 1.334_726_908_836_065_8, 1.298_408_566_359_386_7, 1.262_090_237_162_411_9, 1.225_771_923_459_625_4, 1.189_453_627_654_523_1, 1.153_135_352_376_772, 1.116_817_100_525_785_9, 1.080_498_875_322_336_1, 1.044_180_680_370_245, 1.007_862_519_730_785_6, 9.715_443_980_131_875e-1, 9.352_263_204_856_911e-1, 8.989_082_932_130_182e-1, 8.625_903_232_280_968e-1, 8.262_724_187_486_163e-1, 7.899_545_894_528_804e-1, 7.536_368_468_349_768e-1, 7.173_192_046_673_891e-1, 6.810_016_796_111_442e-1, 6.446_842_920_316_341e-1, 6.083_670_671_059_611e-1, 5.720_500_363_511_798e-1, 5.357_332_397_728_173e-1, 4.994_167_289_487_775_3e-1, 4.631_005_715_608_865e-1, 4.267_848_582_339_84e-1, 3.904_697_131_799_79e-1, 3.541_553_113_674_441_6e-1, 3.178_419_074_113_077_4e-1, 2.815_298_867_038_369e-1, 2.452_198_616_736_214e-1, 2.089_128_675_558_041_3e-1, 1.726_108_022_974_787_2e-1, 1.363_175_571_713_249_3e-1, 1.000_425_397_881_323e-1, 6.381_557_644_960_652e-2, 2.780_129_671_121_636e-2];
-    static EVEN_THETA_ZERO_44: [f64; 44] = [1.553_047_499_032_218_5, 1.517_549_844_221_432_6, 1.482_052_191_561_582_5, 1.446_554_542_510_861, 1.411_056_898_564_365_4, 1.375_559_261_269_981, 1.340_061_632_245_437_7, 1.304_564_013_196_950_3, 1.269_066_405_939_915_5, 1.233_568_812_422_221_3, 1.198_071_234_750_839_3, 1.162_573_675_222_508_9, 1.127_076_136_359_515_4, 1.091_578_620_951_808_8, 1.056_081_132_107_029_1, 1.020_583_673_310_438, 9.850_862_484_973_096e-1, 9.495_888_621_411_026e-1, 9.140_915_193_617_474e-1, 8.785_942_260_597_805e-1, 8.430_969_890_839_84e-1, 8.075_998_164_428_633e-1, 7.721_027_175_741_015e-1, 7.366_057_036_915_555e-1, 7.011_087_882_372_793e-1, 6.656_119_874_777_63e-1, 6.301_153_213_012_085e-1, 5.946_188_142_997_515e-1, 5.591_224_972_630_766e-1, 5.236_264_092_783_025e-1, 4.881_306_007_441_175_7e-1, 4.526_351_377_998_501e-1, 4.171_401_090_099_415e-1, 3.816_456_357_674_021e-1, 3.461_518_890_753_413e-1, 3.106_591_177_837_41e-1, 2.751_676_985_649_013e-1, 2.396_782_299_970_584e-1, 2.041_917_239_104_339_7e-1, 1.687_100_353_513_348_6e-1, 1.332_369_676_454_340_2e-1, 9.778_171_579_501_174e-2, 6.237_343_205_901_608e-2, 2.717_302_558_182_235e-2];
-    static EVEN_THETA_ZERO_45: [f64; 45] = [1.553_439_726_211_154, 1.518_726_525_682_668_7, 1.484_013_327_077_361, 1.449_300_131_698_066_5, 1.414_586_940_879_145_2, 1.379_873_756_000_009_6, 1.345_160_578_499_605_5, 1.310_447_409_892_181, 1.275_734_251_784_724_8, 1.241_021_105_896_515_5, 1.206_307_974_081_314_7, 1.171_594_858_352_843_5, 1.136_881_760_914_326_2, 1.102_168_684_193_068_7, 1.067_455_630_881_287_3, 1.032_742_603_984_709_9, 9.980_296_068_808_996e-1, 9.633_166_433_897_968e-1, 9.286_037_178_597_177e-1, 8.938_908_352_730_484e-1, 8.591_780_013_772_377e-1, 8.244_652_228_485_704e-1, 7.897_525_074_988_289e-1, 7.550_398_645_386_622e-1, 7.203_273_049_167_973e-1, 6.856_148_417_619_669e-1, 6.509_024_909_658_765e-1, 6.161_902_719_627_732e-1, 5.814_782_087_876_726e-1, 5.467_663_315_368_932e-1, 5.120_546_784_214_695e-1, 4.773_432_987_146_161_7e-1, 4.426_322_570_828_637e-1, 4.079_216_401_227_574_5e-1, 3.732_115_665_343_574e-1, 3.385_022_035_318_641e-1, 3.037_937_944_563_406e-1, 2.690_867_076_466_993e-1, 2.343_815_284_441_088_4e-1, 1.996_792_463_094_099e-1, 1.649_816_752_853_099_6e-1, 1.302_925_346_385_956_4e-1, 9.562_081_616_094_949e-2, 6.099_502_786_102_04e-2, 2.657_252_290_854_776_6e-2];
-    static EVEN_THETA_ZERO_46: [f64; 46] = [1.553_814_992_974_904_8, 1.519_852_325_907_741_9, 1.485_889_660_564_341_2, 1.451_926_998_111_647_8, 1.417_964_339_743_631, 1.384_001_686_692_846, 1.350_039_040_242_776_9, 1.316_076_401_741_232_3, 1.282_113_772_615_1, 1.248_151_154_386_817_3, 1.214_188_548_692_984_2, 1.180_225_957_305_622_4, 1.146_263_382_156_703_2, 1.112_300_825_366_699, 1.078_338_289_278_105, 1.044_375_776_495_107_6, 1.010_413_289_930_890_2, 9.764_508_328_644_781e-1, 9.424_884_090_095_589e-1, 9.085_260_225_984_488e-1, 8.745_636_784_853_451e-1, 8.406_013_822_743_46e-1, 8.066_391_404_795_569e-1, 7.726_769_607_271_702e-1, 7.387_148_520_130_368e-1, 7.047_528_250_344_497e-1, 6.707_908_926_224_333e-1, 6.368_290_703_120_276e-1, 6.028_673_771_049_33e-1, 5.689_058_365_047_911e-1, 5.349_444_779_460_832e-1, 5.009_833_388_030_908e-1, 4.670_224_672_735_823e-1, 4.330_619_266_162_572e-1, 3.991_018_015_460_701e-1, 3.651_422_081_877_256_4e-1, 3.311_833_101_314_466_4e-1, 2.972_253_454_486_352_7e-1, 2.632_686_745_061_683_7e-1, 2.293_138_699_815_081_3e-1, 1.953_618_999_343_470_8e-1, 1.614_145_391_777_897_8e-1, 1.274_754_265_555_317e-1, 9.355_335_943_686_297e-2, 5.967_622_944_002_586e-2, 2.599_798_753_052_849e-2];
-    static EVEN_THETA_ZERO_47: [f64; 47] = [1.554_174_376_112_911_6, 1.520_930_475_263_640_3, 1.487_686_575_963_027_1, 1.454_442_679_258_803_2, 1.421_198_786_221_944_2, 1.387_954_897_956_585_4, 1.354_711_015_610_582, 1.321_467_140_386_931_1, 1.288_223_273_556_309_5, 1.254_979_416_471_008_4, 1.221_735_570_580_615_9, 1.188_491_737_449_843_2, 1.155_247_918_778_991_5, 1.122_004_116_427_655_7, 1.088_760_332_442_402, 1.055_516_569_089_340_7, 1.022_272_828_892_741, 9.890_291_146_811_467e-1, 9.557_854_296_428_465e-1, 9.225_417_773_930_867e-1, 8.892_981_620_561_222e-1, 8.560_545_883_661_619e-1, 8.228_110_617_925_68e-1, 7.895_675_886_964_735e-1, 7.563_241_765_284_944e-1, 7.230_808_340_807_682e-1, 6.898_375_718_116_413e-1, 6.565_944_022_687_408e-1, 6.233_513_406_471_279e-1, 5.901_084_055_357_45e-1, 5.568_656_199_307_345e-1, 5.236_230_126_340_485e-1, 4.903_806_202_198_477e-1, 4.571_384_898_571_183e-1, 4.238_966_834_573_972e-1, 3.906_552_839_347_125_5e-1, 3.574_144_049_483_910_5e-1, 3.241_742_066_189_949e-1, 2.909_349_219_721_994e-1, 2.576_969_037_411_283_6e-1, 2.244_607_124_763_75e-1, 1.912_272_957_431_274_7e-1, 1.579_983_907_861_406_7e-1, 1.247_775_594_308_675_6e-1, 9.157_341_285_433_676e-2, 5.841_325_237_532_701e-2, 2.544_777_076_240_816_2e-2];
-    static EVEN_THETA_ZERO_48: [f64; 48] = [1.554_518_863_153_354_6, 1.521_963_936_333_782_7, 1.489_409_010_908_686_4, 1.456_854_087_820_918_6, 1.424_299_168_033_388_5, 1.391_744_252_537_595_2, 1.359_189_342_362_693, 1.326_634_438_585_269_3, 1.294_079_542_340_035, 1.261_524_654_831_669, 1.228_969_777_348_083_7, 1.196_414_911_275_444_5, 1.163_860_058_115_329, 1.131_305_219_504_506_6, 1.098_750_397_237_915, 1.066_195_593_295_557_4, 1.033_640_809_874_213, 1.001_086_049_425_085_3, 9.685_313_146_988_135e-1, 9.359_766_087_996_588e-1, 9.034_219_352_512_048e-1, 8.708_672_980_765_997e-1, 8.383_127_018_973_109e-1, 8.057_581_520_556_424e-1, 7.732_036_547_680_257e-1, 7.406_492_173_185_62e-1, 7.080_948_483_057_715e-1, 6.755_405_579_604_903e-1, 6.429_863_585_601_199e-1, 6.104_322_649_751_623e-1, 5.778_782_954_001_508e-1, 5.453_244_723_459_25e-1, 5.127_708_240_092_148e-1, 4.802_173_861_982_495_5e-1, 4.476_642_050_968_423e-1, 4.151_113_413_261_211_5e-1, 3.825_588_760_747_025_7e-1, 3.500_069_206_395_502_5e-1, 3.174_556_318_161_705e-1, 2.849_052_377_944_113e-1, 2.523_560_839_907_875_7e-1, 2.198_087_193_323_827_2e-1, 1.872_640_717_400_572_7e-1, 1.547_238_424_480_887e-1, 1.221_915_194_567_498_7e-1, 8.967_553_546_914_316e-2, 5.720_262_597_323_678e-2, 2.492_036_059_421_555e-2];
-    static EVEN_THETA_ZERO_49: [f64; 49] = [1.554_849_361_424_471, 1.522_955_431_101_933_8, 1.491_061_502_037_752, 1.459_167_575_082_262, 1.427_273_651_103_158, 1.395_379_730_992_862_2, 1.363_485_815_676_330_8, 1.331_591_906_119_453_4, 1.299_698_003_338_207_3, 1.267_804_108_408_757, 1.235_910_222_478_728_4, 1.204_016_346_779_913_6, 1.172_122_482_642_727_2, 1.140_228_631_512_787_9, 1.108_334_794_970_091_2, 1.076_440_974_751_339, 1.044_547_172_776_127, 1.012_653_391_177_865, 9.807_596_323_405_32e-1, 9.488_658_989_426_542e-1, 9.169_721_940_102_87e-1, 8.850_785_209_812_849e-1, 8.531_848_837_838_286e-1, 8.212_912_869_330_969e-1, 7.893_977_356_512_25e-1, 7.575_042_360_174_186e-1, 7.256_107_951_575_084e-1, 6.937_174_214_856_351e-1, 6.618_241_250_156_436e-1, 6.299_309_177_668_762e-1, 5.980_378_142_995_696e-1, 5.661_448_324_309_071e-1, 5.342_519_942_071_113e-1, 5.023_593_272_451_872e-1, 4.704_668_666_194_035e-1, 4.385_746_575_692_260_3e-1, 4.066_827_594_785_526e-1, 3.747_912_518_813_925_6e-1, 3.429_002_438_089_823e-1, 3.110_098_888_674_705e-1, 2.791_204_106_078_991_6e-1, 2.472_321_474_279_120_5e-1, 2.153_456_371_036_966_7e-1, 1.834_617_887_100_953e-1, 1.515_822_689_338_083_6e-1, 1.197_104_949_484_175_7e-1, 8.785_472_823_121_69e-2, 5.604_116_141_749_524_6e-2, 2.441_436_781_606_819_5e-2];
-    static EVEN_THETA_ZERO_50: [f64; 50] = [1.555_166_706_034_023_7, 1.523_907_464_890_582_4, 1.492_648_224_885_016_6, 1.461_388_986_785_839_2, 1.430_129_751_376_631_1, 1.398_870_519_462_421_8, 1.367_611_291_876_438, 1.336_352_069_487_341_2, 1.305_092_853_207_091_3, 1.273_833_643_999_595_4, 1.242_574_442_890_323_8, 1.211_315_250_977_103_2, 1.180_056_069_442_347_2, 1.148_796_899_567_022_6, 1.117_537_742_746_723_5, 1.086_278_600_510_304_4, 1.055_019_474_541_620_8, 1.023_760_366_705_069_2, 9.925_012_790_757_765e-1, 9.612_422_139_755_203e-1, 9.299_831_740_157_389e-1, 8.987_241_621_493_743e-1, 8.674_651_817_337_867e-1, 8.362_062_366_076_505e-1, 8.049_473_311_856_389e-1, 7.736_884_705_759_381e-1, 7.424_296_607_273_231e-1, 7.111_709_086_148_905e-1, 6.799_122_224_768_92e-1, 6.486_536_121_198_916e-1, 6.173_950_893_164_464e-1, 5.861_366_683_298_16e-1, 5.548_783_666_157_333e-1, 5.236_202_057_751_242e-1, 4.923_622_128_691_229_6e-1, 4.611_044_222_679_868_7e-1, 4.298_468_783_051_183e-1, 3.985_896_391_770_900_6e-1, 3.673_327_828_297_9e-1, 3.360_764_161_195_064_5e-1, 3.048_206_895_905_456_5e-1, 2.735_658_223_403_245_6e-1, 2.423_121_460_275_046_2e-1, 2.110_601_877_217_048_4e-1, 1.798_108_384_023_314_5e-1, 1.485_657_315_840_060_7e-1, 1.173_282_164_330_337_2e-1, 8.610_639_001_623_935e-2, 5.492_592_372_249_737_6e-2, 2.392_851_379_957_687_2e-2];
-
-    static ODD_THETA_ZERO_1: [f64; 1] = [6.847_192_030_022_83e-1];
-    static ODD_THETA_ZERO_2: [f64; 2] = [1.002_176_803_643_121_6, 4.366_349_492_255_221_5e-1];
-    static ODD_THETA_ZERO_3: [f64; 3] = [1.152_892_953_722_227_4, 7.354_466_143_229_52e-1, 3.204_050_902_900_62e-1];
-    static ODD_THETA_ZERO_4: [f64; 4] = [1.240_573_923_404_363_3, 9.104_740_292_261_473e-1, 5.807_869_795_060_066e-1, 2.530_224_166_119_307e-1];
-    static ODD_THETA_ZERO_5: [f64; 5] = [1.297_877_729_331_450_3, 1.025_003_226_369_574_7, 7.522_519_395_990_821e-1, 4.798_534_223_256_743e-1, 2.090_492_874_137_409_3e-1];
-    static ODD_THETA_ZERO_6: [f64; 6] = [1.338_247_676_100_454_5, 1.105_718_066_248_49, 8.732_366_099_401_63e-1, 6.408_663_264_733_868e-1, 4.088_002_373_420_212e-1, 1.780_944_581_262_765_6e-1];
-    static ODD_THETA_ZERO_7: [f64; 7] = [1.368_219_536_992_351_8, 1.165_652_065_603_030_3, 9.631_067_821_301_482e-1, 7.606_069_572_889_919e-1, 5.582_062_109_125_313e-1, 3.560_718_303_314_725e-1, 1.551_231_069_747_375_2e-1];
-    static ODD_THETA_ZERO_8: [f64; 8] = [1.391_350_647_015_287_4, 1.211_909_966_211_469_6, 1.032_480_728_417_239_5, 8.530_732_514_258_506e-1, 6.737_074_594_242_523e-1, 4.944_303_818_194_983_3e-1, 3.153_898_594_929_282_5e-1, 1.373_998_952_992_547_5e-1];
-    static ODD_THETA_ZERO_9: [f64; 9] = [1.409_742_336_767_429, 1.248_691_224_331_339_3, 1.087_646_521_650_455, 9.266_134_127_998_189e-1, 7.656_007_620_508_34e-1, 6.046_261_769_405_451e-1, 4.437_316_659_960_952e-1, 2.830_497_588_453_068e-1, 1.233_108_673_082_312_8e-1];
-    static ODD_THETA_ZERO_10: [f64; 10] = [1.424_715_475_176_742_8, 1.278_636_375_242_898_6, 1.132_561_101_012_537_6, 9.864_925_055_883_794e-1, 8.404_350_520_135_059e-1, 6.943_966_110_110_701e-1, 5.483_930_281_810_39e-1, 4.024_623_099_018_152e-1, 2.567_245_837_448_891e-1, 1.118_422_651_428_890_9e-1];
-    static ODD_THETA_ZERO_11: [f64; 11] = [1.437_141_935_303_526_4, 1.303_488_659_735_581, 1.169_837_785_762_83, 1.036_190_996_404_462_2, 9.025_507_517_347_876e-1, 7.689_210_263_823_625e-1, 6.353_089_402_976_823e-1, 5.017_289_283_414_202e-1, 3.682_157_131_008_289_6e-1, 2.348_791_589_702_580_2e-1, 1.023_252_788_872_632_5e-1];
-    static ODD_THETA_ZERO_12: [f64; 12] = [1.447_620_393_135_667, 1.324_445_197_736_386_7, 1.201_271_573_324_181_2, 1.078_100_568_411_88, 9.549_336_362_382_321e-1, 8.317_729_718_814_276e-1, 7.086_221_837_538_611e-1, 5.854_877_911_108_012e-1, 4.623_830_630_132_757_5e-1, 3.393_399_712_563_371_4e-1, 2.164_597_408_964_339_4e-1, 9.430_083_986_305_52e-2];
-    static ODD_THETA_ZERO_13: [f64; 13] = [1.456_575_541_704_196, 1.342_355_260_834_552_1, 1.228_136_043_468_909_7, 1.113_918_572_282_612, 9.997_037_539_874_954e-1, 8.854_928_869_950_8e-1, 7.712_879_690_777_517e-1, 6.570_923_167_092_416e-1, 5.429_119_513_798_658e-1, 4.287_591_577_660_783_7e-1, 3.146_635_662_674_374e-1, 2.007_190_266_590_380_7e-1, 8.744_338_280_630_3e-2];
-    static ODD_THETA_ZERO_14: [f64; 14] = [1.464_317_002_991_565_3, 1.357_838_033_080_061_7, 1.251_359_804_334_884_8, 1.144_882_777_708_662_8, 1.038_407_544_520_296_7, 9.319_349_156_915_987e-1, 8.254_660_749_671_546e-1, 7.190_028_636_037_068e-1, 6.125_483_562_383_021e-1, 5.061_081_521_563e-1, 3.996_936_914_666_952e-1, 2.933_325_857_619_473e-1, 1.871_123_137_498_062e-1, 8.151_560_650_977_882e-2];
-    static ODD_THETA_ZERO_15: [f64; 15] = [1.471_075_823_713_997_4, 1.371_355_574_944_659, 1.271_635_855_736_122_3, 1.171_916_986_981_364, 1.072_199_368_669_106_5, 9.724_835_301_003_497e-1, 8.727_702_114_891_849e-1, 7.730_605_060_747_958e-1, 6.733_561_257_504_195e-1, 5.736_599_396_529_728e-1, 4.739_771_829_190_733_7e-1, 3.743_185_619_229_329_5e-1, 2.747_099_287_638_327_7e-1, 1.752_332_025_619_508_5e-1, 7.634_046_205_384_429e-2];
-    static ODD_THETA_ZERO_16: [f64; 16] = [1.477_027_911_291_552_3, 1.383_259_682_348_271_7, 1.289_491_840_051_302_7, 1.195_724_613_675_799_5, 1.101_958_282_220_461_4, 1.008_193_204_014_774, 9.144_298_626_454_032e-1, 8.206_689_427_646_12e-1, 7.269_114_630_504_563e-1, 6.331_590_254_855_162e-1, 5.394_143_214_244_184e-1, 4.456_822_679_082_866_3e-1, 3.519_729_273_095_237e-1, 2.583_106_041_071_418e-1, 1.647_723_231_643_112_6e-1, 7.178_317_184_275_122e-2];
-    static ODD_THETA_ZERO_17: [f64; 17] = [1.482_309_554_825_692_4, 1.393_822_922_226_542_1, 1.305_336_577_335_833_6, 1.216_850_687_682_353_3, 1.128_365_453_024_608_6, 1.039_881_123_511_957_6, 9.513_980_267_579_228e-1, 8.629_166_105_524_045e-1, 7.744_375_139_383_605e-1, 6.859_616_923_374_369e-1, 5.974_906_525_247_623e-1, 5.090_269_299_866_796e-1, 4.205_751_610_647_263_7e-1, 3.321_448_379_994_943e-1, 2.437_588_931_448_049e-1, 1.554_900_095_178_924_6e-1, 6.773_932_498_157_585e-2];
-    static ODD_THETA_ZERO_18: [f64; 18] = [1.487_027_983_239_551, 1.403_259_745_496_922_2, 1.319_491_725_464_661_4, 1.235_724_047_968_681_3, 1.151_956_859_289_811_5, 1.068_190_338_689_553_5, 9.844_247_150_109_837e-1, 9.006_602_918_737_365e-1, 8.168_974_877_846_822e-1, 7.331_369_031_796_229e-1, 6.493_794_386_888_65e-1, 5.656_265_174_356_597e-1, 4.818_805_368_222_632e-1, 3.981_458_834_052_59e-1, 3.144_315_409_387_123_4e-1, 2.307_592_167_302_372e-1, 1.471_977_156_945_989_7e-1, 6.412_678_117_309_945e-2];
-    static ODD_THETA_ZERO_19: [f64; 19] = [1.491_268_718_102_344_7, 1.411_741_190_914_640_5, 1.332_213_830_951_015_4, 1.252_686_732_830_81, 1.173_160_005_794_509_3, 1.093_633_781_237_958_8, 1.014_108_223_243_148_4, 9.345_835_440_325_075e-1, 8.550_600_276_575_269e-1, 7.755_380_679_025_249e-1, 6.960_182_317_959_842e-1, 6.165_013_717_819_834e-1, 5.369_888_366_794_913e-1, 4.574_829_005_269_903e-1, 3.779_877_260_196_974e-1, 2.985_118_404_618_625e-1, 2.190_758_506_462_428e-1, 1.397_450_765_119_767_3e-1, 6.088_003_363_863_534_6e-2];
-    static ODD_THETA_ZERO_20: [f64; 20] = [1.495_100_801_651_051_4, 1.419_405_340_110_198_5, 1.343_710_008_748_627_8, 1.268_014_880_389_353, 1.192_320_038_028_903_9, 1.116_625_579_891_689_5, 1.040_931_626_310_454_9, 9.652_383_295_306_943e-1, 8.895_458_882_533_946e-1, 8.138_545_700_535_261e-1, 7.381_647_473_570_305e-1, 6.624_769_578_126_105e-1, 5.867_920_109_947_447e-1, 5.111_111_891_461_745e-1, 4.354_366_553_151_05e-1, 3.597_723_703_299_625e-1, 2.841_264_494_060_559_7e-1, 2.085_185_052_177_155e-1, 1.330_107_089_065_635_5e-1, 5.794_620_170_990_797_5e-2];
-    static ODD_THETA_ZERO_21: [f64; 21] = [1.498_580_583_401_444, 1.426_364_890_228_584_6, 1.354_149_299_629_923_4, 1.281_933_868_420_424, 1.209_718_660_626_713_4, 1.137_503_750_956_414_8, 1.065_289_229_411_734, 9.930_752_076_949_069e-1, 9.208_618_284_397_05e-1, 8.486_492_789_905_562e-1, 7.764_378_127_156_926e-1, 7.042_277_832_708_635e-1, 6.320_197_021_480_768e-1, 5.598_143_404_345_396e-1, 4.876_129_202_946_139_5e-1, 4.154_175_043_169_533_5e-1, 3.432_318_703_096_418e-1, 2.710_637_595_435_203e-1, 1.989_318_822_110_657_5e-1, 1.268_955_503_926_593_2e-1, 5.528_212_871_240_371e-2];
-    static ODD_THETA_ZERO_22: [f64; 22] = [1.501_754_508_594_837_4, 1.432_712_730_475_143_2, 1.363_671_034_069_754_2, 1.294_629_464_249_430_7, 1.225_588_071_083_248_5, 1.156_546_912_269_029_3, 1.087_506_056_298_747_7, 1.018_465_586_752_840_7, 9.494_256_083_335_85e-1, 8.803_862_556_198_168e-1, 8.113_477_061_841_625e-1, 7.423_102_009_244_499e-1, 6.732_740_767_851_639e-1, 6.042_398_217_472_142e-1, 5.352_081_720_899_523e-1, 4.661_802_954_366_277e-1, 3.971_581_629_712_621_6e-1, 3.281_453_857_685_808e-1, 2.591_493_642_052_662e-1, 1.901_879_854_885_491_8e-1, 1.213_179_541_186_130_7e-1, 5.285_224_511_635_143_5e-2];
-    static ODD_THETA_ZERO_23: [f64; 23] = [1.504_661_202_517_196_4, 1.438_526_110_541_037_2, 1.372_391_084_315_255_8, 1.306_256_159_670_932, 1.240_121_376_243_315_9, 1.173_986_779_205_849_4, 1.107_852_421_486_856_1, 1.041_718_366_715_156_8, 9.755_846_932_657_443e-1, 9.094_514_999_854_932e-1, 8.433_189_145_364_798e-1, 7.771_871_059_265_139e-1, 7.110_563_039_566_126e-1, 6.449_268_305_419_475e-1, 5.787_991_523_675_322e-1, 5.126_739_740_395_089e-1, 4.465_524_134_105_889_3e-1, 3.804_363_581_140_941_5e-1, 3.143_292_666_717_729_7e-1, 2.482_382_273_986_418_5e-1, 1.821_803_739_336_923_5e-1, 1.162_100_228_791_666_2e-1, 5.062_697_144_246_345e-2];
-    static ODD_THETA_ZERO_24: [f64; 24] = [1.507_333_049_739_684_4, 1.443_869_798_951_040_7, 1.380_406_601_553_595_6, 1.316_943_486_448_336_5, 1.253_480_485_358_734, 1.190_017_634_088_428_8, 1.126_554_974_102_287, 1.063_092_554_588_577_2, 9.996_304_352_342_33e-1, 9.361_686_900_661_624e-1, 8.727_074_129_127_596e-1, 8.092_467_253_835_331e-1, 7.457_867_888_716_805e-1, 6.823_278_231_980_089e-1, 6.188_701_366_516_796e-1, 5.554_141_765_061_554e-1, 4.919_606_183_965_743_5e-1, 4.285_105_345_527_886e-1, 3.650_657_359_209_552e-1, 3.016_295_408_979_54e-1, 2.382_087_510_453_128_6e-1, 1.748_198_074_104_535_3e-1, 1.115_148_317_291_502_1e-1, 4.858_150_828_905_664e-2];
-    static ODD_THETA_ZERO_25: [f64; 25] = [1.509_797_405_521_643_7, 1.448_798_505_784_201_9, 1.387_799_649_767_640_8, 1.326_800_860_997_572_3, 1.265_802_165_120_213_6, 1.204_803_590_828_283_7, 1.143_805_171_007_496_1, 1.082_806_944_206_958_6, 1.021_808_956_582_037_2, 9.608_112_645_303_607e-1, 8.998_139_383_584_991e-1, 8.388_170_675_106_567e-1, 7.778_207_682_214_244e-1, 7.168_251_950_382_156e-1, 6.558_305_587_295_081e-1, 5.948_371_551_492_265e-1, 5.338_454_137_827_293e-1, 4.728_559_836_463_23e-1, 4.118_698_949_811_841e-1, 3.508_888_880_839_026e-1, 2.899_161_521_835_467_7e-1, 2.289_582_244_272_697e-1, 1.680_309_071_251_71e-1, 1.071_842_976_730_454_6e-1, 4.669_490_825_917_858e-2];
-    static ODD_THETA_ZERO_26: [f64; 26] = [1.512_077_535_592_702_7, 1.453_358_762_182_399_3, 1.394_640_024_852_448_2, 1.335_921_342_914_185_2, 1.277_202_737_290_683_6, 1.218_484_231_207_691_8, 1.159_765_851_037_557_2, 1.101_047_627_365_156_2, 1.042_329_596_373_083_6, 9.836_118_016_874_52e-1, 9.248_942_968_954_766e-1, 8.661_771_490_588_063e-1, 8.074_604_437_333_369e-1, 7.487_442_923_247_565e-1, 6.900_288_431_709_55e-1, 6.313_142_987_730_108e-1, 5.726_009_435_739_572e-1, 5.138_891_906_843_944e-1, 4.551_796_645_660_731e-1, 3.964_733_566_771_858_6e-1, 3.377_719_420_068_963_6e-1, 2.790_784_903_284_342_4e-1, 2.203_992_941_938_221_2e-1, 1.617_495_649_772_923_1e-1, 1.031_775_271_253_784_8e-1, 4.494_935_602_951_386e-2];
-    static ODD_THETA_ZERO_27: [f64; 27] = [1.514_193_352_804_82, 1.457_590_393_617_468_8, 1.400_987_464_419_153_1, 1.344_384_581_184_662, 1.287_781_761_126_833_8, 1.231_179_023_218_584_2, 1.174_576_388_822_640_8, 1.117_973_882_475_943_8, 1.061_371_532_893_653_5, 1.004_769_374_285_310_8, 9.481_674_481_184_789e-1, 8.915_658_055_327_279e-1, 8.349_645_107_156_934e-1, 7.783_636_457_331_087e-1, 7.217_633_176_118_4e-1, 6.651_636_690_166_557e-1, 6.085_648_948_549_621e-1, 5.519_672_690_500_085e-1, 4.953_711_895_788_267e-1, 4.387_772_581_729_22e-1, 3.821_864_303_519_236e-1, 3.256_003_205_491_779e-1, 2.690_218_877_324_958e-1, 2.124_571_975_249_336e-1, 1.559_209_129_891_515_3e-1, 9.945_952_063_842_375e-2, 4.332_960_406_341_033e-2];
-    static ODD_THETA_ZERO_28: [f64; 28] = [1.516_162_000_094_549_3, 1.461_527_685_790_782_4, 1.406_893_396_579_229_6, 1.352_259_145_769_087, 1.297_624_947_629_923, 1.242_990_817_790_598, 1.188_356_773_715_062_2, 1.133_722_835_287_525_7, 1.079_089_025_551_156, 1.024_455_371_662_101_4, 9.698_219_061_474_76e-1, 9.151_886_685_974_01e-1, 8.605_557_079_864_861e-1, 8.059_230_859_253_163e-1, 7.512_908_813_164_714e-1, 6.966_591_971_861_112e-1, 6.420_281_709_850_566e-1, 5.873_979_906_122_764e-1, 5.327_689_202_536_826e-1, 4.781_413_438_508_069e-1, 4.235_158_420_269_503_6e-1, 3.688_933_369_002_844e-1, 3.142_753_865_947_702e-1, 2.596_648_470_121_556_4e-1, 2.050_675_726_616_484_3e-1, 1.504_977_164_639_767_8e-1, 9.600_014_792_058_155e-2, 4.182_252_607_645_932_4e-2];
-    static ODD_THETA_ZERO_29: [f64; 29] = [1.517_998_315_905_975_8, 1.465_200_315_462_026_6, 1.412_402_336_143_181, 1.359_604_389_111_228_2, 1.306_806_486_279_734_8, 1.254_008_640_622_089_2, 1.201_210_866_535_131, 1.148_413_180_281_18, 1.095_615_600_538_999_4, 1.042_818_149_105_710_5, 9.900_208_518_088_601e-1, 9.372_237_397_138_955e-1, 8.844_268_507_524_555e-1, 8.316_302_319_600_398e-1, 7.788_339_426_133_211e-1, 7.260_380_587_255_163e-1, 6.732_426_796_448_046e-1, 6.204_479_380_061_241e-1, 5.676_540_152_134_466e-1, 5.148_611_664_077_888e-1, 4.620_697_624_728_054e-1, 4.092_803_643_735_033_5e-1, 3.564_938_631_002_461e-1, 3.037_117_642_790_043_7e-1, 2.509_368_276_982_061e-1, 1.981_747_109_679_033e-1, 1.454_390_911_823_840_7e-1, 9.277_332_955_453_467e-2, 4.041_676_055_113_025_6e-2];
-    static ODD_THETA_ZERO_30: [f64; 30] = [1.519_715_208_823_086_7, 1.468_634_099_702_062_6, 1.417_553_008_469_014_7, 1.366_471_944_542_347_2, 1.315_390_917_933_947, 1.264_309_939_489_363_8, 1.213_229_021_168_654, 1.162_148_176_384_137_3, 1.111_067_420_416_500_8, 1.059_986_770_938_296_6, 1.008_906_248_685_091_7, 9.578_258_783_312_408e-1, 9.067_456_896_525_242e-1, 8.556_657_190_967_861e-1, 8.045_860_119_448_479e-1, 7.535_066_253_423_996e-1, 7.024_276_326_462_753e-1, 6.513_491_298_057_894e-1, 6.002_712_449_887_427e-1, 5.491_941_535_583_391e-1, 4.981_181_022_276_018_3e-1, 4.470_434_496_975_185e-1, 3.959_707_385_770_102e-1, 3.449_008_307_748_737e-1, 2.938_351_828_535_981e-1, 2.427_764_647_581_323_6e-1, 1.917_301_500_230_701_2e-1, 1.407_094_708_800_750_4e-1, 8.975_637_836_633_63e-2, 3.910_242_380_354_419_4e-2];
-    static ODD_THETA_ZERO_31: [f64; 31] = [1.521_323_961_422_700_5, 1.471_851_603_590_422, 1.422_379_260_986_849_5, 1.372_906_941_604_798_5, 1.323_434_653_909_308, 1.273_962_407_026_590_5, 1.224_490_210_963_055_8, 1.175_018_076_866_133_6, 1.125_546_017_342_156_3, 1.076_074_046_851_682_3, 1.026_602_182_210_094_6, 9.771_304_432_322_302e-1, 9.276_588_535_760_336e-1, 8.781_874_418_647_316e-1, 8.287_162_432_047_307e-1, 7.792_453_012_756_761e-1, 7.297_746_712_644_485e-1, 6.803_044_240_724_808e-1, 6.308_346_524_943_159e-1, 5.813_654_805_388_74e-1, 5.318_970_779_332_963e-1, 4.824_296_835_154_055_5e-1, 4.329_636_445_908_698_3e-1, 3.834_994_865_870_752e-1, 3.340_380_441_799_942e-1, 2.845_807_279_748_544_5e-1, 2.351_301_237_470_960_8e-1, 1.856_915_325_646_991e-1, 1.362_777_698_319_135e-1, 8.692_946_525_012_055e-2, 3.787_087_726_949_234e-2];
-    static ODD_THETA_ZERO_32: [f64; 32] = [1.522_834_478_472_358_6, 1.474_872_636_605_138_3, 1.426_910_807_768_284_4, 1.378_948_998_781_055_3, 1.330_987_216_841_224_7, 1.283_025_469_674_968_6, 1.235_063_765_709_223, 1.187_102_114_275_073_7, 1.139_140_525_853_183, 1.091_179_012_375_759_6, 1.043_217_587_604_604_8, 9.952_562_676_120_37e-1, 9.472_950_714_021_223e-1, 8.993_340_217_254_078e-1, 8.513_731_461_641_338e-1, 8.034_124_786_014_694e-1, 7.554_520_612_457_603e-1, 7.074_919_474_732_165e-1, 6.595_322_059_052_657e-1, 6.115_729_263_971_504e-1, 5.636_142_290_734_364e-1, 5.156_562_783_879_918e-1, 4.676_993_058_012_954e-1, 4.197_436_479_350_834e-1, 3.717_898_140_987_174e-1, 3.238_386_134_116_157e-1, 2.758_914_133_405_792e-1, 2.279_507_206_431_424_5e-1, 1.800_216_744_637_006_5e-1, 1.321_166_988_439_841_6e-1, 8.427_518_284_958_235e-2, 3.671_453_742_186_897_6e-2];
-    static ODD_THETA_ZERO_33: [f64; 33] = [1.524_255_491_013_576_8, 1.477_714_660_784_952_7, 1.431_173_841_758_652_8, 1.384_633_039_781_787, 1.338_092_261_006_965_7, 1.291_551_512_012_124_8, 1.245_010_799_937_3, 1.198_470_132_644_670_5, 1.151_929_518_909_907_2, 1.105_388_968_655_282_7, 1.058_848_493_238_442, 1.012_308_105_815_651_3, 9.657_678_218_054_127e-1, 9.192_276_594_886_802e-1, 8.726_876_407_972_168e-1, 8.261_477_923_647_281e-1, 7.796_081_469_509_05e-1, 7.330_687_454_042_533e-1, 6.865_296_394_193_01e-1, 6.399_908_954_920_467e-1, 5.934_526_007_301_574e-1, 5.469_148_716_199_144e-1, 5.003_778_676_688_562e-1, 4.538_418_134_105_091_7e-1, 4.073_070_354_279_486e-1, 3.607_740_278_788_823e-1, 3.142_435_758_510_728e-1, 2.677_170_062_389_944_7e-1, 2.211_967_514_739_567_6e-1, 1.746_877_983_807_874_3e-1, 1.282_022_028_383_48e-1, 8.177_818_680_168_765e-2, 3.562_671_947_817_428e-2];
-    static ODD_THETA_ZERO_34: [f64; 34] = [1.525_594_725_214_770_8, 1.480_393_128_432_045_7, 1.435_191_541_323_085_6, 1.389_989_968_924_959_7, 1.344_788_416_522_907_8, 1.299_586_889_746_828, 1.254_385_394_680_662, 1.209_183_937_989_395_2, 1.163_982_527_069_6, 1.118_781_170_231_154_8, 1.073_579_876_920_155, 1.028_378_657_996_412_7, 9.831_775_260_837_211e-1, 9.379_764_960_179_657e-1, 8.927_755_854_282_049e-1, 8.475_748_155_007_348e-1, 8.023_742_119_985_848e-1, 7.571_738_066_433_709e-1, 7.119_736_390_205_872e-1, 6.667_737_592_565_46e-1, 6.215_742_318_591_893e-1, 5.763_751_413_603_714e-1, 5.311_766_008_298_876e-1, 4.859_787_651_249_622e-1, 4.407_818_522_612_534e-1, 3.955_861_793_705_505e-1, 3.503_922_263_398_634e-1, 3.052_007_556_167_344_3e-1, 2.600_130_558_662_051e-1, 2.148_314_894_784_556e-1, 1.696_608_997_322_034e-1, 1.245_129_955_389_27e-1, 7.942_489_891_978_154e-2, 3.460_150_809_198_017e-2];
-    static ODD_THETA_ZERO_35: [f64; 35] = [1.526_859_042_890_589_5, 1.482_921_763_148_403_7, 1.438_984_491_795_164_6, 1.395_047_233_189_252_5, 1.351_109_991_891_878, 1.307_172_772_745_304_6, 1.263_235_580_960_969, 1.219_298_422_221_050_8, 1.175_361_302_797_916_6, 1.131_424_229_697_065_8, 1.087_487_210_830_888, 1.043_550_255_232_887_2, 9.996_133_733_253_191e-1, 9.556_765_772_578_536e-1, 9.117_398_813_415_957e-1, 8.678_033_026_125_662e-1, 8.238_668_615_732_248e-1, 7.799_305_831_824_294e-1, 7.359_944_981_977_458e-1, 6.920_586_450_266_629e-1, 6.481_230_723_279_65e-1, 6.041_878_427_445_001e-1, 5.602_530_383_870_994e-1, 5.163_187_691_099_713e-1, 4.723_851_853_891_5e-1, 4.284_524_990_953_311e-1, 3.845_210_184_454_249_7e-1, 3.405_912_098_612_419e-1, 2.966_638_144_233_703_6e-1, 2.527_400_847_124_078_3e-1, 2.088_223_170_017_057_8e-1, 1.649_152_190_599_722_8e-1, 1.210_301_722_471_160_1e-1, 7.720_326_018_898_818e-2, 3.363_364_974_516_995e-2];
-    static ODD_THETA_ZERO_36: [f64; 36] = [1.528_054_559_083_405_2, 1.485_312_794_997_097_7, 1.442_571_038_214_470_7, 1.399_829_292_522_32, 1.357_087_561_874_166_7, 1.314_345_850_454_078_8, 1.271_604_162_748_143_4, 1.228_862_503_626_297, 1.186_120_878_437_839_4, 1.143_379_293_124_832, 1.100_637_754_358_770_9, 1.057_896_269_707_576_2, 1.015_154_847_842_238, 9.724_134_987_956_584e-1, 9.296_722_342_907_947e-1, 8.869_310_681_617_368e-1, 8.441_900_169_008_687e-1, 8.014_491_003_793_524e-1, 7.587_083_428_093_936e-1, 7.159_677_740_493_625e-1, 6.732_274_314_040_502e-1, 6.304_873_621_547_357e-1, 5.877_476_271_899_241e-1, 5.450_083_063_396_327e-1, 5.022_695_064_252_395e-1, 4.595_313_737_871_712e-1, 4.167_941_144_922_007e-1, 3.740_580_283_336_802_4e-1, 3.313_235_690_067_747e-1, 2.885_914_573_933_48e-1, 2.458_629_119_584_249_4e-1, 2.031_401_664_615_301_6e-1, 1.604_278_005_405_711_5e-1, 1.177_368_858_339_244_4e-1, 7.510_252_408_650_087e-2, 3.271_846_270_775_479e-2];
-    static ODD_THETA_ZERO_37: [f64; 37] = [1.529_186_740_959_505, 1.487_577_158_293_388_8, 1.445_967_582_009_979_3, 1.404_358_015_412_336_5, 1.362_748_461_941_311_5, 1.321_138_925_227_93, 1.279_529_409_151_733_3, 1.237_919_917_907_157, 1.196_310_456_080_473, 1.154_701_028_740_456_7, 1.113_091_641_546_798, 1.071_482_300_881_451_4, 1.029_873_014_009_735_9, 9.882_637_892_802_374e-1, 9.466_546_363_756_945e-1, 9.050_455_666_314_926e-1, 8.634_365_934_447_507e-1, 8.218_277_328_062_565e-1, 7.802_190_040_012_227e-1, 7.386_104_305_454_958e-1, 6.970_020_414_556_032e-1, 6.553_938_730_008_771e-1, 6.137_859_711_661_063e-1, 5.721_783_951_857_431e-1, 5.305_712_227_365_694e-1, 4.889_645_577_740_233e-1, 4.473_585_427_277_744e-1, 4.057_533_781_735_039_3e-1, 3.641_493_559_322_687e-1, 3.225_469_176_515_179_4e-1, 2.809_467_650_889_194e-1, 2.393_500_844_055_271e-1, 1.977_590_501_629_603e-1, 1.561_781_206_604_067e-1, 1.146_180_742_271_483_3e-1, 7.311_308_274_978_66e-2, 3.185_176_130_791_400_6e-2];
-    static ODD_THETA_ZERO_38: [f64; 38] = [1.530_260_491_394_766_4, 1.489_724_658_775_115_2, 1.449_188_831_753_177_4, 1.408_653_013_220_734_9, 1.368_117_206_184_034, 1.327_581_413_807_02, 1.287_045_639_459_248_7, 1.246_509_886_770_075_5, 1.205_974_159_691_065, 1.165_438_462_569_017_7, 1.124_902_800_232_641_8, 1.084_367_178_096_737_4, 1.043_831_602_288_925_6, 1.003_296_079_805_520_6, 9.627_606_187_053_432e-1, 9.222_252_283_533_213e-1, 8.816_899_197_300_536e-1, 8.411_547_058_297_167e-1, 8.006_196_021_777_24e-1, 7.600_846_275_129_128e-1, 7.195_498_046_991_649e-1, 6.790_151_619_622_966e-1, 6.384_807_345_966_276e-1, 5.979_465_673_637_772e-1, 5.574_127_179_353_943e-1, 5.168_792_619_515_767e-1, 4.763_463_006_547_496e-1, 4.358_139_727_703_203_6e-1, 3.952_824_736_706_232e-1, 3.547_520_876_199_503_6e-1, 3.142_232_448_436_673_7e-1, 2.736_966_289_659_02e-1, 2.331_733_955_144_496_3e-1, 1.926_556_629_116_315_8e-1, 1.521_477_743_835_989_4e-1, 1.116_602_300_918_232_4e-1, 7.122_632_005_925_39e-2, 3.102_979_192_734_514e-2];
-    static ODD_THETA_ZERO_39: [f64; 39] = [1.531_280_219_945_530_9, 1.491_764_115_543_711_6, 1.452_248_016_067_723_1, 1.412_731_924_058_150_6, 1.373_215_842_151_127_1, 1.333_699_773_114_208_2, 1.294_183_719_885_94, 1.254_667_685_620_366_7, 1.215_151_673_737_978_4, 1.175_635_687_984_935, 1.136_119_732_502_868_2, 1.096_603_811_912_170_4, 1.057_087_931_412_518_6, 1.017_572_096_905_509, 9.780_563_151_458_207e-1, 9.385_405_939_294_599e-1, 8.990_249_423_306_287e-1, 8.595_093_710_029_676e-1, 8.199_938_925_669_823e-1, 7.804_785_221_142_635e-1, 7.409_632_778_721_44e-1, 7.014_481_820_920_566e-1, 6.619_332_622_550_151e-1, 6.224_185_527_349_886e-1, 5.829_040_971_371_158e-1, 5.433_899_516_536_147e-1, 5.038_761_899_947_553e-1, 4.643_629_108_305_196e-1, 4.248_502_493_722_176_6e-1, 3.853_383_960_541_811e-1, 3.458_276_279_674_768e-1, 3.063_183_644_932_167e-1, 2.668_112_720_373_341e-1, 2.273_074_770_384_765_6e-1, 1.878_090_446_069_578_5e-1, 1.483_202_086_882_449_1e-1, 1.088_512_052_741_322_6e-1, 6.943_448_689_600_674e-2, 3.024_917_865_720_923_3e-2];
-    static ODD_THETA_ZERO_40: [f64; 40] = [1.532_249_903_371_281_8, 1.493_703_482_108_998_7, 1.455_157_065_195_200_4, 1.416_610_654_869_340_2, 1.378_064_253_450_997_7, 1.339_517_863_369_794, 1.300_971_487_198_245_4, 1.262_425_127_688_525, 1.223_878_787_814_308_3, 1.185_332_470_819_113_3, 1.146_786_180_272_904_7, 1.108_239_920_139_165_7, 1.069_693_694_855_262_9, 1.031_147_509_429_735, 9.926_013_695_612_463e-1, 9.540_552_817_854_49e-1, 9.155_092_536_580_933e-1, 8.769_632_939_856_247e-1, 8.384_174_131_186_299e-1, 7.998_716_233_293_993e-1, 7.613_259_393_034_546e-1, 7.227_803_787_876_119e-1, 6.842_349_634_562_861e-1, 6.456_897_200_871_629e-1, 6.071_446_821_835_497e-1, 5.685_998_922_550_279e-1, 5.300_554_050_908_43e-1, 4.915_112_925_697_217_6e-1, 4.529_676_509_187_803e-1, 4.144_246_120_108_054_4e-1, 3.758_823_615_873_93e-1, 3.373_411_699_211_847_4e-1, 2.988_014_460_838_619_4e-1, 2.602_638_401_106_843e-1, 2.217_294_507_811_754_2e-1, 1.832_002_925_124_018_3e-1, 1.446_804_953_347_050_6e-1, 1.061_800_440_374_660_7e-1, 6.773_059_476_567_832e-2, 2.950_687_695_527_422_2e-2];
-    static ODD_THETA_ZERO_41: [f64; 41] = [1.533_173_137_460_634_5, 1.495_549_950_040_734_3, 1.457_926_766_471_341, 1.420_303_588_732_694_5, 1.382_680_418_872_520_8, 1.345_057_259_031_1, 1.307_434_111_468_679, 1.269_810_978_596_001_7, 1.232_187_863_008_871_6, 1.194_564_767_527_852, 1.156_941_695_244_461, 1.119_318_649_575_559_6, 1.081_695_634_328_068, 1.044_072_653_776_750_8, 1.006_449_712_758_602_5, 9.688_268_167_884_442e-1, 9.312_039_722_018_275e-1, 8.935_811_863_333_634e-1, 8.559_584_677_414_483e-1, 8.183_358_264_943_739e-1, 7.807_132_745_385_689e-1, 7.430_908_261_781_095e-1, 7.054_684_987_070_4e-1, 6.678_463_132_547_298e-1, 6.302_242_959_332_083e-1, 5.926_024_794_204_98e-1, 5.549_809_051_864_956e-1, 5.173_596_266_878_258e-1, 4.797_387_140_623_364e-1, 4.421_182_612_140_318_6e-1, 4.044_983_968_396_638e-1, 3.668_793_022_152_994_6e-1, 3.292_612_411_240_57e-1, 2.916_446_128_242_035e-1, 2.540_300_517_665_934_3e-1, 2.164_186_303_985_620_2e-1, 1.788_123_148_742_007_7e-1, 1.412_151_362_884_411_8e-1, 1.036_368_402_634_645_1e-1, 6.610_832_470_916_41e-2, 2.880_013_396_280_84e-2];
-    static ODD_THETA_ZERO_42: [f64; 42] = [1.534_053_181_584_449_1, 1.497_310_038_074_501, 1.460_566_897_984_002_7, 1.423_823_763_069_233, 1.387_080_635_143_548, 1.350_337_516_098_480_7, 1.313_594_407_926_723_9, 1.276_851_312_747_612_7, 1.240_108_232_835_827_3, 1.203_365_170_654_181_7, 1.166_622_128_891_556_8, 1.129_879_110_507_285, 1.093_136_118_783_624_4, 1.056_393_157_388_405_8, 1.019_650_230_450_503, 9.829_073_426_515_786e-1, 9.461_644_993_385_943e-1, 9.094_217_066_630_321e-1, 8.726_789_717_547_518e-1, 8.359_363_029_411_929e-1, 7.991_937_100_265_522e-1, 7.624_512_046_511_992e-1, 7.257_088_007_597_791e-1, 6.889_665_152_185_69e-1, 6.522_243_686_409_074e-1, 6.154_823_865_075_504e-1, 5.787_406_007_128_421e-1, 5.419_990_517_384_126e-1, 5.052_577_917_731_961e-1, 4.685_168_892_980_174e-1, 4.317_764_360_047_099_4e-1, 3.950_365_575_646_973e-1, 3.582_974_309_994_31e-1, 3.215_593_139_080_008e-1, 2.848_225_961_961_619e-1, 2.480_878_974_611_689e-1, 2.113_562_650_517_154_8e-1, 1.746_296_191_183_898e-1, 1.379_118_964_507_339e-1, 1.012_126_146_941_469_4e-1, 6.456_194_899_726_138e-2, 2.812_645_439_079_299_4e-2];
-    static ODD_THETA_ZERO_43: [f64; 43] = [1.534_892_997_139_557_3, 1.498_989_668_998_897_5, 1.463_086_343_903_285_7, 1.427_183_023_414_814_2, 1.391_279_709_144_040_5, 1.355_376_402_767_822, 1.319_473_106_048_673, 1.283_569_820_856_137_5, 1.247_666_549_190_743, 1.211_763_293_211_231_4, 1.175_860_055_265_884_2, 1.139_956_837_928_964, 1.104_053_644_043_539, 1.068_150_476_772_278_3, 1.032_247_339_658_243_7, 9.963_442_366_982_619e-1, 9.604_411_724_322_427e-1, 9.245_381_520_528_253e-1, 8.886_351_815_411_563e-1, 8.527_322_678_365_407e-1, 8.168_294_190_504_262e-1, 7.809_266_447_390_143e-1, 7.450_239_562_542_93e-1, 7.091_213_672_012_905e-1, 6.732_188_940_411_854e-1, 6.373_165_568_977_467e-1, 6.014_143_806_519_714e-1, 5.655_123_964_528_13e-1, 5.296_106_438_411_04e-1, 4.937_091_737_981_756_4e-1, 4.578_080_532_255_782e-1, 4.219_073_717_059_785_3e-1, 3.860_072_520_255_396e-1, 3.501_078_671_472_636e-1, 3.142_094_687_704_932_7e-1, 2.783_124_378_775_218_6e-1, 2.424_173_798_924_625_4e-1, 2.065_253_182_141_071_6e-1, 1.706_381_290_938_671_6e-1, 1.347_596_593_282_315_3e-1, 9.889_920_900_871_123e-2, 6.308_626_356_388_784e-2, 2.748_357_108_440_508_2e-2];
-    static ODD_THETA_ZERO_44: [f64; 44] = [1.535_695_280_838_63, 1.500_594_236_235_067_7, 1.465_493_194_350_303_8, 1.430_392_156_577_492_6, 1.395_291_124_351_096_8, 1.360_190_099_162_024_2, 1.325_089_082_574, 1.289_988_076_241_572, 1.254_887_081_930_202_7, 1.219_786_101_538_995, 1.184_685_137_126_702_2, 1.149_584_190_941_820_7, 1.114_483_265_457_749_5, 1.079_382_363_414_242_8, 1.044_281_487_866_709, 1.009_180_642_245_317_8, 9.740_798_304_264_51e-1, 9.389_790_568_197_675e-1, 9.038_783_264_751_75e-1, 8.687_776_452_153_702e-1, 8.336_770_198_015_192e-1, 7.985_764_581_422_97e-1, 7.634_759_695_602_61e-1, 7.283_755_651_349_081e-1, 6.932_752_581_495_918e-1, 6.581_750_646_810_479e-1, 6.230_750_043_877_162e-1, 5.879_751_015_798_286e-1, 5.528_753_866_962_97e-1, 5.177_758_983_811_02e-1, 4.826_766_864_637_186_4e-1, 4.475_778_163_386_701_6e-1, 4.124_793_755_752_883_5e-1, 3.773_814_842_049_053_6e-1, 3.422_843_113_148_581_5e-1, 3.071_881_029_697_498e-1, 2.720_932_316_284_943e-1, 2.370_002_891_767_127_6e-1, 2.019_102_761_348_422e-1, 1.668_250_268_181_993e-1, 1.317_483_020_532_982_7e-1, 9.668_919_410_176_594e-2, 6.167_652_949_817_792e-2, 2.686_941_953_400_763e-2];
-    static ODD_THETA_ZERO_45: [f64; 45] = [1.536_462_493_634_653_6, 1.502_128_661_685_489_4, 1.467_794_832_169_950_3, 1.433_461_006_333_747_4, 1.399_127_185_457_927_2, 1.364_793_370_871_767_4, 1.330_459_563_966_682_5, 1.296_125_766_211_457, 1.261_791_979_169_174_5, 1.227_458_204_516_276_3, 1.193_124_444_064_268_8, 1.158_790_699_784_705_5, 1.124_456_973_838_220_9, 1.090_123_268_608_563_3, 1.055_789_586_742_829, 1.021_455_931_199_402_3, 9.871_223_053_055_24e-1, 9.527_887_128_269_591e-1, 9.184_551_580_529_615e-1, 8.841_216_459_007_313e-1, 8.497_881_820_448_998e-1, 8.154_547_730_794_464e-1, 7.811_214_267_220_411e-1, 7.467_881_520_744_806e-1, 7.124_549_599_581_423e-1, 6.781_218_633_510_391e-1, 6.437_888_779_643_722e-1, 6.094_560_230_135_453e-1, 5.751_233_222_647_906e-1, 5.407_908_054_797_11e-1, 5.064_585_104_462_233e-1, 4.721_264_858_937_837e-1, 4.377_947_957_771_643e-1, 4.034_635_257_416_918_6e-1, 3.691_327_931_855_416_5e-1, 3.348_027_634_909_946_4e-1, 3.004_736_773_353_657_3e-1, 2.661_458_990_278_704e-1, 2.318_200_075_085_118e-1, 1.974_969_814_205_034_5e-1, 1.631_786_149_772_797e-1, 1.288_685_867_945_150_2e-1, 9.457_579_039_019_365e-2, 6.032_842_220_945_916_5e-2, 2.628_211_572_883_546e-2];
-    static ODD_THETA_ZERO_46: [f64; 46] = [1.537_196_885_933_572_3, 1.503_597_446_159_129_7, 1.469_998_008_568_304_2, 1.436_398_574_277_729_4, 1.402_799_144_434_368_4, 1.369_199_720_226_542_2, 1.335_600_302_895_785_7, 1.302_000_893_749_787_8, 1.268_401_494_176_718_3, 1.234_802_105_661_283, 1.201_202_729_802_928_6, 1.167_603_368_336_689, 1.134_004_023_157_288_5, 1.100_404_696_347_243_4, 1.066_805_390_209_896, 1.033_206_107_308_545_7, 9.996_068_505_131_473e-1, 9.660_076_230_564_559e-1, 9.324_084_286_020_319e-1, 8.988_092_713_272_343e-1, 8.652_101_560_253_048e-1, 8.316_110_882_319_595e-1, 7.980_120_743_837_287e-1, 7.644_131_220_178_279e-1, 7.308_142_400_269_309e-1, 6.972_154_389_873_656e-1, 6.636_167_315_867_435e-1, 6.300_181_331_881_123e-1, 5.964_196_625_844_131e-1, 5.628_213_430_226_633e-1, 5.292_232_036_175_456e-1, 4.956_252_813_388_603e-1, 4.620_276_238_643_497e-1, 4.284_302_937_718_012_5e-1, 3.948_333_748_659_561_5e-1, 3.612_369_820_255_325e-1, 3.276_412_770_872_601e-1, 2.940_464_955_725_917e-1, 2.604_529_939_906_063e-1, 2.268_613_388_903_867_3e-1, 1.932_724_879_746_856_9e-1, 1.596_881_970_714_452_4e-1, 1.261_120_661_032_951_7e-1, 9.255_279_834_764_232e-2, 5.903_798_711_627_596e-2, 2.571_993_685_288_741_2e-2];
-    static ODD_THETA_ZERO_47: [f64; 47] = [1.537_900_519_639_177_3, 1.505_004_713_461_118_9, 1.472_108_909_246_876_7, 1.439_213_107_999_753_9, 1.406_317_310_749_171_8, 1.373_421_518_560_135_7, 1.340_525_732_543_379, 1.307_629_953_866_4, 1.274_734_183_765_634_5, 1.241_838_423_560_042_4, 1.208_942_674_666_441_6, 1.176_046_938_616_99, 1.143_151_217_079_297, 1.110_255_511_879_752_1, 1.077_359_825_030_803_2, 1.044_464_158_763_086_5, 1.011_568_515_563_551, 9.786_728_982_210_094e-1, 9.457_773_098_809_58e-1, 9.128_817_541_120_208e-1, 8.799_862_349_870_846e-1, 8.470_907_571_831_348e-1, 8.141_953_261_050_969e-1, 7.812_999_480_407_721e-1, 7.484_046_303_564_402e-1, 7.155_093_817_462_244e-1, 6.826_142_125_533_466e-1, 6.497_191_351_887_404e-1, 6.168_241_646_833_333e-1, 5.839_293_194_266_533e-1, 5.510_346_221_695_15e-1, 5.181_401_014_079_634e-1, 4.852_457_933_290_632_6e-1, 4.523_517_446_039_431e-1, 4.194_580_164_920_722_6e-1, 3.865_646_910_356_375e-1, 3.536_718_807_003_19e-1, 3.207_797_439_266_498e-1, 2.878_885_112_969_848_4e-1, 2.549_985_318_477_515_6e-1, 2.221_103_602_568_508_3e-1, 1.892_249_341_643_785_2e-1, 1.563_439_726_212_394_7e-1, 1.234_710_001_537_068_2e-1, 9.061_453_776_736_62e-2, 5.780_160_090_309_369e-2, 2.518_130_440_638_251_6e-2];
-    static ODD_THETA_ZERO_48: [f64; 48] = [1.538_575_287_485_045_8, 1.506_354_249_056_545_8, 1.474_133_212_398_093_5, 1.441_912_178_413_208_5, 1.409_691_148_027_974, 1.377_470_122_199_186_3, 1.345_249_101_923_067_2, 1.313_028_088_244_711_4, 1.280_807_082_268_469_4, 1.248_586_085_169_490_5, 1.216_365_098_206_699_1, 1.184_144_122_737_519, 1.151_923_160_234_735_7, 1.119_702_212_305_964_1, 1.087_481_280_716_290_8, 1.055_260_367_414_81, 1.023_039_474_565_930_1, 9.908_186_045_865_675e-1, 9.585_977_601_906_32e-1, 9.263_769_444_426_037e-1, 8.941_561_608_225_062e-1, 8.619_354_133_052_818e-1, 8.297_147_064_584_917e-1, 7.974_940_455_635_383e-1, 7.652_734_367_673_509e-1, 7.330_528_872_739_118e-1, 7.008_324_055_884_452e-1, 6.686_120_018_320_298e-1, 6.363_916_881_515_75e-1, 6.041_714_792_607_29e-1, 5.719_513_931_632_927e-1, 5.397_314_521_353e-1, 5.075_116_840_805_377e-1, 4.752_921_244_363_892e-1, 4.430_728_189_095_547_4e-1, 4.108_538_274_961_112_7e-1, 3.786_352_305_487_998_3e-1, 3.464_171_382_200_184_6e-1, 3.141_997_056_941_599e-1, 2.819_831_588_178_046_7e-1, 2.497_678_394_619_649_3e-1, 2.175_542_909_210_22e-1, 1.853_434_315_961_135_8e-1, 1.531_369_452_704_970_4e-1, 1.209_382_841_678_252_6e-1, 8.875_579_450_016_283e-2, 5.661_593_754_525_191e-2, 2.466_476_940_450_737e-2];
-    static ODD_THETA_ZERO_49: [f64; 49] = [1.539_222_930_035_210_3, 1.507_649_534_071_73, 1.476_076_139_707_032_6, 1.444_502_747_756_546_6, 1.412_929_359_055_252_5, 1.381_355_974_464_721_5, 1.349_782_594_880_622_8, 1.318_209_221_240_839_4, 1.286_635_854_534_357_5, 1.255_062_495_811_113, 1.223_489_146_193_015_5, 1.191_915_806_886_406, 1.160_342_479_196_260_4, 1.128_769_164_542_51, 1.097_195_864_478_936_6, 1.065_622_580_715_200_7, 1.034_049_315_142_698_6, 1.002_476_069_865_111, 9.709_028_472_347_33e-1, 9.393_296_498_959_608e-1, 9.077_564_808_376_97e-1, 8.761_833_434_569_334e-1, 8.446_102_416_364_528e-1, 8.130_371_798_404_961e-1, 7.814_641_632_334_84e-1, 7.498_911_978_285_965e-1, 7.183_182_906_753_955e-1, 6.867_454_500_990_592e-1, 6.551_726_860_086_247e-1, 6.236_000_102_986_843e-1, 5.920_274_373_793_84e-1, 5.604_549_848_852_622e-1, 5.288_826_746_375_584e-1, 4.973_105_339_724_572e-1, 4.657_385_976_085_971e-1, 4.341_669_103_277_771e-1, 4.025_955_309_141_879e-1, 3.710_245_380_997_234e-1, 3.394_540_398_171_456e-1, 3.078_841_881_262_277_5e-1, 2.763_152_043_287_541e-1, 2.447_474_234_189_502_6e-1, 2.131_813_777_658_572e-1, 1.816_179_673_056_091e-1, 1.500_588_419_721_174_4e-1, 1.185_073_845_935_281_6e-1, 8.697_177_361_567_243e-2, 5.547_793_843_128_156_3e-2, 2.416_899_936_118_312e-2];
-
-    /// The nodes in theta-space for odd order
-    pub static ODD_THETA_ZEROS: [&[f64]; 49] = [&ODD_THETA_ZERO_1, &ODD_THETA_ZERO_2, &ODD_THETA_ZERO_3, &ODD_THETA_ZERO_4, &ODD_THETA_ZERO_5, &ODD_THETA_ZERO_6, &ODD_THETA_ZERO_7, &ODD_THETA_ZERO_8, &ODD_THETA_ZERO_9, &ODD_THETA_ZERO_10, &ODD_THETA_ZERO_11, &ODD_THETA_ZERO_12, &ODD_THETA_ZERO_13, &ODD_THETA_ZERO_14, &ODD_THETA_ZERO_15, &ODD_THETA_ZERO_16, &ODD_THETA_ZERO_17, &ODD_THETA_ZERO_18, &ODD_THETA_ZERO_19, &ODD_THETA_ZERO_20, &ODD_THETA_ZERO_21, &ODD_THETA_ZERO_22, &ODD_THETA_ZERO_23, &ODD_THETA_ZERO_24, &ODD_THETA_ZERO_25, &ODD_THETA_ZERO_26, &ODD_THETA_ZERO_27, &ODD_THETA_ZERO_28, &ODD_THETA_ZERO_29, &ODD_THETA_ZERO_30, &ODD_THETA_ZERO_31, &ODD_THETA_ZERO_32, &ODD_THETA_ZERO_33, &ODD_THETA_ZERO_34, &ODD_THETA_ZERO_35, &ODD_THETA_ZERO_36, &ODD_THETA_ZERO_37, &ODD_THETA_ZERO_38, &ODD_THETA_ZERO_39, &ODD_THETA_ZERO_40, &ODD_THETA_ZERO_41, &ODD_THETA_ZERO_42, &ODD_THETA_ZERO_43, &ODD_THETA_ZERO_44, &ODD_THETA_ZERO_45, &ODD_THETA_ZERO_46, &ODD_THETA_ZERO_47, &ODD_THETA_ZERO_48, &ODD_THETA_ZERO_49];
-    /// The nodes in theta-space for even order
-    pub static EVEN_THETA_ZEROS: [&[f64]; 50] = [&EVEN_THETA_ZERO_1, &EVEN_THETA_ZERO_2, &EVEN_THETA_ZERO_3, &EVEN_THETA_ZERO_4, &EVEN_THETA_ZERO_5, &EVEN_THETA_ZERO_6, &EVEN_THETA_ZERO_7, &EVEN_THETA_ZERO_8, &EVEN_THETA_ZERO_9, &EVEN_THETA_ZERO_10, &EVEN_THETA_ZERO_11, &EVEN_THETA_ZERO_12, &EVEN_THETA_ZERO_13, &EVEN_THETA_ZERO_14, &EVEN_THETA_ZERO_15, &EVEN_THETA_ZERO_16, &EVEN_THETA_ZERO_17, &EVEN_THETA_ZERO_18, &EVEN_THETA_ZERO_19, &EVEN_THETA_ZERO_20, &EVEN_THETA_ZERO_21, &EVEN_THETA_ZERO_22, &EVEN_THETA_ZERO_23, &EVEN_THETA_ZERO_24, &EVEN_THETA_ZERO_25, &EVEN_THETA_ZERO_26, &EVEN_THETA_ZERO_27, &EVEN_THETA_ZERO_28, &EVEN_THETA_ZERO_29, &EVEN_THETA_ZERO_30, &EVEN_THETA_ZERO_31, &EVEN_THETA_ZERO_32, &EVEN_THETA_ZERO_33, &EVEN_THETA_ZERO_34, &EVEN_THETA_ZERO_35, &EVEN_THETA_ZERO_36, &EVEN_THETA_ZERO_37, &EVEN_THETA_ZERO_38, &EVEN_THETA_ZERO_39, &EVEN_THETA_ZERO_40, &EVEN_THETA_ZERO_41, &EVEN_THETA_ZERO_42, &EVEN_THETA_ZERO_43, &EVEN_THETA_ZERO_44, &EVEN_THETA_ZERO_45, &EVEN_THETA_ZERO_46, &EVEN_THETA_ZERO_47, &EVEN_THETA_ZERO_48, &EVEN_THETA_ZERO_49, &EVEN_THETA_ZERO_50];
-
-    // The weights of the nodes
-    static EVEN_W_1: [f64; 1] = [1.0];
-    static EVEN_W_2: [f64; 2] = [6.521_451_548_625_461e-1, 3.478_548_451_374_538_5e-1];
-    static EVEN_W_3: [f64; 3] = [4.679_139_345_726_910_4e-1, 3.607_615_730_481_386e-1, 1.713_244_923_791_703_6e-1];
-    static EVEN_W_4: [f64; 4] = [3.626_837_833_783_62e-1, 3.137_066_458_778_872_7e-1, 2.223_810_344_533_744_8e-1, 1.012_285_362_903_762_6e-1];
-    static EVEN_W_5: [f64; 5] = [2.955_242_247_147_528_7e-1, 2.692_667_193_099_963_5e-1, 2.190_863_625_159_820_4e-1, 1.494_513_491_505_806e-1, 6.667_134_430_868_814e-2];
-    static EVEN_W_6: [f64; 6] = [2.491_470_458_134_027_7e-1, 2.334_925_365_383_548e-1, 2.031_674_267_230_659_2e-1, 1.600_783_285_433_462_2e-1, 1.069_393_259_953_184_3e-1, 4.717_533_638_651_183e-2];
-    static EVEN_W_7: [f64; 7] = [2.152_638_534_631_578e-1, 2.051_984_637_212_956e-1, 1.855_383_974_779_378_2e-1, 1.572_031_671_581_935_5e-1, 1.215_185_706_879_031_9e-1, 8.015_808_715_976_021e-2, 3.511_946_033_175_186e-2];
-    static EVEN_W_8: [f64; 8] = [1.894_506_104_550_685e-1, 1.826_034_150_449_235_8e-1, 1.691_565_193_950_025_4e-1, 1.495_959_888_165_767_4e-1, 1.246_289_712_555_338_8e-1, 9.515_851_168_249_279e-2, 6.225_352_393_864_789_4e-2, 2.715_245_941_175_409_6e-2];
-    static EVEN_W_9: [f64; 9] = [1.691_423_829_631_436e-1, 1.642_764_837_458_327_3e-1, 1.546_846_751_262_652_4e-1, 1.406_429_146_706_506_5e-1, 1.225_552_067_114_784_6e-1, 1.009_420_441_062_871_7e-1, 7.642_573_025_488_905e-2, 4.971_454_889_496_98e-2, 2.161_601_352_648_331e-2];
-    static EVEN_W_10: [f64; 10] = [1.527_533_871_307_258_4e-1, 1.491_729_864_726_037_4e-1, 1.420_961_093_183_820_4e-1, 1.316_886_384_491_766_4e-1, 1.181_945_319_615_184_1e-1, 1.019_301_198_172_404_4e-1, 8.327_674_157_670_475e-2, 6.267_204_833_410_907e-2, 4.060_142_980_038_694e-2, 1.761_400_713_915_211_8e-2];
-    static EVEN_W_11: [f64; 11] = [1.392_518_728_556_319_8e-1, 1.365_414_983_460_151_7e-1, 1.311_735_047_870_623_8e-1, 1.232_523_768_105_124_2e-1, 1.129_322_960_805_392_2e-1, 1.004_141_444_428_809_6e-1, 8.594_160_621_706_773e-2, 6.979_646_842_452_049e-2, 5.229_333_515_268_328_6e-2, 3.377_490_158_481_415e-2, 1.462_799_529_827_22e-2];
-    static EVEN_W_12: [f64; 12] = [1.279_381_953_467_521_6e-1, 1.258_374_563_468_283e-1, 1.216_704_729_278_033_9e-1, 1.155_056_680_537_256e-1, 1.074_442_701_159_656_3e-1, 9.761_865_210_411_388e-2, 8.619_016_153_195_327e-2, 7.334_648_141_108_03e-2, 5.929_858_491_543_678e-2, 4.427_743_881_741_981e-2, 2.853_138_862_893_366_3e-2, 1.234_122_979_998_72e-2];
-    static EVEN_W_13: [f64; 13] = [1.183_214_152_792_622_8e-1, 1.166_604_434_852_965_8e-1, 1.133_618_165_463_196_6e-1, 1.084_718_405_285_765_9e-1, 1.020_591_610_944_254_2e-1, 9.421_380_035_591_415e-2, 8.504_589_431_348_523e-2, 7.468_414_976_565_975e-2, 6.327_404_632_957_484e-2, 5.097_582_529_714_781e-2, 3.796_238_329_436_276_6e-2, 2.441_785_109_263_191e-2, 1.055_137_261_734_300_6e-2];
-    static EVEN_W_14: [f64; 14] = [1.100_470_130_164_752e-1, 1.087_111_922_582_941_3e-1, 1.060_557_659_228_464_2e-1, 1.021_129_675_780_607_6e-1, 9.693_065_799_792_992e-2, 9.057_174_439_303_284e-2, 8.311_341_722_890_121e-2, 7.464_621_423_456_878e-2, 6.527_292_396_699_96e-2, 5.510_734_567_571_675e-2, 4.427_293_475_900_423e-2, 3.290_142_778_230_438e-2, 2.113_211_259_277_126e-2, 9.124_282_593_094_517e-3];
-    static EVEN_W_15: [f64; 15] = [1.028_526_528_935_588_4e-1, 1.017_623_897_484_055e-1, 9.959_342_058_679_527e-2, 9.636_873_717_464_425e-2, 9.212_252_223_778_612e-2, 8.689_978_720_108_298e-2, 8.075_589_522_942_021e-2, 7.375_597_473_770_52e-2, 6.597_422_988_218_049e-2, 5.749_315_621_761_906_5e-2, 4.840_267_283_059_405e-2, 3.879_919_256_962_705e-2, 2.878_470_788_332_337e-2, 1.846_646_831_109_096e-2, 7.968_192_496_166_605e-3];
-    static EVEN_W_16: [f64; 16] = [9.654_008_851_472_78e-2, 9.563_872_007_927_486e-2, 9.384_439_908_080_457e-2, 9.117_387_869_576_389e-2, 8.765_209_300_440_381e-2, 8.331_192_422_694_675e-2, 7.819_389_578_707_031e-2, 7.234_579_410_884_85e-2, 6.582_222_277_636_185e-2, 5.868_409_347_853_554_4e-2, 5.099_805_926_237_617_5e-2, 4.283_589_802_222_668e-2, 3.427_386_291_302_143e-2, 2.539_206_530_926_206e-2, 1.627_439_473_090_567e-2, 7.018_610_009_470_096e-3];
-    static EVEN_W_17: [f64; 17] = [9.095_674_033_025_987e-2, 9.020_304_437_064_074e-2, 8.870_189_783_569_386e-2, 8.646_573_974_703_575e-2, 8.351_309_969_984_566e-2, 7.986_844_433_977_185e-2, 7.556_197_466_003_194e-2, 7.062_937_581_425_573e-2, 6.511_152_155_407_642e-2, 5.905_413_582_752_449_4e-2, 5.250_741_457_267_811e-2, 4.552_561_152_335_327e-2, 3.816_659_379_638_752e-2, 3.049_138_063_844_613e-2, 2.256_372_198_549_497e-2, 1.445_016_274_859_503_6e-2, 6.229_140_555_908_685e-3];
-    static EVEN_W_18: [f64; 18] = [8.598_327_567_039_475e-2, 8.534_668_573_933_862e-2, 8.407_821_897_966_193e-2, 8.218_726_670_433_97e-2, 7.968_782_891_207_16e-2, 7.659_841_064_587_067e-2, 7.294_188_500_565_306e-2, 6.874_532_383_573_645e-2, 6.403_979_735_501_548e-2, 5.886_014_424_532_482e-2, 5.324_471_397_775_992e-2, 4.723_508_349_026_598e-2, 4.087_575_092_364_489e-2, 3.421_381_077_030_723e-2, 2.729_862_149_856_878e-2, 2.018_151_529_773_547_2e-2, 1.291_594_728_406_557_4e-2, 5.565_719_664_245_045_5e-3];
-    static EVEN_W_19: [f64; 19] = [8.152_502_928_038_578e-2, 8.098_249_377_059_71e-2, 7.990_103_324_352_782e-2, 7.828_784_465_821_095e-2, 7.615_366_354_844_64e-2, 7.351_269_258_474_345e-2, 7.038_250_706_689_896e-2, 6.678_393_797_914_041e-2, 6.274_093_339_213_306e-2, 5.828_039_914_699_72e-2, 5.343_201_991_033_232e-2, 4.822_806_186_075_868e-2, 4.270_315_850_467_443e-2, 3.689_408_159_402_474e-2, 3.083_950_054_517_505_3e-2, 2.457_973_973_823_237_4e-2, 1.815_657_770_961_323_6e-2, 1.161_344_471_646_867_5e-2, 5.002_880_749_639_346e-3];
-    static EVEN_W_20: [f64; 20] = [7.750_594_797_842_48e-2, 7.703_981_816_424_797e-2, 7.611_036_190_062_624e-2, 7.472_316_905_796_826e-2, 7.288_658_239_580_406e-2, 7.061_164_739_128_678e-2, 6.791_204_581_523_39e-2, 6.480_401_345_660_104e-2, 6.130_624_249_292_894e-2, 5.743_976_909_939_155e-2, 5.322_784_698_393_682e-2, 4.869_580_763_507_223e-2, 4.387_090_818_567_327e-2, 3.878_216_797_447_201_6e-2, 3.346_019_528_254_784_4e-2, 2.793_700_698_002_34e-2, 2.224_584_919_416_696e-2, 1.642_105_838_190_789e-2, 1.049_828_453_115_281_3e-2, 4.521_277_098_533_191e-3];
-    static EVEN_W_21: [f64; 21] = [7.386_423_423_217_288e-2, 7.346_081_345_346_753e-2, 7.265_617_524_380_41e-2, 7.145_471_426_517_098e-2, 6.986_299_249_259_416e-2, 6.788_970_337_652_195e-2, 6.554_562_436_490_897e-2, 6.284_355_804_500_258e-2, 5.979_826_222_758_665_6e-2, 5.642_636_935_801_838e-2, 5.274_629_569_917_407e-2, 4.877_814_079_280_324_4e-2, 4.454_357_777_196_588e-2, 4.006_573_518_069_226_5e-2, 3.536_907_109_759_211e-2, 3.047_924_069_960_346_7e-2, 2.542_295_952_611_304_7e-2, 2.022_786_956_905_264_4e-2, 1.492_244_369_735_749_3e-2, 9.536_220_301_748_503e-3, 4.105_998_604_649_085e-3];
-    static EVEN_W_22: [f64; 22] = [7.054_915_778_935_407e-2, 7.019_768_547_355_822e-2, 6.949_649_186_157_258e-2, 6.844_907_026_936_665e-2, 6.706_063_890_629_366e-2, 6.533_811_487_918_144e-2, 6.329_007_973_320_386e-2, 6.092_673_670_156_197e-2, 5.825_985_987_759_549e-2, 5.530_273_556_372_805_6e-2, 5.207_009_609_170_446e-2, 4.857_804_644_835_203_6e-2, 4.484_398_408_197_003e-2, 4.088_651_231_034_622e-2, 3.672_534_781_380_887e-2, 3.238_122_281_206_982e-2, 2.787_578_282_128_101e-2, 2.323_148_190_201_921e-2, 1.847_148_173_681_475e-2, 1.361_958_675_557_998_5e-2, 8.700_481_367_524_843e-3, 3.745_404_803_112_777_6e-3];
-    static EVEN_W_23: [f64; 23] = [6.751_868_584_903_646e-2, 6.721_061_360_067_818e-2, 6.659_587_476_845_488e-2, 6.567_727_426_778_121e-2, 6.445_900_346_713_906e-2, 6.294_662_106_439_451e-2, 6.114_702_772_465_048e-2, 5.906_843_459_554_632e-2, 5.672_032_584_399_123e-2, 5.411_341_538_585_675_4e-2, 5.125_959_800_714_302e-2, 4.817_189_510_171_22e-2, 4.486_439_527_731_813e-2, 4.135_219_010_967_873e-2, 3.765_130_535_738_607e-2, 3.377_862_799_910_69e-2, 2.975_182_955_220_275_6e-2, 2.558_928_639_713_001_2e-2, 2.130_999_875_413_65e-2, 1.693_351_400_783_624e-2, 1.247_988_377_098_868_5e-2, 7.969_898_229_724_623e-3, 3.430_300_868_107_048_3e-3];
-    static EVEN_W_24: [f64; 24] = [6.473_769_681_268_392e-2, 6.446_616_443_595_009e-2, 6.392_423_858_464_819e-2, 6.311_419_228_625_402e-2, 6.203_942_315_989_266_5e-2, 6.070_443_916_589_388e-2, 5.911_483_969_839_563_5e-2, 5.727_729_210_040_321_4e-2, 5.519_950_369_998_416_5e-2, 5.289_018_948_519_367e-2, 5.035_903_555_385_447e-2, 4.761_665_849_249_048e-2, 4.467_456_085_669_428e-2, 4.154_508_294_346_475e-2, 3.824_135_106_583_071e-2, 3.477_722_256_477_044e-2, 3.116_722_783_279_809e-2, 2.742_650_970_835_694_8e-2, 2.357_076_083_932_438e-2, 1.961_616_045_735_553e-2, 1.557_931_572_294_384_9e-2, 1.147_723_457_923_454e-2, 7.327_553_901_276_262_5e-3, 3.153_346_052_305_838_5e-3];
-    static EVEN_W_25: [f64; 25] = [6.217_661_665_534_726e-2, 6.193_606_742_068_324e-2, 6.145_589_959_031_666_5e-2, 6.073_797_084_177_022e-2, 5.978_505_870_426_546e-2, 5.860_084_981_322_244_4e-2, 5.718_992_564_772_838e-2, 5.555_774_480_621_252e-2, 5.371_062_188_899_624_5e-2, 5.165_570_306_958_114e-2, 4.940_093_844_946_632e-2, 4.695_505_130_394_843_4e-2, 4.432_750_433_880_327_4e-2, 4.152_846_309_014_769_6e-2, 3.856_875_661_258_768e-2, 3.545_983_561_514_615e-2, 3.221_372_822_357_801_4e-2, 2.884_299_358_053_519_7e-2, 2.536_067_357_001_239e-2, 2.178_024_317_012_479_4e-2, 1.811_556_071_348_939_2e-2, 1.438_082_276_148_557_4e-2, 1.059_054_838_365_096_9e-2, 6.759_799_195_745_401e-3, 2.908_622_553_155_141e-3];
-    static EVEN_W_26: [f64; 26] = [5.981_036_574_529_186e-2, 5.959_626_017_124_816e-2, 5.916_881_546_604_297e-2, 5.852_956_177_181_387e-2, 5.768_078_745_252_682_6e-2, 5.662_553_090_236_86e-2, 5.536_756_966_930_265e-2, 5.391_140_693_275_726e-2, 5.226_225_538_390_699e-2, 5.042_601_856_634_238e-2, 4.840_926_974_407_49e-2, 4.621_922_837_278_479e-2, 4.386_373_425_900_040_6e-2, 4.135_121_950_056_027e-2, 3.869_067_831_042_398e-2, 3.589_163_483_509_723e-2, 3.296_410_908_971_88e-2, 2.991_858_114_714_394_6e-2, 2.676_595_374_650_401_3e-2, 2.351_751_355_398_446_3e-2, 2.018_489_150_798_079_3e-2, 1.678_002_339_630_073_7e-2, 1.331_511_498_234_096_1e-2, 9.802_634_579_462_751e-3, 6.255_523_962_973_277e-3, 2.691_316_950_047_111_3e-3];
-    static EVEN_W_27: [f64; 27] = [5.761_753_670_714_702_5e-2, 5.742_613_705_411_211e-2, 5.704_397_355_879_46e-2, 5.647_231_573_062_596_5e-2, 5.571_306_256_058_998_5e-2, 5.476_873_621_305_798_6e-2, 5.364_247_364_755_361e-2, 5.233_801_619_829_875e-2, 5.085_969_714_618_815e-2, 4.921_242_732_452_888_6e-2, 4.740_167_880_644_499e-2, 4.543_346_672_827_671_5e-2, 4.331_432_930_959_701e-2, 4.105_130_613_664_497_6e-2, 3.865_191_478_210_252e-2, 3.612_412_584_038_355_4e-2, 3.347_633_646_437_265e-2, 3.071_734_249_787_067_7e-2, 2.785_630_931_059_587e-2, 2.490_274_146_720_877_4e-2, 2.186_645_142_285_308_4e-2, 1.875_752_762_146_938e-2, 1.558_630_303_592_413_1e-2, 1.236_332_812_884_764_4e-2, 9.099_369_455_509_397e-3, 5.805_611_015_239_985e-3, 2.497_481_835_761_586e-3];
-    static EVEN_W_28: [f64; 28] = [5.557_974_630_651_439_7e-2, 5.540_795_250_324_512e-2, 5.506_489_590_176_242_4e-2, 5.455_163_687_088_942_4e-2, 5.386_976_186_571_449e-2, 5.302_137_852_401_076_6e-2, 5.200_910_915_174_14e-2, 5.083_608_261_779_848_4e-2, 4.950_592_468_304_758e-2, 4.802_274_679_360_026e-2, 4.639_113_337_300_189_4e-2, 4.461_612_765_269_228e-2, 4.270_321_608_466_709e-2, 4.065_831_138_474_452e-2, 3.848_773_425_924_766e-2, 3.619_819_387_231_519e-2, 3.379_676_711_561_176e-2, 3.129_087_674_731_044_5e-2, 2.868_826_847_382_274e-2, 2.599_698_705_839_195_4e-2, 2.322_535_156_256_531_5e-2, 2.038_192_988_240_257e-2, 1.747_551_291_140_094_6e-2, 1.451_508_927_802_147_2e-2, 1.150_982_434_038_338_3e-2, 8.469_063_163_307_887e-3, 5.402_522_246_015_338e-3, 2.323_855_375_773_215_6e-3];
-    static EVEN_W_29: [f64; 29] = [5.368_111_986_333_485e-2, 5.352_634_330_405_825_5e-2, 5.321_723_644_657_901e-2, 5.275_469_052_637_083_6e-2, 5.214_003_918_366_982e-2, 5.137_505_461_828_572_5e-2, 5.046_194_247_995_313e-2, 4.940_333_550_896_239e-2, 4.820_228_594_541_775e-2, 4.686_225_672_902_634_4e-2, 4.538_711_151_481_980_6e-2, 4.378_110_353_364_025e-2, 4.204_886_332_958_212_4e-2, 4.019_538_540_986_78e-2, 3.822_601_384_585_843_5e-2, 3.614_642_686_708_727e-2, 3.396_262_049_341_601e-2, 3.168_089_125_380_933e-2, 2.930_781_804_416_049e-2, 2.685_024_318_198_187e-2, 2.431_525_272_496_395_2e-2, 2.171_015_614_014_623_6e-2, 1.904_246_546_189_340_7e-2, 1.631_987_423_497_096_4e-2, 1.355_023_711_298_881_2e-2, 1.074_155_353_287_877_3e-2, 7.901_973_849_998_675e-3, 5.039_981_612_650_243e-3, 2.167_723_249_627_45e-3];
-    static EVEN_W_30: [f64; 30] = [5.190_787_763_122_063_6e-2, 5.176_794_317_491_019e-2, 5.148_845_150_098_093_5e-2, 5.107_015_606_985_563e-2, 5.051_418_453_250_937_4e-2, 4.982_203_569_055_018e-2, 4.899_557_545_575_683_5e-2, 4.803_703_181_997_118e-2, 4.694_898_884_891_22e-2, 4.573_437_971_611_448_6e-2, 4.439_647_879_578_711e-2, 4.293_889_283_593_564e-2, 4.136_555_123_558_475_3e-2, 3.968_069_545_238_08e-2, 3.788_886_756_924_344_4e-2, 3.599_489_805_108_45e-2, 3.400_389_272_494_642e-2, 3.192_121_901_929_633e-2, 2.975_249_150_078_894_4e-2, 2.750_355_674_992_479e-2, 2.518_047_762_152_124_7e-2, 2.278_951_694_399_782e-2, 2.033_712_072_945_728_6e-2, 1.782_990_101_420_772e-2, 1.527_461_859_678_479_9e-2, 1.267_816_647_681_596e-2, 1.004_755_718_228_798_4e-2, 7.389_931_163_345_456e-3, 4.712_729_926_953_568e-3, 2.026_811_968_873_758_5e-3];
-    static EVEN_W_31: [f64; 31] = [5.024_800_037_525_628e-2, 5.012_106_956_904_329e-2, 4.986_752_859_495_239_4e-2, 4.948_801_791_969_929e-2, 4.898_349_622_051_783_5e-2, 4.835_523_796_347_767_5e-2, 4.760_483_018_410_123_5e-2, 4.673_416_847_841_552e-2, 4.574_545_221_457_018e-2, 4.464_117_897_712_441e-2, 4.342_413_825_804_742e-2, 4.209_740_441_038_51e-2, 4.066_432_888_241_744_4e-2, 3.912_853_175_196_308e-2, 3.749_389_258_228_003e-2, 3.576_454_062_276_814e-2, 3.394_484_437_941_054_6e-2, 3.203_940_058_162_468e-2, 3.005_302_257_398_987e-2, 2.799_072_816_331_464e-2, 2.585_772_695_402_469_7e-2, 2.365_940_720_868_279_4e-2, 2.140_132_227_766_996_7e-2, 1.908_917_665_857_32e-2, 1.672_881_179_017_731_6e-2, 1.432_619_182_380_651_8e-2, 1.188_739_011_701_050_1e-2, 9.418_579_428_420_388e-3, 6.926_041_901_830_961e-3, 4.416_333_456_930_905e-3, 1.899_205_679_513_690_5e-3];
-    static EVEN_W_32: [f64; 32] = [4.869_095_700_913_972_4e-2, 4.857_546_744_150_343e-2, 4.834_476_223_480_295_4e-2, 4.799_938_859_645_831e-2, 4.754_016_571_483_031e-2, 4.696_818_281_621_002e-2, 4.628_479_658_131_441_6e-2, 4.549_162_792_741_814e-2, 4.459_055_816_375_656_6e-2, 4.358_372_452_932_345e-2, 4.247_351_512_365_359e-2, 4.126_256_324_262_353e-2, 3.995_374_113_272_034e-2, 3.855_015_317_861_562_6e-2, 3.705_512_854_024_005e-2, 3.547_221_325_688_238_6e-2, 3.380_516_183_714_160_6e-2, 3.205_792_835_485_155e-2, 3.023_465_707_240_247_8e-2, 2.833_967_261_425_948_3e-2, 2.637_746_971_505_466e-2, 2.435_270_256_871_087_4e-2, 2.227_017_380_838_325_3e-2, 2.013_482_315_353_021e-2, 1.795_171_577_569_734_3e-2, 1.572_603_047_602_471_8e-2, 1.346_304_789_671_864_3e-2, 1.116_813_946_013_112_8e-2, 8.846_759_826_363_947e-3, 6.504_457_968_978_363e-3, 4.147_033_260_562_468e-3, 1.783_280_721_696_433e-3];
-    static EVEN_W_33: [f64; 33] = [4.722_748_126_299_855_5e-2, 4.712_209_828_764_473_6e-2, 4.691_156_748_762_083e-2, 4.659_635_863_958_410_5e-2, 4.617_717_509_791_597_5e-2, 4.565_495_222_527_305_6e-2, 4.503_085_530_544_15e-2, 4.430_627_694_315_316_5e-2, 4.348_283_395_666_748e-2, 4.256_236_377_005_572e-2, 4.154_692_031_324_188e-2, 4.043_876_943_895_497_6e-2, 3.924_038_386_682_833e-2, 3.795_443_766_594_162e-2, 3.658_380_028_813_909_6e-2, 3.513_153_016_547_255e-2, 3.360_086_788_611_223e-2, 3.199_522_896_404_688_5e-2, 3.031_819_621_886_852e-2, 2.857_351_178_293_187e-2, 2.676_506_875_425_000_2e-2, 2.489_690_251_475_737_4e-2, 2.297_318_173_532_665_6e-2, 2.099_819_909_186_462_6e-2, 1.897_636_172_277_132_6e-2, 1.691_218_147_224_521_7e-2, 1.481_026_500_273_396e-2, 1.267_530_398_126_168_2e-2, 1.051_206_598_770_575_5e-2, 8.325_388_765_990_902e-3, 6.120_192_018_447_937e-3, 3.901_625_641_744_248e-3, 1.677_653_744_007_238_5e-3];
-    static EVEN_W_34: [f64; 34] = [4.584_938_738_725_097e-2, 4.575_296_541_606_795e-2, 4.556_032_425_064_828_6e-2, 4.527_186_901_844_378e-2, 4.488_820_634_542_666_6e-2, 4.441_014_308_035_276e-2, 4.383_868_459_795_605e-2, 4.317_503_268_464_422_5e-2, 4.242_058_301_114_25e-2, 4.157_692_219_740_291_6e-2, 4.064_582_447_595_408e-2, 3.962_924_796_071_231e-2, 3.852_933_052_910_672e-2, 3.734_838_532_618_666_6e-2, 3.608_889_590_017_987e-2, 3.475_351_097_975_151e-2, 3.334_503_890_398_069e-2, 3.186_644_171_682_106e-2, 3.032_082_893_855_398e-2, 2.871_145_102_748_499_2e-2, 2.704_169_254_590_396e-2, 2.531_506_504_517_64e-2, 2.353_519_968_587_633_5e-2, 2.170_583_961_037_808e-2, 1.983_083_208_795_549_7e-2, 1.791_412_045_792_315_3e-2, 1.595_973_590_961_38e-2, 1.397_178_917_445_765_5e-2, 1.195_446_231_976_944_3e-2, 9.912_001_251_585_938e-3, 7.848_711_393_177_167e-3, 5.768_969_918_729_952e-3, 3.677_366_595_011_730_5e-3, 1.581_140_256_372_913e-3];
-    static EVEN_W_35: [f64; 35] = [4.454_941_715_975_467e-2, 4.446_096_841_724_637e-2, 4.428_424_653_905_540_4e-2, 4.401_960_239_018_346e-2, 4.366_756_139_720_144e-2, 4.322_882_250_506_87e-2, 4.270_425_678_944_978e-2, 4.209_490_572_728_441e-2, 4.140_197_912_904_521e-2, 4.062_685_273_678_961e-2, 3.977_106_549_277_657e-2, 3.883_631_648_407_34e-2, 3.782_446_156_922_282e-2, 3.673_750_969_367_27e-2, 3.557_761_890_129_238e-2, 3.434_709_204_990_654e-2, 3.304_837_223_937_242e-2, 3.168_403_796_130_848e-2, 3.025_679_798_015_423_7e-2, 2.876_948_595_580_828e-2, 2.722_505_481_866_441_6e-2, 2.562_657_090_846_848_2e-2, 2.397_720_788_910_029_4e-2, 2.228_024_045_225_659_7e-2, 2.053_903_782_432_645_4e-2, 1.875_705_709_313_342e-2, 1.693_783_637_630_293e-2, 1.508_498_786_544_312_8e-2, 1.320_219_081_467_674_7e-2, 1.129_318_464_993_153_7e-2, 9.361_762_769_699_028e-3, 7.411_769_363_190_21e-3, 5.447_111_874_217_219e-3, 3.471_894_893_078_143_2e-3, 1.492_721_288_844_515_8e-3];
-    static EVEN_W_36: [f64; 36] = [4.332_111_216_548_653_6e-2, 4.323_978_130_522_261_6e-2, 4.307_727_227_491_37e-2, 4.283_389_016_833_881_5e-2, 4.251_009_191_005_772_3e-2, 4.210_648_539_758_646e-2, 4.162_382_836_013_86e-2, 4.106_302_693_607_506e-2, 4.042_513_397_173_397e-2, 3.971_134_704_483_49e-2, 3.892_300_621_616_966e-2, 3.806_159_151_380_216_6e-2, 3.712_872_015_450_29e-2, 3.612_614_350_763_799_4e-2, 3.505_574_380_721_787e-2, 3.391_953_061_828_606e-2, 3.271_963_706_429_384_4e-2, 3.145_831_582_256_181e-2, 3.013_793_489_537_548e-2, 2.876_097_316_470_176e-2, 2.733_001_573_895_093_4e-2, 2.584_774_910_065_589e-2, 2.431_695_606_441_916_5e-2, 2.274_051_055_503_575_5e-2, 2.112_137_221_644_055_2e-2, 1.946_258_086_329_427_7e-2, 1.776_725_078_920_065_5e-2, 1.603_856_495_028_515_4e-2, 1.427_976_905_455_419_4e-2, 1.249_416_561_987_375_8e-2, 1.068_510_816_535_189_8e-2, 8.855_996_073_706_154e-3, 7.010_272_321_861_863e-3, 5.151_436_018_790_887e-3, 3.283_169_774_667_496e-3, 1.411_516_393_973_434_2e-3];
-    static EVEN_W_37: [f64; 37] = [4.215_870_660_994_342e-2, 4.208_374_996_915_697_4e-2, 4.193_396_995_777_702e-2, 4.170_963_287_924_075e-2, 4.141_113_759_675_351e-2, 4.103_901_482_412_726_5e-2, 4.059_392_618_219_473e-2, 4.007_666_302_247_697e-2, 3.948_814_502_019_647e-2, 3.882_941_853_913_771e-2, 3.810_165_477_126_324_6e-2, 3.730_614_765_439_415_5e-2, 3.644_431_157_165_857e-2, 3.551_767_883_680_096e-2, 3.452_789_696_982_646e-2, 3.347_672_576_782_876_5e-2, 3.236_603_417_621_7e-2, 3.119_779_696_591_542_7e-2, 2.997_409_122_246_118_6e-2, 2.869_709_265_326_987_6e-2, 2.736_907_171_967_935_3e-2, 2.599_238_960_072_378_8e-2, 2.456_949_399_594_276_7e-2, 2.310_291_477_491_582_5e-2, 2.159_525_948_167_589e-2, 2.004_920_870_279_494_3e-2, 1.846_751_130_897_988e-2, 1.685_297_958_202_485_2e-2, 1.520_848_424_340_123_4e-2, 1.353_694_941_178_749_4e-2, 1.184_134_754_749_966_7e-2, 1.012_469_453_828_730_5e-2, 8.390_045_433_971_398e-3, 6.640_492_909_114_358e-3, 4.879_179_758_594_145e-3, 3.109_420_149_896_754_6e-3, 1.336_761_650_069_883_5e-3];
-    static EVEN_W_38: [f64; 38] = [4.105_703_691_622_942_4e-2, 4.098_780_546_479_395e-2, 4.084_945_930_182_849e-2, 4.064_223_171_029_474e-2, 4.036_647_212_284_402e-2, 4.002_264_553_259_682e-2, 3.961_133_170_906_206e-2, 3.913_322_422_051_844e-2, 3.858_912_926_450_673_6e-2, 3.797_996_430_840_528_5e-2, 3.730_675_654_238_161e-2, 3.657_064_114_732_961_5e-2, 3.577_285_938_071_395e-2, 3.491_475_648_355_077e-2, 3.399_777_941_205_638e-2, 3.302_347_439_779_174e-2, 3.199_348_434_042_16e-2, 3.090_954_603_749_159_6e-2, 2.977_348_725_590_504e-2, 2.858_722_365_005_400_2e-2, 2.735_275_553_182_752_2e-2, 2.607_216_449_798_598_4e-2, 2.474_760_992_065_967e-2, 2.338_132_530_701_118_5e-2, 2.197_561_453_441_625e-2, 2.053_284_796_790_802e-2, 1.905_545_846_719_058_2e-2, 1.754_593_729_147_423e-2, 1.600_682_991_224_857_2e-2, 1.444_073_174_827_668e-2, 1.285_028_384_751_014_4e-2, 1.123_816_856_966_768_8e-2, 9.607_105_414_713_754e-3, 7.959_847_477_239_735e-3, 6.299_180_497_328_445_5e-3, 4.627_935_228_037_421e-3, 2.949_102_953_642_474_8e-3, 1.267_791_634_085_359_6e-3];
-    static EVEN_W_39: [f64; 39] = [4.001_146_511_842_048_5e-2, 3.994_739_036_908_802e-2, 3.981_934_348_036_409e-2, 3.962_752_950_781_054_5e-2, 3.937_225_562_423_312_5e-2, 3.905_393_062_777_341e-2, 3.867_306_428_725_767_6e-2, 3.823_026_652_585_099e-2, 3.772_624_644_432_425e-2, 3.716_181_118_549_838_3e-2, 3.653_786_464_168_47e-2, 3.585_540_600_719_17e-2, 3.511_552_817_821_719e-2, 3.431_941_600_268_909e-2, 3.346_834_438_285_898e-2, 3.256_367_623_368_905e-2, 3.160_686_030_030_479_5e-2, 3.059_942_883_801_304_6e-2, 2.954_299_515_860_694_6e-2, 2.843_925_104_689_751_6e-2, 2.728_996_405_162_436_4e-2, 2.609_697_465_510_883_4e-2, 2.486_219_332_622_245e-2, 2.358_759_746_145_747e-2, 2.227_522_821_911_388_8e-2, 2.092_718_725_187_772_7e-2, 1.954_563_334_339_992_3e-2, 1.813_277_895_498_232_8e-2, 1.669_088_668_934_389_3e-2, 1.522_226_568_017_845_2e-2, 1.372_926_792_014_414_8e-2, 1.221_428_454_978_988_6e-2, 1.067_974_215_748_111_4e-2, 9.128_099_227_255_087e-3, 7.561_843_189_439_719e-3, 5.983_489_944_440_408e-3, 4.395_596_039_460_346e-3, 2.800_868_811_838_630_4e-3, 1.204_024_566_067_353_3e-3];
-    static EVEN_W_40: [f64; 40] = [3.901_781_365_630_665e-2, 3.895_839_596_276_953e-2, 3.883_965_105_905_197e-2, 3.866_175_977_407_646_7e-2, 3.842_499_300_695_942_4e-2, 3.812_971_131_447_763_5e-2, 3.777_636_436_200_14e-2, 3.736_549_023_873_049e-2, 3.689_771_463_827_601e-2, 3.637_374_990_583_598e-2, 3.579_439_395_341_605_5e-2, 3.516_052_904_474_759e-2, 3.447_312_045_175_393e-2, 3.373_321_498_461_152e-2, 3.294_193_939_764_54e-2, 3.210_049_867_348_777e-2, 3.121_017_418_811_47e-2, 3.027_232_175_955_798e-2, 2.928_836_958_326_785e-2, 2.825_981_605_727_686_4e-2, 2.718_822_750_048_638e-2, 2.607_523_576_756_511_8e-2, 2.492_253_576_411_549_2e-2, 2.373_188_286_593_01e-2, 2.250_509_024_633_246_3e-2, 2.124_402_611_578_200_5e-2, 1.995_061_087_814_2e-2, 1.862_681_420_829_903_3e-2, 1.727_465_205_626_930_7e-2, 1.589_618_358_372_568_7e-2, 1.449_350_804_050_907_6e-2, 1.306_876_159_240_134e-2, 1.162_411_412_079_782_6e-2, 1.016_176_604_110_306_5e-2, 8.683_945_269_260_859e-3, 7.192_904_768_117_313e-3, 5.690_922_451_403_199e-3, 4.180_313_124_694_896e-3, 2.663_533_589_512_681_8e-3, 1.144_950_003_186_941_4e-3];
-    static EVEN_W_41: [f64; 41] = [3.807_230_964_014_187e-2, 3.801_710_843_143_527e-2, 3.790_678_605_050_578_5e-2, 3.774_150_245_427_587e-2, 3.752_149_728_818_502e-2, 3.724_708_953_872_766_4e-2, 3.691_867_707_095_445_5e-2, 3.653_673_605_160_765_6e-2, 3.610_182_025_872_702_6e-2, 3.561_456_027_872_747_5e-2, 3.507_566_259_211_269e-2, 3.448_590_854_915_071e-2, 3.384_615_323_699_685_7e-2, 3.315_732_423_990_721_4e-2, 3.242_042_029_434_060_6e-2, 3.163_650_984_090_024_5e-2, 3.080_672_947_521_563e-2, 2.993_228_230_001_272_4e-2, 2.901_443_618_076_440_5e-2, 2.805_452_190_745_423e-2, 2.705_393_126_512_477_3e-2, 2.601_411_501_601_702_5e-2, 2.493_658_079_624_075_4e-2, 2.382_289_093_004_782_6e-2, 2.267_466_016_491_410_2e-2, 2.149_355_333_077_484_4e-2, 2.028_128_292_691_216e-2, 1.903_960_664_017_892_4e-2, 1.777_032_479_849_840_7e-2, 1.647_527_776_398_371e-2, 1.515_634_327_076_256_3e-2, 1.381_543_371_412_646e-2, 1.245_449_340_114_210_4e-2, 1.107_549_578_175_989_6e-2, 9.680_440_704_371_074e-3, 8.271_351_818_383_685e-3, 6.850_274_534_183_526e-3, 5.419_276_232_446_765e-3, 3.980_457_937_856_075e-3, 2.536_054_696_856_106e-3, 1.090_118_595_275_831e-3];
-    static EVEN_W_42: [f64; 42] = [3.717_153_701_903_407e-2, 3.712_016_261_260_209e-2, 3.701_748_480_379_452e-2, 3.686_364_550_259_031e-2, 3.665_885_732_875_908e-2, 3.640_340_331_800_212_5e-2, 3.609_763_653_077_257e-2, 3.574_197_956_431_531e-2, 3.533_692_396_860_127_5e-2, 3.488_302_956_696_330_6e-2, 3.438_092_368_237_27e-2, 3.383_130_027_042_598e-2, 3.323_491_896_024_044_5e-2, 3.259_260_400_458_425_4e-2, 3.190_524_314_069_273e-2, 3.117_378_636_334_566e-2, 3.039_924_461_190_247e-2, 2.958_268_837_311_084_4e-2, 2.872_524_620_162_180_2e-2, 2.782_810_316_025_840_7e-2, 2.689_249_918_219_764e-2, 2.591_972_735_733_464_8e-2, 2.491_113_214_520_643e-2, 2.386_810_751_695_824e-2, 2.279_209_502_894_213e-2, 2.168_458_183_064_482e-2, 2.054_709_860_975_628e-2, 1.938_121_747_731_881e-2, 1.818_854_979_605_655e-2, 1.697_074_395_521_161e-2, 1.572_948_309_558_36e-2, 1.446_648_278_916_118_7e-2, 1.318_348_867_918_234_6e-2, 1.188_227_408_980_122_4e-2, 1.056_463_762_300_824_6e-2, 9.232_400_784_190_247e-3, 7.887_405_752_648_147e-3, 6.531_513_687_713_654e-3, 5.166_605_182_746_808_5e-3, 3.794_591_650_452_35e-3, 2.417_511_265_443_123e-3, 1.039_133_516_451_972e-3];
-    static EVEN_W_43: [f64; 43] = [3.631_239_537_581_333_6e-2, 3.626_450_208_420_238_5e-2, 3.616_877_866_860_064e-2, 3.602_535_138_093_526e-2, 3.583_440_939_092_405_5e-2, 3.559_620_453_657_55e-2, 3.531_105_099_203_421e-2, 3.497_932_485_321_01e-2, 3.460_146_364_173_769e-2, 3.417_796_572_791_99e-2, 3.370_938_967_341_756e-2, 3.319_635_349_455_16e-2, 3.263_953_384_718_992e-2, 3.203_966_513_429_401_5e-2, 3.139_753_853_730_286_5e-2, 3.071_400_097_263_205_3e-2, 2.998_995_397_466_493_2e-2, 2.922_635_250_670_994_5e-2, 2.842_420_370_149_349_4e-2, 2.758_456_553_285_124_8e-2, 2.670_854_542_037_221e-2, 2.579_729_876_883_953_7e-2, 2.485_202_744_439_983_5e-2, 2.387_397_818_947_900_6e-2, 2.286_444_097_854_800_6e-2, 2.182_474_731_692_763e-2, 2.075_626_848_490_914_2e-2, 1.966_041_372_956_218e-2, 1.853_862_840_670_985_8e-2, 1.739_239_207_569_054_4e-2, 1.622_321_654_972_902e-2, 1.503_264_390_508_137_8e-2, 1.382_224_445_276_667e-2, 1.259_361_467_806_969_7e-2, 1.134_837_515_617_770_4e-2, 1.008_816_846_038_610_6e-2, 8.814_657_101_954_816e-3, 7.529_521_612_194_562e-3, 6.234_459_139_140_123e-3, 4.931_184_096_960_104e-3, 3.621_439_249_610_901_6e-3, 2.307_087_488_809_903e-3, 9.916_432_666_203_635e-4];
-    static EVEN_W_44: [f64; 44] = [3.549_206_430_171_454_6e-2, 3.544_734_460_447_077e-2, 3.535_796_155_642_384_5e-2, 3.522_402_777_945_910_6e-2, 3.504_571_202_900_426_4e-2, 3.482_323_898_139_936e-2, 3.455_688_895_080_709e-2, 3.424_699_753_602_008e-2, 3.389_395_519_761_026e-2, 3.349_820_676_595_309_5e-2, 3.306_025_088_074_67e-2, 3.258_063_936_273_211e-2, 3.205_997_651_840_639e-2, 3.149_891_837_860_489e-2, 3.089_817_187_191_219_7e-2, 3.025_849_393_394_352_4e-2, 2.958_069_055_361_934_8e-2, 2.886_561_575_763_543e-2, 2.811_417_053_440_861_2e-2, 2.732_730_169_885_533e-2, 2.650_600_069_943_473_8e-2, 2.565_130_236_896_195e-2, 2.476_428_362_076_873_3e-2, 2.384_606_209_185_966e-2, 2.289_779_473_478_114e-2, 2.192_067_635_998_985_4e-2, 2.091_593_813_057_662_4e-2, 1.988_484_601_127_411_5e-2, 1.882_869_917_375_545_2e-2, 1.774_882_836_032_407_4e-2, 1.664_659_420_821_765_6e-2, 1.552_338_553_693_355_4e-2, 1.438_061_760_129_994_4e-2, 1.321_973_031_362_791_3e-2, 1.204_218_643_958_121_2e-2, 1.084_946_977_542_927_2e-2, 9.643_083_322_053_204e-3, 8.424_547_492_702_474e-3, 7.195_398_459_796_372e-3, 5.957_186_996_138_047e-3, 4.711_479_279_598_662e-3, 3.459_867_667_862_796_5e-3, 2.204_058_563_143_696_7e-3, 9.473_355_981_619_273e-4];
-    static EVEN_W_45: [f64; 45] = [3.470_797_248_895_006e-2, 3.466_615_208_568_824e-2, 3.458_256_166_949_689e-2, 3.445_730_196_032_425e-2, 3.429_052_388_637_504e-2, 3.408_242_840_225_4e-2, 3.383_326_624_683_169e-2, 3.354_333_764_112_428e-2, 3.321_299_192_655_132e-2, 3.284_262_714_400_751e-2, 3.243_268_955_425_561_4e-2, 3.198_367_310_021_857_4e-2, 3.149_611_881_181_864e-2, 3.097_061_415_408_092_2e-2, 3.040_779_231_928_695_4e-2, 2.980_833_146_403_127_6e-2, 2.917_295_389_210_074_2e-2, 2.850_242_518_416_141_7e-2, 2.779_755_327_530_227_7e-2, 2.705_918_748_154_796e-2, 2.628_821_747_651_458_7e-2, 2.548_557_221_944_323e-2, 2.465_221_883_590_485_2e-2, 2.378_916_145_252_872_2e-2, 2.289_743_998_716_318_3e-2, 2.197_812_889_593_413_5e-2, 2.103_233_587_872_256_3e-2, 2.006_120_054_463_959_5e-2, 1.906_589_303_913_732e-2, 1.804_761_263_446_023_5e-2, 1.700_758_628_522_267_7e-2, 1.594_706_715_100_664e-2, 1.486_733_308_804_332_4e-2, 1.376_968_511_233_709_3e-2, 1.265_544_583_716_813e-2, 1.152_595_788_914_805_9e-2, 1.038_258_230_989_321_5e-2, 9.226_696_957_741_991e-3, 8.059_694_944_620_015e-3, 6.882_983_208_463_284_5e-3, 5.697_981_560_747_353e-3, 4.506_123_613_674_978e-3, 3.308_867_243_336_018e-3, 2.107_778_774_526_33e-3, 9.059_323_712_148_331e-4];
-    static EVEN_W_46: [f64; 46] = [3.395_777_082_810_235e-2, 3.391_860_442_372_255e-2, 3.384_031_678_893_36e-2, 3.372_299_821_957_387e-2, 3.356_678_402_920_367_6e-2, 3.337_185_439_303_681e-2, 3.313_843_414_012_938e-2, 3.286_679_249_406_566e-2, 3.255_724_276_244_004e-2, 3.221_014_197_549_333e-2, 3.182_589_047_432_008_6e-2, 3.140_493_144_912_218e-2, 3.094_775_042_804_103_3e-2, 3.045_487_471_715_832_2e-2, 2.992_687_279_231_107_5e-2, 2.936_435_364_342_281_4e-2, 2.876_796_607_210_717_7e-2, 2.813_839_794_335_440_4e-2, 2.747_637_539_216_417_3e-2, 2.678_266_198_604_032e-2, 2.605_805_784_431_418e-2, 2.530_339_871_531_322_7e-2, 2.451_955_501_244_097_4e-2, 2.370_743_081_028_191_3e-2, 2.286_796_280_189_254_4e-2, 2.200_211_921_848_586e-2, 2.111_089_871_276_246_3e-2, 2.019_532_920_718_748_4e-2, 1.925_646_670_855_947_5e-2, 1.829_539_409_026_755_7e-2, 1.731_321_984_368_977_6e-2, 1.631_107_680_025_596e-2, 1.529_012_082_579_650_1e-2, 1.425_152_948_895_392_5e-2, 1.319_650_070_571_113_9e-2, 1.212_625_136_263_771_1e-2, 1.104_201_592_263_539_4e-2, 9.945_045_019_726_081e-3, 8.836_604_056_467_877e-3, 7.717_971_837_373_569e-3, 6.590_439_334_214_895e-3, 5.455_308_908_000_871e-3, 4.313_895_331_861_7e-3, 3.167_535_943_396_097_8e-3, 2.017_671_366_262_838_6e-3, 8.671_851_787_671_422e-4];
-    static EVEN_W_47: [f64; 47] = [3.323_930_891_781_532e-2, 3.320_257_661_860_686_5e-2, 3.312_915_261_254_696_7e-2, 3.301_911_803_949_165e-2, 3.287_259_449_712_959e-2, 3.268_974_390_660_630_4e-2, 3.247_076_833_358_768e-2, 3.221_590_976_496_030_4e-2, 3.192_544_984_141_561e-2, 3.159_970_954_621_32e-2, 3.123_904_885_046_741_7e-2, 3.084_386_631_534_919e-2, 3.041_459_865_164_271_3e-2, 2.995_172_023_714_386_8e-2, 2.945_574_259_243_367_5e-2, 2.892_721_381_560_625_6e-2, 2.836_671_797_657_610_8e-2, 2.777_487_447_163_422e-2, 2.715_233_733_896_656_5e-2, 2.649_979_453_589_17e-2, 2.581_796_717_861_673e-2, 2.510_760_874_535_240_5e-2, 2.436_950_424_366_898_8e-2, 2.360_446_934_301_438_4e-2, 2.281_334_947_335_523_7e-2, 2.199_701_889_094_008e-2, 2.115_637_971_222_139e-2, 2.029_236_091_701_113_3e-2, 1.940_591_732_198_200_5e-2, 1.849_802_852_566_591e-2, 1.756_969_782_614_325_3e-2, 1.662_195_111_266_549_8e-2, 1.565_583_573_251_556e-2, 1.467_241_933_449_946_4e-2, 1.367_278_869_060_687_9e-2, 1.265_804_849_763_899_4e-2, 1.162_932_016_112_241_4e-2, 1.058_774_056_495_412_2e-2, 9.534_460_832_865_159e-3, 8.470_645_094_534_636e-3, 7.397_469_288_142_357e-3, 6.316_120_091_036_448e-3, 5.227_794_289_507_768e-3, 4.133_699_875_407_777e-3, 3.035_065_891_038_628e-3, 1.933_219_888_725_419e-3, 8.308_716_126_821_624e-4];
-    static EVEN_W_48: [f64; 48] = [3.255_061_449_236_317e-2, 3.251_611_871_386_883_6e-2, 3.244_716_371_406_427e-2, 3.234_382_256_857_593e-2, 3.220_620_479_403_025e-2, 3.203_445_623_199_266_4e-2, 3.182_875_889_441_101e-2, 3.158_933_077_072_717e-2, 3.131_642_559_686_135_4e-2, 3.101_033_258_631_383_6e-2, 3.067_137_612_366_914_8e-2, 3.029_991_542_082_759_5e-2, 2.989_634_413_632_838_4e-2, 2.946_108_995_816_790_5e-2, 2.899_461_415_055_523_7e-2, 2.849_741_106_508_538_5e-2, 2.797_000_761_684_833_4e-2, 2.741_296_272_602_924_3e-2, 2.682_686_672_559_176e-2, 2.621_234_073_567_241_3e-2, 2.557_003_600_534_936e-2, 2.490_063_322_248_361e-2, 2.420_484_179_236_469_2e-2, 2.348_339_908_592_622e-2, 2.273_706_965_832_937_2e-2, 2.196_664_443_874_435e-2, 2.117_293_989_219_13e-2, 2.035_679_715_433_332_3e-2, 1.951_908_114_014_502e-2, 1.866_067_962_741_146_6e-2, 1.778_250_231_604_526e-2, 1.688_547_986_424_517_4e-2, 1.597_056_290_256_229_3e-2, 1.503_872_102_699_493_7e-2, 1.409_094_177_231_486e-2, 1.312_822_956_696_157_3e-2, 1.215_160_467_108_832e-2, 1.116_210_209_983_849_9e-2, 1.016_077_053_500_841_6e-2, 9.148_671_230_783_386e-3, 8.126_876_925_698_759e-3, 7.096_470_791_153_865e-3, 6.058_545_504_235_961e-3, 5.014_202_742_927_518e-3, 3.964_554_338_444_687e-3, 2.910_731_817_934_946_5e-3, 1.853_960_788_946_921_7e-3, 7.967_920_655_520_124e-4];
-    static EVEN_W_49: [f64; 49] = [3.188_987_535_287_647e-2, 3.185_743_815_812_401e-2, 3.179_259_676_252_863e-2, 3.169_541_712_034_925_5e-2, 3.156_599_807_910_805e-2, 3.140_447_127_904_656e-2, 3.121_100_101_922_626_5e-2, 3.098_578_409_040_993_5e-2, 3.072_904_957_489_367e-2, 3.044_105_861_349_326e-2, 3.012_210_413_992_189_7e-2, 2.977_251_058_282_947_8e-2, 2.939_263_353_580_649e-2, 2.898_285_939_568_834_3e-2, 2.854_360_496_952_788_5e-2, 2.807_531_705_063_614e-2, 2.757_847_196_412_239_3e-2, 2.705_357_508_239_612_8e-2, 2.650_116_031_112_364e-2, 2.592_178_954_616_245e-2, 2.531_605_210_202_609_8e-2, 2.468_456_411_246_099_4e-2, 2.402_796_790_374_55e-2, 2.334_693_134_134_927_3e-2, 2.264_214_715_061_843_2e-2, 2.191_433_221_217_865e-2, 2.116_422_683_277_485_7e-2, 2.039_259_399_229_191_6e-2, 1.960_021_856_772_633e-2, 1.878_790_653_490_468_7e-2, 1.795_648_414_877_063e-2, 1.710_679_710_308_99e-2, 1.623_970_967_045_369_4e-2, 1.535_610_382_349_775_6e-2, 1.445_687_833_830_440_1e-2, 1.354_294_788_102_946_4e-2, 1.261_524_207_892_195_3e-2, 1.167_470_457_713_812_5e-2, 1.072_229_208_322_431_8e-2, 9.758_973_402_174_097e-3, 8.785_728_467_392_263e-3, 7.803_547_379_100_755e-3, 6.813_429_479_165_216e-3, 5.816_382_546_439_639e-3, 4.813_422_398_586_771e-3, 3.805_574_085_352_359_7e-3, 2.793_881_135_722_131e-3, 1.779_477_041_014_528_8e-3, 7.647_669_822_743_135e-4];
-    static EVEN_W_50: [f64; 50] = [3.125_542_345_386_335_4e-2, 3.122_488_425_484_935_8e-2, 3.116_383_569_620_990_7e-2, 3.107_233_742_756_651_5e-2, 3.095_047_885_049_098_7e-2, 3.079_837_903_115_259e-2, 3.061_618_658_398_044_7e-2, 3.040_407_952_645_482e-2, 3.016_226_510_516_914_5e-2, 2.989_097_959_333_283_2e-2, 2.959_048_805_991_264_2e-2, 2.926_108_411_063_827_6e-2, 2.890_308_960_112_52e-2, 2.851_685_432_239_509_8e-2, 2.810_275_565_910_117_3e-2, 2.766_119_822_079_239e-2, 2.719_261_344_657_688e-2, 2.669_745_918_357_096_4e-2, 2.617_621_923_954_567_6e-2, 2.562_940_291_020_811_6e-2, 2.505_754_448_157_959e-2, 2.446_120_270_795_705_2e-2, 2.384_096_026_596_820_7e-2, 2.319_742_318_525_412_3e-2, 2.253_122_025_633_627_3e-2, 2.184_300_241_624_738_8e-2, 2.113_344_211_252_764_2e-2, 2.040_323_264_620_943_3e-2, 1.965_308_749_443_530_5e-2, 1.888_373_961_337_490_3e-2, 1.809_594_072_212_811_6e-2, 1.729_046_056_832_358_3e-2, 1.646_808_617_614_521_3e-2, 1.562_962_107_754_600_4e-2, 1.477_588_452_744_130_2e-2, 1.390_771_070_371_877_3e-2, 1.302_594_789_297_154_2e-2, 1.213_145_766_297_949_8e-2, 1.122_511_402_318_597_7e-2, 1.030_780_257_486_897e-2, 9.380_419_653_694_457e-3, 8.443_871_469_668_97e-3, 7.499_073_255_464_711e-3, 6.546_948_450_845_322e-3, 5.588_428_003_865_515e-3, 4.624_450_063_422_119_6e-3, 3.655_961_201_326_375_4e-3, 2.683_925_371_553_482_6e-3, 1.709_392_653_518_105_2e-3, 7.346_344_905_056_717e-4];
-
-    static ODD_W_1: [f64; 1] = [0.555_555_555_555_555_6];
-    static ODD_W_2: [f64; 2] = [4.786_286_704_993_664_7e-1, 2.369_268_850_561_890_8e-1];
-    static ODD_W_3: [f64; 3] = [3.818_300_505_051_189e-1, 2.797_053_914_892_766_4e-1, 1.294_849_661_688_697e-1];
-    static ODD_W_4: [f64; 4] = [3.123_470_770_400_028_6e-1, 2.606_106_964_029_354_4e-1, 1.806_481_606_948_574e-1, 8.127_438_836_157_441e-2];
-    static ODD_W_5: [f64; 5] = [2.628_045_445_102_466_5e-1, 2.331_937_645_919_904_8e-1, 1.862_902_109_277_342_6e-1, 1.255_803_694_649_046e-1, 5.566_856_711_617_366e-2];
-    static ODD_W_6: [f64; 6] = [2.262_831_802_628_972_3e-1, 2.078_160_475_368_885e-1, 1.781_459_807_619_457_4e-1, 1.388_735_102_197_872_5e-1, 9.212_149_983_772_845e-2, 4.048_400_476_531_588e-2];
-    static ODD_W_7: [f64; 7] = [1.984_314_853_271_115_8e-1, 1.861_610_000_155_622e-1, 1.662_692_058_169_939_2e-1, 1.395_706_779_261_543_2e-1, 1.071_592_204_671_719_4e-1, 7.036_604_748_810_812e-2, 3.075_324_199_611_727e-2];
-    static ODD_W_8: [f64; 8] = [1.765_627_053_669_926_4e-1, 1.680_041_021_564_500_4e-1, 1.540_457_610_768_102_8e-1, 1.351_363_684_685_254_8e-1, 1.118_838_471_934_039_7e-1, 8.503_614_831_717_918e-2, 5.545_952_937_398_72e-2, 2.414_830_286_854_793e-2];
-    static ODD_W_9: [f64; 9] = [1.589_688_433_939_543_4e-1, 1.527_660_420_658_596_7e-1, 1.426_067_021_736_066e-1, 1.287_539_625_393_362_1e-1, 1.115_666_455_473_339_9e-1, 9.149_002_162_245e-2, 6.904_454_273_764_123e-2, 4.481_422_676_569_96e-2, 1.946_178_822_972_647_8e-2];
-    static ODD_W_10: [f64; 10] = [1.445_244_039_899_700_5e-1, 1.398_873_947_910_731_5e-1, 1.322_689_386_333_374_7e-1, 1.218_314_160_537_285_3e-1, 1.087_972_991_671_483_8e-1, 9.344_442_345_603_386e-2, 7.610_011_362_837_93e-2, 5.713_442_542_685_720_5e-2, 3.695_378_977_085_249_4e-2, 1.601_722_825_777_433_5e-2];
-    static ODD_W_11: [f64; 11] = [1.324_620_394_046_966e-1, 1.289_057_221_880_821_6e-1, 1.230_490_843_067_295_3e-1, 1.149_966_402_224_113_6e-1, 1.048_920_914_645_414_1e-1, 9.291_576_606_003_515e-2, 7.928_141_177_671_895e-2, 6.423_242_140_852_585e-2, 4.803_767_173_108_467e-2, 3.098_800_585_697_944_5e-2, 1.341_185_948_714_177_1e-2];
-    static ODD_W_12: [f64; 12] = [1.222_424_429_903_100_4e-1, 1.194_557_635_357_847_7e-1, 1.148_582_591_457_116_4e-1, 1.085_196_244_742_636_5e-1, 1.005_359_490_670_506_4e-1, 9.102_826_198_296_365e-2, 8.014_070_033_500_102e-2, 6.803_833_381_235_691e-2, 5.490_469_597_583_519_4e-2, 4.093_915_670_130_631_6e-2, 2.635_498_661_503_213_7e-2, 1.139_379_850_102_628_8e-2];
-    static ODD_W_13: [f64; 13] = [1.134_763_461_089_651_5e-1, 1.112_524_883_568_451_9e-1, 1.075_782_857_885_331_9e-1, 1.025_016_378_177_458e-1, 9.608_872_737_002_85e-2, 8.842_315_854_375_694e-2, 7.960_486_777_305_777e-2, 6.974_882_376_624_56e-2, 5.898_353_685_983_359_6e-2, 4.744_941_252_061_506e-2, 3.529_705_375_741_971e-2, 2.268_623_159_618_062_3e-2, 9.798_996_051_294_36e-3];
-    static ODD_W_14: [f64; 14] = [1.058_761_550_973_209_4e-1, 1.040_733_100_777_293_8e-1, 1.010_912_737_599_149_6e-1, 9.696_383_409_440_86e-2, 9.173_775_713_925_876e-2, 8.547_225_736_617_253e-2, 7.823_832_713_576_379e-2, 7.011_793_325_505_128e-2, 6.120_309_065_707_913_6e-2, 5.159_482_690_249_793e-2, 4.140_206_251_868_283_6e-2, 3.074_049_220_209_362_4e-2, 1.973_208_505_612_270_7e-2, 8.516_903_878_746_41e-3];
-    static ODD_W_15: [f64; 15] = [9.922_501_122_667_231e-2, 9.774_333_538_632_872e-2, 9.529_024_291_231_951e-2, 9.189_011_389_364_148e-2, 8.757_674_060_847_788e-2, 8.239_299_176_158_926e-2, 7.639_038_659_877_662e-2, 6.962_858_323_541_037e-2, 6.217_478_656_102_843e-2, 5.410_308_242_491_685_5e-2, 4.549_370_752_720_110_4e-2, 3.643_227_391_238_547e-2, 2.700_901_918_497_942_3e-2, 1.731_862_079_031_058_4e-2, 7.470_831_579_248_775_5e-3];
-    static ODD_W_16: [f64; 16] = [9.335_642_606_559_612e-2, 9.212_398_664_331_685e-2, 9.008_195_866_063_858e-2, 8.724_828_761_884_433e-2, 8.364_787_606_703_87e-2, 7.931_236_479_488_674e-2, 7.427_985_484_395_415e-2, 6.859_457_281_865_672e-2, 6.230_648_253_031_748e-2, 5.547_084_663_166_356e-2, 4.814_774_281_871_169_6e-2, 4.040_154_133_166_959_4e-2, 3.230_035_863_232_895_4e-2, 2.391_554_810_174_948e-2, 1.532_170_151_293_467_6e-2, 6.606_227_847_587_378e-3];
-    static ODD_W_17: [f64; 17] = [8.814_053_043_027_546e-2, 8.710_444_699_718_353e-2, 8.538_665_339_209_912e-2, 8.300_059_372_885_658e-2, 7.996_494_224_232_427e-2, 7.630_345_715_544_205e-2, 7.204_479_477_256_007e-2, 6.722_228_526_908_69e-2, 6.187_367_196_608_018_6e-2, 5.604_081_621_237_013e-2, 4.976_937_040_135_353e-2, 4.310_842_232_617_021_6e-2, 3.611_011_586_346_338e-2, 2.882_926_010_889_425_4e-2, 2.132_297_991_148_358_2e-2, 1.365_082_834_836_149_3e-2, 5.883_433_420_443_085e-3];
-    static ODD_W_18: [f64; 18] = [8.347_457_362_586_279e-2, 8.259_527_223_643_726e-2, 8.113_662_450_846_502e-2, 7.910_886_183_752_938e-2, 7.652_620_757_052_923e-2, 7.340_677_724_848_817e-2, 6.977_245_155_570_035e-2, 6.564_872_287_275_125e-2, 6.106_451_652_322_599e-2, 5.605_198_799_827_492e-2, 5.064_629_765_482_460_4e-2, 4.488_536_466_243_716_5e-2, 3.880_960_250_193_454e-2, 3.246_163_984_752_148_4e-2, 2.588_603_699_055_893_5e-2, 1.912_904_448_908_396_6e-2, 1.223_878_010_030_755_7e-2, 5.273_057_279_497_939e-3];
-    static ODD_W_19: [f64; 19] = [7.927_622_256_836_847e-2, 7.852_361_328_737_117e-2, 7.727_455_254_468_202e-2, 7.553_693_732_283_606e-2, 7.332_175_341_426_862e-2, 7.064_300_597_060_875e-2, 6.751_763_096_623_127e-2, 6.396_538_813_868_238e-2, 6.000_873_608_859_615e-2, 5.567_269_034_091_63e-2, 5.098_466_529_212_94e-2, 4.597_430_110_891_663_5e-2, 4.067_327_684_793_384e-2, 3.511_511_149_813_133e-2, 2.933_495_598_390_338e-2, 2.336_938_483_217_816_3e-2, 1.725_622_909_372_491_8e-2, 1.103_478_893_916_459_4e-2, 4.752_944_691_635_101e-3];
-    static ODD_W_20: [f64; 20] = [7.547_874_709_271_582e-2, 7.482_962_317_622_155e-2, 7.375_188_202_722_346e-2, 7.225_169_686_102_308e-2, 7.033_766_062_081_749e-2, 6.802_073_676_087_676e-2, 6.531_419_645_352_741e-2, 6.223_354_258_096_632e-2, 5.879_642_094_987_195e-2, 5.502_251_924_257_874e-2, 5.093_345_429_461_749e-2, 4.655_264_836_901_434e-2, 4.190_519_519_590_969e-2, 3.701_771_670_350_799e-2, 3.191_821_173_169_928e-2, 2.663_589_920_711_044_4e-2, 2.120_106_336_877_955_5e-2, 1.564_493_840_781_859e-2, 9.999_938_773_905_946e-3, 4.306_140_358_164_887_5e-3];
-    static ODD_W_21: [f64; 21] = [7.202_750_197_142_198e-2, 7.146_373_425_251_414e-2, 7.052_738_776_508_503e-2, 6.922_334_419_365_668e-2, 6.755_840_222_936_517e-2, 6.554_124_212_632_28e-2, 6.318_238_044_939_611e-2, 6.049_411_524_999_129e-2, 5.749_046_195_691_052e-2, 5.418_708_031_888_179e-2, 5.060_119_278_439_015e-2, 4.675_149_475_434_658e-2, 4.265_805_719_798_208e-2, 3.834_222_219_413_266e-2, 3.382_649_208_686_029e-2, 2.913_441_326_149_849_4e-2, 2.429_045_661_383_881_5e-2, 1.931_990_142_368_39e-2, 1.424_875_643_157_648_6e-2, 9.103_996_637_401_404e-3, 3.919_490_253_844_127_4e-3];
-    static ODD_W_22: [f64; 22] = [6.887_731_697_766_132e-2, 6.838_457_737_866_968e-2, 6.756_595_416_360_754e-2, 6.642_534_844_984_253e-2, 6.496_819_575_072_343e-2, 6.320_144_007_381_993e-2, 6.113_350_083_106_652e-2, 5.877_423_271_884_174e-2, 5.613_487_875_978_647_6e-2, 5.322_801_673_126_895e-2, 5.006_749_923_795_203e-2, 4.666_838_771_837_336_6e-2, 4.304_688_070_916_497e-2, 3.922_023_672_930_245e-2, 3.520_669_220_160_902e-2, 3.102_537_493_451_546_6e-2, 2.669_621_396_757_766_4e-2, 2.223_984_755_057_873_2e-2, 1.767_753_525_793_759e-2, 1.303_110_499_158_278_5e-2, 8.323_189_296_218_24e-3, 3.582_663_155_283_559e-3];
-    static ODD_W_23: [f64; 23] = [6.599_053_358_881_048e-2, 6.555_737_776_654_974e-2, 6.483_755_623_894_573e-2, 6.383_421_660_571_703e-2, 6.255_174_622_092_166e-2, 6.099_575_300_873_965e-2, 5.917_304_094_233_888e-2, 5.709_158_029_323_154e-2, 5.476_047_278_153_023e-2, 5.218_991_178_005_714e-2, 4.939_113_774_736_117e-2, 4.637_638_908_650_591e-2, 4.315_884_864_847_954e-2, 3.975_258_612_253_100_6e-2, 3.617_249_658_417_495e-2, 3.243_423_551_518_475_5e-2, 2.855_415_070_064_338_7e-2, 2.454_921_165_965_882e-2, 2.043_693_814_766_842_8e-2, 1.623_533_314_643_306e-2, 1.196_284_846_431_232_1e-2, 7.638_616_295_848_833e-3, 3.287_453_842_528_015e-3];
-    static ODD_W_24: [f64; 24] = [6.333_550_929_649_175e-2, 6.295_270_746_519_57e-2, 6.231_641_732_005_726_4e-2, 6.142_920_097_919_294e-2, 6.029_463_095_315_202e-2, 5.891_727_576_002_727e-2, 5.730_268_153_018_748e-2, 5.545_734_967_480_359e-2, 5.338_871_070_825_897e-2, 5.110_509_433_014_459e-2, 4.861_569_588_782_824e-2, 4.593_053_935_559_585e-2, 4.306_043_698_125_959_5e-2, 4.001_694_576_637_302e-2, 3.681_232_096_300_069e-2, 3.345_946_679_162_218e-2, 2.997_188_462_058_382_5e-2, 2.636_361_892_706_601_6e-2, 2.264_920_158_744_667_5e-2, 1.884_359_585_308_945_8e-2, 1.496_214_493_562_465e-2, 1.102_055_103_159_358e-2, 7.035_099_590_086_451_4e-3, 3.027_278_988_922_905e-3];
-    static ODD_W_25: [f64; 25] = [6.088_546_484_485_634_5e-2, 6.054_550_693_473_779e-2, 5.998_031_577_750_325_6e-2, 5.919_199_392_296_154_6e-2, 5.818_347_398_259_214e-2, 5.695_850_772_025_866_4e-2, 5.552_165_209_573_869e-2, 5.387_825_231_304_556e-2, 5.203_442_193_669_708_5e-2, 4.999_702_015_005_741e-2, 4.777_362_624_062_310_4e-2, 4.537_251_140_765_007e-2, 4.280_260_799_788_008_4e-2, 4.007_347_628_549_646e-2, 3.719_526_892_326_029e-2, 3.417_869_320_418_833_6e-2, 3.103_497_129_016_001e-2, 2.777_579_859_416_247_7e-2, 2.441_330_057_378_143_4e-2, 2.095_998_840_170_321e-2, 1.742_871_472_340_105_2e-2, 1.383_263_400_647_782_2e-2, 1.018_519_129_782_173e-2, 6.500_337_783_252_6e-3, 2.796_807_171_089_895_4e-3];
-    static ODD_W_26: [f64; 26] = [5.861_758_623_272_026_6e-2, 5.831_431_136_225_601e-2, 5.781_001_499_171_32e-2, 5.710_643_553_626_719_4e-2, 5.620_599_838_173_971e-2, 5.511_180_752_393_36e-2, 5.382_763_486_873_103e-2, 5.235_790_722_987_272e-2, 5.070_769_106_929_272e-2, 4.888_267_503_269_914e-2, 4.688_915_034_075_031_4e-2, 4.473_398_910_367_281e-2, 4.242_462_063_452_001_6e-2, 3.996_900_584_354_038_5e-2, 3.737_560_980_348_291_6e-2, 3.465_337_258_353_424e-2, 3.181_167_845_901_932_6e-2, 2.886_032_361_782_373_7e-2, 2.580_948_251_075_751_8e-2, 2.266_967_305_707_020_7e-2, 1.945_172_110_763_689_4e-2, 1.616_672_525_668_746_3e-2, 1.282_602_614_424_037_9e-2, 9.441_202_284_940_345e-3, 6.024_276_226_948_674e-3, 2.591_683_720_567_032e-3];
-    static ODD_W_27: [f64; 27] = [5.651_231_824_977_2e-2, 5.624_063_407_108_436_5e-2, 5.578_879_419_528_409e-2, 5.515_824_600_250_869e-2, 5.435_100_932_991_110_4e-2, 5.336_967_000_160_547_4e-2, 5.221_737_154_563_209e-2, 5.089_780_512_449_398e-2, 4.941_519_771_155_174e-2, 4.777_429_855_120_069_6e-2, 4.598_036_394_628_384e-2, 4.403_914_042_160_659e-2, 4.195_684_631_771_876e-2, 3.974_015_187_433_718e-2, 3.739_615_786_796_554_6e-2, 3.493_237_287_358_988_4e-2, 3.235_668_922_618_583_5e-2, 2.967_735_776_516_104e-2, 2.690_296_145_639_627e-2, 2.404_238_800_972_562e-2, 2.110_480_166_801_645_4e-2, 1.809_961_452_072_906_4e-2, 1.503_645_833_351_178_8e-2, 1.192_516_071_984_861_2e-2, 8.775_746_107_058_528e-3, 5.598_632_266_560_767_5e-3, 2.408_323_619_979_789e-3];
-    static ODD_W_28: [f64; 28] = [5.455_280_360_476_188e-2, 5.430_847_145_249_864e-2, 5.390_206_148_329_857_6e-2, 5.333_478_658_481_916e-2, 5.260_833_972_917_743e-2, 5.172_488_892_051_782_5e-2, 5.068_707_072_492_741e-2, 4.949_798_240_201_968e-2, 4.816_117_266_168_775e-2, 4.668_063_107_364_15e-2, 4.506_077_616_138_115_5e-2, 4.330_644_221_621_52e-2, 4.142_286_487_080_111e-2, 3.941_566_547_548_011_6e-2, 3.729_083_432_441_731_4e-2, 3.505_471_278_231_262e-2, 3.271_397_436_637_157e-2, 3.027_560_484_269_4e-2, 2.774_688_140_218_019_3e-2, 2.513_535_099_091_812_3e-2, 2.244_880_789_077_643_6e-2, 1.969_527_069_948_852e-2, 1.688_295_902_344_155e-2, 1.402_027_079_075_355_6e-2, 1.111_576_373_233_599e-2, 8.178_160_067_821_233e-3, 5.216_533_474_718_78e-3, 2.243_753_872_250_663e-3];
-    static ODD_W_29: [f64; 29] = [5.272_443_385_912_793e-2, 5.250_390_264_782_874e-2, 5.213_703_364_837_539_4e-2, 5.162_484_939_089_148e-2, 5.096_877_742_539_391_4e-2, 5.017_064_634_299_69e-2, 4.923_268_067_936_198_4e-2, 4.815_749_471_460_644e-2, 4.694_808_518_696_201_6e-2, 4.560_782_294_050_977e-2, 4.414_044_353_029_738_4e-2, 4.255_003_681_106_763_6e-2, 4.084_103_553_868_671e-2, 3.901_820_301_616_001e-2, 3.708_661_981_887_092_4e-2, 3.505_166_963_640_011e-2, 3.291_902_427_104_527_5e-2, 3.069_462_783_611_168_2e-2, 2.838_468_020_053_48e-2, 2.599_561_973_129_85e-2, 2.353_410_539_371_336_4e-2, 2.100_699_828_843_718_6e-2, 1.842_134_275_361_003e-2, 1.578_434_731_308_146_8e-2, 1.310_336_630_634_519e-2, 1.038_588_550_099_586_2e-2, 7.639_529_453_487_575_6e-3, 4.872_239_168_265_285e-3, 2.095_492_284_541_223_5e-3];
-    static ODD_W_30: [f64; 30] = [5.101_448_703_869_726_5e-2, 5.081_476_366_881_834e-2, 5.048_247_038_679_741e-2, 5.001_847_410_817_825e-2, 4.942_398_534_673_559e-2, 4.870_055_505_641_153e-2, 4.785_007_058_509_560_5e-2, 4.687_475_075_080_907e-2, 4.577_714_005_314_596e-2, 4.456_010_203_508_349e-2, 4.322_681_181_249_609_5e-2, 4.178_074_779_088_849_4e-2, 4.022_568_259_099_825e-2, 3.856_567_320_700_817_6e-2, 3.680_505_042_315_481_6e-2, 3.494_840_751_653_335e-2, 3.300_058_827_590_741e-2, 3.096_667_436_839_739_6e-2, 2.885_197_208_818_34e-2, 2.666_199_852_415_089e-2, 2.440_246_718_754_420_3e-2, 2.207_927_314_831_904_5e-2, 1.969_847_774_610_118e-2, 1.726_629_298_761_374_3e-2, 1.478_906_588_493_791_5e-2, 1.227_326_350_781_210_4e-2, 9.725_461_830_356_134e-3, 7.152_354_991_749_09e-3, 4.560_924_006_012_417e-3, 1.961_453_361_670_283e-3];
-    static ODD_W_31: [f64; 31] = [4.941_183_303_991_818e-2, 4.923_038_042_374_756e-2, 4.892_845_282_051_199e-2, 4.850_678_909_788_385e-2, 4.796_642_113_799_513e-2, 4.730_867_131_226_891_6e-2, 4.653_514_924_538_37e-2, 4.564_774_787_629_261e-2, 4.464_863_882_594_139_6e-2, 4.354_026_708_302_759e-2, 4.232_534_502_081_582e-2, 4.100_684_575_966_64e-2, 3.958_799_589_154_409_6e-2, 3.807_226_758_434_955_5e-2, 3.646_337_008_545_729e-2, 3.476_524_064_535_587_6e-2, 3.298_203_488_377_934e-2, 3.111_811_662_221_982e-2, 2.917_804_720_828_052_7e-2, 2.716_657_435_909_793_4e-2, 2.508_862_055_334_498_7e-2, 2.294_927_100_488_993_2e-2, 2.075_376_125_803_909e-2, 1.850_746_446_016_127e-2, 1.621_587_841_033_834e-2, 1.388_461_261_611_561_1e-2, 1.151_937_607_688_004_2e-2, 9.125_968_676_326_656e-3, 6.710_291_765_960_137e-3, 4.278_508_346_863_762e-3, 1.839_874_595_577_084_2e-3];
-    static ODD_W_32: [f64; 32] = [4.790_669_250_049_586e-2, 4.774_134_868_124_062e-2, 4.746_619_823_288_550_4e-2, 4.708_187_401_045_452e-2, 4.658_925_997_223_349_6e-2, 4.598_948_914_665_17e-2, 4.528_394_102_630_023_4e-2, 4.447_423_839_508_297e-2, 4.356_224_359_580_048_6e-2, 4.255_005_424_675_58e-2, 4.143_999_841_724_029e-2, 4.023_462_927_300_553_6e-2, 3.893_671_920_405_119_4e-2, 3.754_925_344_825_770_6e-2, 3.607_542_322_556_528e-2, 3.451_861_839_854_905_6e-2, 3.288_241_967_636_857_5e-2, 3.117_059_038_018_914_2e-2, 2.938_706_778_931_066_8e-2, 2.753_595_408_845_034_6e-2, 2.562_150_693_803_775_7e-2, 2.364_812_969_128_723_6e-2, 2.162_036_128_493_406_4e-2, 1.954_286_583_675_006e-2, 1.742_042_199_767_024_7e-2, 1.525_791_214_644_831e-2, 1.306_031_163_999_484_7e-2, 1.083_267_878_959_796_9e-2, 8.580_148_266_881_46e-3, 6.307_942_578_971_755e-3, 4.021_524_172_003_736e-3, 1.729_258_251_300_250_8e-3];
-    static ODD_W_33: [f64; 33] = [4.649_043_816_026_463e-2, 4.633_935_168_241_562e-2, 4.608_790_448_976_158e-2, 4.573_664_116_106_369e-2, 4.528_632_245_466_953e-2, 4.473_792_366_088_983e-2, 4.409_263_248_975_102e-2, 4.335_184_649_869_952e-2, 4.251_717_006_583_049_4e-2, 4.159_041_091_519_924_4e-2, 4.057_357_620_174_452_4e-2, 3.946_886_816_430_888e-2, 3.827_867_935_617_948e-2, 3.700_558_746_349_258_4e-2, 3.565_234_972_274_501e-2, 3.422_189_694_953_664_5e-2, 3.271_732_719_153_120_5e-2, 3.114_189_901_947_282_3e-2, 2.949_902_447_094_567e-2, 2.779_226_166_243_677e-2, 2.602_530_708_621_324e-2, 2.420_198_760_967_316_4e-2, 2.232_625_219_645_207_8e-2, 2.040_216_337_134_354e-2, 1.843_388_845_680_457_2e-2, 1.642_569_062_253_088e-2, 1.438_191_982_720_055_1e-2, 1.230_700_384_928_815e-2, 1.020_544_003_410_244_1e-2, 8.081_790_299_023_136e-3, 5.940_693_177_582_235e-3, 3.787_008_301_825_508_3e-3, 1.628_325_035_240_012_9e-3];
-    static ODD_W_34: [f64; 34] = [4.515_543_023_614_546e-2, 4.501_700_814_039_98e-2, 4.478_661_887_831_255_5e-2, 4.446_473_312_204_713_6e-2, 4.405_200_846_590_929e-2, 4.354_928_808_292_674e-2, 4.295_759_900_230_521e-2, 4.227_815_001_128_051e-2, 4.151_232_918_565_45e-2, 4.066_170_105_406_160_4e-2, 3.972_800_340_176_164e-2, 3.871_314_372_049_251e-2, 3.761_919_531_164_091e-2, 3.644_839_305_070_051_6e-2, 3.520_312_882_168_348_5e-2, 3.388_594_663_083_229e-2, 3.249_953_740_964_611e-2, 3.104_673_351_789_054e-2, 2.953_050_295_790_671_2e-2, 2.795_394_331_218_770_5e-2, 2.632_027_541_686_948_4e-2, 2.463_283_678_454_245_7e-2, 2.289_507_479_074_078_7e-2, 2.111_053_963_987_189_6e-2, 1.928_287_712_884_940_4e-2, 1.741_582_123_196_983e-2, 1.551_318_654_340_616_6e-2, 1.357_886_064_907_567_2e-2, 1.161_679_661_067_196_5e-2, 9.631_006_150_415_575e-3, 7.625_555_931_201_511e-3, 5.604_579_927_870_595e-3, 3.572_416_739_397_372_5e-3, 1.535_976_952_792_084e-3];
-    static ODD_W_35: [f64; 35] = [4.389_487_921_178_858_4e-2, 4.376_774_491_340_214e-2, 4.355_612_710_410_853_4e-2, 4.326_043_426_324_127e-2, 4.288_123_715_758_044e-2, 4.241_926_773_962_459_4e-2, 4.187_541_773_473_301e-2, 4.125_073_691_986_603e-2, 4.054_643_109_724_69e-2, 3.976_385_976_685_758e-2, 3.890_453_350_226_294_5e-2, 3.797_011_103_483_116e-2, 3.696_239_605_198_203e-2, 3.588_333_371_564_891e-2, 3.473_500_690_768_218_5e-2, 3.351_963_220_945_403e-2, 3.223_955_562_344_353e-2, 3.089_724_804_509_072e-2, 2.949_530_049_370_881_2e-2, 2.803_641_911_174_149e-2, 2.652_341_994_215_790_7e-2, 2.495_922_349_431_387_4e-2, 2.334_684_910_922_325e-2, 2.168_940_913_598_537e-2, 1.999_010_293_235_011_3e-2, 1.825_221_070_467_867e-2, 1.647_908_720_746_239_8e-2, 1.467_415_533_461_152_9e-2, 1.284_089_966_808_780_7e-2, 1.098_286_015_429_855_2e-2, 9.103_626_461_992_005e-3, 7.206_835_281_831_493e-3, 5.296_182_844_025_892_6e-3, 3.375_555_496_730_676e-3, 1.451_267_330_029_397_3e-3];
-    static ODD_W_36: [f64; 36] = [4.270_273_086_485_722e-2, 4.258_568_982_601_839e-2, 4.239_085_899_223_159e-2, 4.211_859_425_425_564e-2, 4.176_939_294_869_285_6e-2, 4.134_389_294_952_549e-2, 4.084_287_150_293_886e-2, 4.026_724_380_756_003_4e-2, 3.961_806_134_270_614_4e-2, 3.889_650_994_769_673_6e-2, 3.810_390_765_573_98e-2, 3.724_170_228_634_977e-2, 3.631_146_880_069_778_5e-2, 3.531_490_642_472_829e-2, 3.425_383_554_530_221_6e-2, 3.313_019_438_504_384e-2, 3.194_603_546_197_670_5e-2, 3.070_352_184_043_350_5e-2, 2.940_492_318_011_656e-2, 2.805_261_159_057_206e-2, 2.664_905_729_872_748_4e-2, 2.519_682_413_753_831_4e-2, 2.369_856_486_421_897_6e-2, 2.215_701_631_704_007_3e-2, 2.057_499_442_036_117e-2, 1.895_538_904_867_002_2e-2, 1.730_115_876_248_908_3e-2, 1.561_532_543_359_142_3e-2, 1.390_096_878_831_465_1e-2, 1.216_122_092_928_111_2e-2, 1.039_926_099_500_053_2e-2, 8.618_310_479_532_248e-3, 6.821_631_349_174_792e-3, 5.012_538_571_606_19e-3, 3.194_524_377_289_034_5e-3, 1.373_376_462_759_619_2e-3];
-    static ODD_W_37: [f64; 37] = [4.157_356_944_178_127_6e-2, 4.146_558_103_261_909e-2, 4.128_580_808_246_718_6e-2, 4.103_456_181_139_21e-2, 4.071_227_717_293_733e-2, 4.031_951_210_114_157_6e-2, 3.985_694_654_465_635e-2, 3.932_538_128_963_516e-2, 3.872_573_657_343_257_6e-2, 3.805_905_049_151_360_5e-2, 3.732_647_720_033_209e-2, 3.652_928_491_929_034e-2, 3.566_885_373_524_045e-2, 3.474_667_321_333_041e-2, 3.376_433_981_833_409e-2, 3.272_355_415_093_422e-2, 3.162_611_800_374_965e-2, 3.047_393_124_221_454e-2, 2.926_898_851_572_598_5e-2, 2.801_337_580_478_054e-2, 2.670_926_681_012_085e-2, 2.535_891_919_021_637_8e-2, 2.396_467_065_371_696e-2, 2.252_893_491_386_577_5e-2, 2.105_419_751_228_284e-2, 1.954_301_152_012_788_8e-2, 1.799_799_312_564_505e-2, 1.642_181_711_902_464e-2, 1.481_721_228_981_446_9e-2, 1.318_695_676_282_480_1e-2, 1.153_387_332_830_449_6e-2, 9.860_824_916_114_019e-3, 8.170_710_707_327_826e-3, 6.466_464_907_037_539e-3, 4.751_069_185_015_274e-3, 3.027_671_014_606_041e-3, 1.301_591_717_375_856e-3];
-    static ODD_W_38: [f64; 38] = [4.050_253_572_678_803e-2, 4.040_269_003_221_775_4e-2, 4.023_646_282_485_108_4e-2, 4.000_412_721_559_124e-2, 3.970_606_493_128_931e-2, 3.934_276_568_757_015_5e-2, 3.891_482_638_423_378e-2, 3.842_295_012_455_452_4e-2, 3.786_794_506_008_932e-2, 3.725_072_306_289_372e-2, 3.657_229_822_732_745e-2, 3.583_378_520_391_196_4e-2, 3.503_639_736_797_828e-2, 3.418_144_482_611_568e-2, 3.327_033_226_369_854e-2, 3.230_455_663_703_097_5e-2, 3.128_570_471_390_543e-2, 3.021_545_046_662_3e-2, 2.909_555_232_176_876e-2, 2.792_785_027_127_697e-2, 2.671_426_284_955_789_2e-2, 2.545_678_398_169_440_3e-2, 2.415_747_970_795_584_6e-2, 2.281_848_479_012_952e-2, 2.144_199_920_545_613_5e-2, 2.003_028_453_431_617_7e-2, 1.858_566_024_834_148_7e-2, 1.711_049_990_653_110_5e-2, 1.560_722_726_874_913_2e-2, 1.407_831_234_002_700_4e-2, 1.252_626_736_922_736_5e-2, 1.095_364_285_391_135_4e-2, 9.363_023_692_386_431e-3, 7.757_025_950_083_071e-3, 6.138_296_159_756_341_5e-3, 4.509_523_600_205_835e-3, 2.873_553_083_652_691_6e-3, 1.235_291_177_139_409_6e-3];
-    static ODD_W_39: [f64; 39] = [3.948_525_740_129_116e-2, 3.939_275_600_474_300_6e-2, 3.923_874_749_659_464_5e-2, 3.902_347_234_287_98e-2, 3.874_726_667_023_997e-2, 3.841_056_174_110_418e-2, 3.801_388_328_032_605e-2, 3.755_785_065_432_977e-2, 3.704_317_590_404_678_6e-2, 3.647_066_263_315_342_6e-2, 3.584_120_475_334_575_5e-2, 3.515_578_508_861_113_4e-2, 3.441_547_384_067_66e-2, 3.362_142_691_803_093e-2, 3.277_488_413_113_082e-2, 3.187_716_725_661_117e-2, 3.092_967_797_352_483_6e-2, 2.993_389_567_483_836_4e-2, 2.889_137_515_760_726_8e-2, 2.780_374_419_544_706e-2, 2.667_270_099_710_555_5e-2, 2.550_001_155_512_877_2e-2, 2.428_750_688_879_949_3e-2, 2.303_708_018_571_902_5e-2, 2.175_068_384_660_808e-2, 2.043_032_643_814_086e-2, 1.907_806_955_893_749e-2, 1.769_602_462_431_041_7e-2, 1.628_634_957_619_168_3e-2, 1.485_124_552_635_006_8e-2, 1.339_295_334_482_567_6e-2, 1.191_375_021_511_699_8e-2, 1.041_594_620_451_338_3e-2, 8.901_880_982_652_486e-3, 7.373_921_131_330_177e-3, 5.834_459_868_763_466e-3, 4.285_929_113_126_531_5e-3, 2.730_907_065_754_856e-3, 1.173_930_129_956_613e-3];
-    static ODD_W_40: [f64; 40] = [3.851_778_959_688_47e-2, 3.843_192_958_037_517_5e-2, 3.828_897_129_558_352e-2, 3.808_912_713_547_56e-2, 3.783_269_400_830_056e-2, 3.752_005_289_647_583_6e-2, 3.715_166_829_056_371e-2, 3.672_808_749_918_044e-2, 3.624_993_983_586_341_5e-2, 3.571_793_568_410_457e-2, 3.513_286_544_193_938e-2, 3.449_559_834_765_98e-2, 3.380_708_118_839_624_6e-2, 3.306_833_689_348_8e-2, 3.228_046_301_473_269e-2, 3.144_463_009_577_406_3e-2, 3.056_207_993_305_266_2e-2, 2.963_412_373_090_559_8e-2, 2.866_214_015_356_067_7e-2, 2.764_757_327_692_492_6e-2, 2.659_193_044_321_992_2e-2, 2.549_678_002_166_567_6e-2, 2.436_374_907_856_309_8e-2, 2.319_452_096_027_392e-2, 2.199_083_279_275_163_4e-2, 2.075_447_290_144_561e-2, 1.948_727_815_560_191_7e-2, 1.819_113_124_125_576e-2, 1.686_795_786_763_514e-2, 1.551_972_391_246_436_4e-2, 1.414_843_251_323_606_5e-2, 1.275_612_111_513_442_2e-2, 1.134_485_849_541_625_6e-2, 9.916_741_809_595_876e-3, 8.473_893_785_345_566e-3, 7.018_460_484_931_626e-3, 5.552_611_370_256_279e-3, 4.078_551_113_421_395e-3, 2.598_622_299_928_953e-3, 1.117_029_847_124_606_6e-3];
-    static ODD_W_41: [f64; 41] = [3.759_656_394_395_518e-2, 3.751_672_450_373_727_5e-2, 3.738_378_433_575_740_4e-2, 3.719_793_160_197_673e-2, 3.695_942_935_618_497e-2, 3.666_861_517_167_809e-2, 3.632_590_066_346_228_6e-2, 3.593_177_090_566_064_5e-2, 3.548_678_374_494_710_5e-2, 3.499_156_901_097_965_5e-2, 3.444_682_762_495_051_4e-2, 3.385_333_060_751_52e-2, 3.321_191_798_750_501_6e-2, 3.252_349_761_296_806_4e-2, 3.178_904_386_622_215e-2, 3.100_959_628_473_919_4e-2, 3.018_625_808_981_441_7e-2, 2.932_019_462_510_452_8e-2, 2.841_263_170_724_764_3e-2, 2.746_485_389_090_326_2e-2, 2.647_820_265_067_376_4e-2, 2.545_407_448_248_949_6e-2, 2.439_391_892_715_856e-2, 2.329_923_651_890_055e-2, 2.217_157_666_180_362_4e-2, 2.101_253_543_726_991_7e-2, 1.982_375_334_565_493_8e-2, 1.860_691_298_547_847_3e-2, 1.736_373_667_382_462_2e-2, 1.609_598_401_193_537e-2, 1.480_544_940_071_787_7e-2, 1.349_395_951_237_523_6e-2, 1.216_337_072_779_861_3e-2, 1.081_556_655_803_715_9e-2, 9.452_455_092_479_7e-3, 8.075_966_593_123_452e-3, 6.688_051_635_243_685e-3, 5.290_681_445_859_866e-3, 3.885_859_435_353_202e-3, 2.475_719_322_545_939_5e-3, 1.064_168_219_666_567_8e-3];
-    static ODD_W_42: [f64; 42] = [3.671_834_473_341_962e-2, 3.664_397_593_378_570_4e-2, 3.652_013_948_874_488_5e-2, 3.634_700_257_169_52e-2, 3.612_479_890_936_246e-2, 3.585_382_846_628_081e-2, 3.553_445_703_985_57e-2, 3.516_711_576_655_579e-2, 3.475_230_053_990_064e-2, 3.429_057_134_102_985e-2, 3.378_255_148_275_753e-2, 3.322_892_676_813_277e-2, 3.263_044_456_464_217_5e-2, 3.198_791_279_530_468e-2, 3.130_219_884_802_087e-2, 3.057_422_840_464_999_6e-2, 2.980_498_419_139_589e-2, 2.899_550_465_219_015e-2, 2.814_688_254_686_507_6e-2, 2.726_026_347_601_116_4e-2, 2.633_684_433_451_436e-2, 2.537_787_169_586_609e-2, 2.438_464_012_943_568_3e-2, 2.335_849_045_298_989e-2, 2.230_080_792_283_937_3e-2, 2.121_302_036_408_938e-2, 2.009_659_624_357_542e-2, 1.895_304_268_818_284e-2, 1.778_390_345_139_817_3e-2, 1.659_075_683_115_467e-2, 1.537_521_354_238_962_7e-2, 1.413_891_454_840_083_2e-2, 1.288_352_885_649_808_4e-2, 1.161_075_128_670_389_7e-2, 1.032_230_023_052_424_5e-2, 9.019_915_439_993_632e-3, 7.705_355_960_382_757e-3, 6.380_398_587_897_515e-3, 5.046_838_426_924_442_5e-3, 3.706_500_125_759_316_7e-3, 2.361_331_704_285_021e-3, 1.014_971_908_967_743_7e-3];
-    static ODD_W_43: [f64; 43] = [3.588_019_106_018_701_4e-2, 3.581_080_434_383_374e-2, 3.569_525_919_440_943e-2, 3.553_370_454_416_059e-2, 3.532_634_862_941_021e-2, 3.507_345_872_215_153_5e-2, 3.477_536_078_554_782_6e-2, 3.443_243_905_378_224_5e-2, 3.404_513_553_679_937e-2, 3.361_394_945_057_693_4e-2, 3.313_943_657_366_202e-2, 3.262_220_853_080_144_4e-2, 3.206_293_200_458_967e-2, 3.146_232_787_615_076e-2, 3.082_117_029_596_223_6e-2, 3.014_028_568_601_882_4e-2, 2.942_055_167_462_304_8e-2, 2.866_289_596_517_621_7e-2, 2.786_829_514_042_920_8e-2, 2.703_777_340_373_580_6e-2, 2.617_240_125_893_356e-2, 2.527_329_413_055_707_4e-2, 2.434_161_092_616_763_2e-2, 2.337_855_254_266_017e-2, 2.238_536_031_848_547_8e-2, 2.136_331_443_380_253e-2, 2.031_373_226_065_557e-2, 1.923_796_666_535_656e-2, 1.813_740_426_535_425_2e-2, 1.701_346_364_300_153_3e-2, 1.586_759_351_882_631_8e-2, 1.470_127_088_723_984_3e-2, 1.351_599_911_824_565_9e-2, 1.231_330_603_004_803_7e-2, 1.109_474_194_056_072e-2, 9.861_877_713_701_827e-3, 8.616_302_838_488_951e-3, 7.359_623_648_818_063_5e-3, 6.093_462_047_634_872e-3, 4.819_456_238_501_885_5e-3, 3.539_271_655_388_628_4e-3, 2.254_690_753_752_853_3e-3, 9.691_097_381_770_754e-4];
-    static ODD_W_44: [f64; 44] = [3.507_942_401_790_202_5e-2, 3.501_458_416_619_644e-2, 3.490_660_650_856_071e-2, 3.475_562_407_298_142e-2, 3.456_182_286_913_781e-2, 3.432_544_165_923_908_5e-2, 3.404_677_166_387_109e-2, 3.372_615_620_321_457e-2, 3.336_399_027_407_732e-2, 3.296_072_006_326_112e-2, 3.251_684_239_786_320_5e-2, 3.203_290_413_318_958_5e-2, 3.150_950_147_903_429e-2, 3.094_727_926_515_484_4e-2, 3.034_693_014_684_912_8e-2, 2.970_919_375_161_245_8e-2, 2.903_485_576_792_681e-2, 2.832_474_697_730_520_7e-2, 2.757_974_223_078_458_3e-2, 2.680_075_937_112_917_6e-2, 2.598_875_810_207_383_6e-2, 2.514_473_880_600_257e-2, 2.426_974_131_152_234e-2, 2.336_484_361_245_544_7e-2, 2.243_116_053_983_637e-2, 2.146_984_238_856_114e-2, 2.048_207_350_040_027e-2, 1.946_907_080_515_187_4e-2, 1.843_208_232_178_411_6e-2, 1.737_238_562_150_240_2e-2, 1.629_128_625_479_238_4e-2, 1.519_011_614_466_612_3e-2, 1.407_023_194_864_448_3e-2, 1.293_301_339_260_267_8e-2, 1.177_986_158_087_489_2e-2, 1.061_219_728_997_218_7e-2, 9.431_459_260_797_89e-3, 8.239_102_525_389_08e-3, 7.036_596_870_989_114e-3, 5.825_425_788_770_108e-3, 4.607_087_343_463_241_5e-3, 3.383_104_792_407_455_2e-3, 2.155_112_582_219_114e-3, 9.262_871_051_934_728e-4];
-    static ODD_W_45: [f64; 45] = [3.431_359_817_623_14e-2, 3.425_291_647_165_106e-2, 3.415_185_977_541_012_5e-2, 3.401_054_720_622_908e-2, 3.382_914_533_369_793_5e-2, 3.360_786_798_193_575_5e-2, 3.334_697_597_754_984e-2, 3.304_677_684_219_179e-2, 3.270_762_443_007_278e-2, 3.232_991_851_086_54e-2, 3.191_410_429_848_369_5e-2, 3.146_067_192_629_709e-2, 3.097_015_586_939_654_4e-2, 3.044_313_431_459_439_6e-2, 2.988_022_847_890_037_6e-2, 2.928_210_187_727_748e-2, 2.864_945_954_054_102_4e-2, 2.798_304_718_432_316_5e-2, 2.728_365_033_008_298e-2, 2.655_209_337_919_890_7e-2, 2.578_923_864_123_601_5e-2, 2.499_598_531_753_496e-2, 2.417_326_844_132_288e-2, 2.332_205_777_559_880_4e-2, 2.244_335_667_009_737_5e-2, 2.153_820_087_868_566_6e-2, 2.060_765_733_859_846e-2, 1.965_282_291_296_914_5e-2, 1.867_482_309_816_812_7e-2, 1.767_481_069_752_190_4e-2, 1.665_396_446_306_124e-2, 1.561_348_770_705_006e-2, 1.455_460_688_520_869_6e-2, 1.347_857_015_383_098e-2, 1.238_664_590_355_674_3e-2, 1.128_012_127_376_968_3e-2, 1.016_030_065_441_547_7e-2, 9.028_504_189_234_487e-3, 7.886_066_314_628_902e-3, 6.734_334_432_268_884_6e-3, 5.574_668_047_479_789e-3, 4.408_439_747_302_676_4e-3, 3.237_045_507_972_105e-3, 2.061_987_122_032_229_7e-3, 8.862_412_406_694_142e-4];
-    static ODD_W_46: [f64; 46] = [3.358_047_670_273_291e-2, 3.352_360_509_236_69e-2, 3.342_889_041_048_296e-2, 3.329_643_957_561_579e-2, 3.312_640_210_470_322_4e-2, 3.291_896_994_430_459e-2, 3.267_437_725_392_241_5e-2, 3.239_290_014_167_229_4e-2, 3.207_485_635_259_922e-2, 3.172_060_490_999_231e-2, 3.133_054_571_010_280_4e-2, 3.090_511_907_072_293_6e-2, 3.044_480_523_413_531e-2, 2.995_012_382_499_392_3e-2, 2.942_163_326_374_897_7e-2, 2.885_993_013_627_770_7e-2, 2.826_564_852_043_306_6e-2, 2.763_945_927_027_072e-2, 2.698_206_925_876_273_4e-2, 2.629_422_057_985_327_3e-2, 2.557_668_971_075_783_8e-2, 2.483_028_663_545_258_3e-2, 2.405_585_393_034_465_8e-2, 2.325_426_581_315_775_3e-2, 2.242_642_715_610_957_2e-2, 2.157_327_246_449_981_8e-2, 2.069_576_482_186_873_5e-2, 1.979_489_480_292_793e-2, 1.887_167_935_550_803_5e-2, 1.792_716_065_281_371_4e-2, 1.696_240_491_732_901e-2, 1.597_850_121_778_211_8e-2, 1.497_656_024_067_188_2e-2, 1.395_771_303_800_797e-2, 1.292_310_975_318_535_1e-2, 1.187_391_832_744_712_6e-2, 1.081_132_319_054_249e-2, 9.736_523_941_887_688e-3, 8.650_734_035_428_648e-3, 7.555_179_500_769_821e-3, 6.451_097_794_311_276e-3, 5.339_737_098_169_215e-3, 4.222_357_382_406_608e-3, 3.100_240_403_099_316_7e-3, 1.974_768_768_686_808_2e-3, 8.487_371_680_679_111e-4];
-    static ODD_W_47: [f64; 47] = [3.287_800_959_763_195e-2, 3.282_463_569_369_918_6e-2, 3.273_574_336_068_393e-2, 3.261_142_878_598_215e-2, 3.245_182_648_620_326e-2, 3.225_710_916_161_441_5e-2, 3.202_748_750_926_77e-2, 3.176_320_999_501_228e-2, 3.146_456_258_463_84e-2, 3.113_186_843_444_399_8e-2, 3.076_548_754_155_891_3e-2, 3.036_581_635_440_506_8e-2, 2.993_328_734_371_411_2e-2, 2.946_836_853_456_688_2e-2, 2.897_156_299_996_101_3e-2, 2.844_340_831_645_486_2e-2, 2.788_447_598_247_691_5e-2, 2.729_537_079_993_022_2e-2, 2.667_673_021_976_135_4e-2, 2.602_922_365_220_227_2e-2, 2.535_355_174_243_201_4e-2, 2.465_044_561_244_262e-2, 2.392_066_606_993_061e-2, 2.316_500_278_507_139e-2, 2.238_427_343_606_939_2e-2, 2.157_932_282_441_14e-2, 2.075_102_196_078_49e-2, 1.990_026_712_265_721e-2, 1.902_797_888_454_570_5e-2, 1.813_510_112_204_514_6e-2, 1.722_259_999_071_698_6e-2, 1.629_146_288_099_104_4e-2, 1.534_269_735_028_835_6e-2, 1.437_733_003_365_908_1e-2, 1.339_640_553_436_828_6e-2, 1.240_098_529_611_606e-2, 1.139_214_645_908_584_5e-2, 1.037_098_070_311_609_7e-2, 9.338_593_083_876_396e-3, 8.296_100_874_530_99e-3, 7.244_632_443_933_2e-3, 6.185_326_261_033_324e-3, 5.119_330_329_927_718e-3, 4.047_803_316_371_759_5e-3, 2.971_924_240_818_191e-3, 1.892_968_377_922_935_9e-3, 8.135_642_494_541_165e-4];
-    static ODD_W_48: [f64; 48] = [3.220_431_459_661_350_6e-2, 3.215_415_737_958_550_5e-2, 3.207_061_987_527_28e-2, 3.195_378_880_670_864e-2, 3.180_378_546_007_149e-2, 3.162_076_555_877_401_5e-2, 3.140_491_910_180_172e-2, 3.115_647_016_646_904_2e-2, 3.087_567_667_579_765_5e-2, 3.056_283_013_075_858_4e-2, 3.021_825_530_765_601_5e-2, 2.984_230_992_096_703e-2, 2.943_538_425_198_732_2e-2, 2.899_790_074_366_843_2e-2, 2.853_031_356_206_718_6e-2, 2.803_310_812_486_267_7e-2, 2.750_680_059_743_034_3e-2, 2.695_193_735_699_644e-2, 2.636_909_442_542_935e-2, 2.575_887_687_125_678_5e-2, 2.512_191_818_153_004_8e-2, 2.445_887_960_418_785e-2, 2.377_044_946_160_306_7e-2, 2.305_734_243_602_599_5e-2, 2.232_029_882_766_713e-2, 2.156_008_378_619_171_7e-2, 2.077_748_651_642_657e-2, 1.997_331_945_910_804_7e-2, 1.914_841_744_752_813e-2, 1.830_363_684_096_414e-2, 1.743_985_463_580_780_3e-2, 1.655_796_755_534_245_6e-2, 1.565_889_111_915_692e-2, 1.474_355_869_323_695e-2, 1.381_292_052_185_304_3e-2, 1.286_794_274_249_338_6e-2, 1.190_960_638_533_075_7e-2, 1.093_890_635_919_594_9e-2, 9.956_850_427_084_045e-3, 8.964_458_176_698e-3, 7.962_759_997_865_495e-3, 6.952_796_096_469_405e-3, 5.935_615_630_788_223e-3, 4.912_276_262_166_028e-3, 3.883_845_329_489_294_3e-3, 2.851_409_243_213_055_7e-3, 1.816_146_398_210_039_6e-3, 7.805_332_219_425_613e-4];
-    static ODD_W_49: [f64; 49] = [3.155_766_036_791_123e-2, 3.151_046_648_162_835e-2, 3.143_186_227_722_155e-2, 3.132_192_610_907_518e-2, 3.118_076_756_395_816e-2, 3.100_852_735_178_559_5e-2, 3.080_537_716_535_628e-2, 3.057_151_950_920_578e-2, 3.030_718_749_774_580_6e-2, 3.001_264_462_289_103_5e-2, 2.968_818_449_140_51e-2, 2.933_413_053_222_750_5e-2, 2.895_083_567_407_331e-2, 2.853_868_199_362_695e-2, 2.809_808_033_468_091_2e-2, 2.762_946_989_859_901e-2, 2.713_331_780_651_255e-2, 2.661_011_863_368_585_2e-2, 2.606_039_391_651_548_3e-2, 2.548_469_163_265_475_5e-2, 2.488_358_565_478_194_7e-2, 2.425_767_517_855_707_8e-2, 2.360_758_412_533_789_4e-2, 2.293_396_052_025_105_4e-2, 2.223_747_584_623_937_3e-2, 2.151_882_437_473_022_5e-2, 2.077_872_247_359_421_2e-2, 2.001_790_789_308_656_4e-2, 1.923_713_903_048_718_6e-2, 1.843_719_417_417_85e-2, 1.761_887_072_792_438e-2, 1.678_298_441_613_870_8e-2, 1.593_036_847_096_085e-2, 1.506_187_280_199_023_3e-2, 1.417_836_314_957_944_5e-2, 1.328_072_022_265_728_4e-2, 1.236_983_882_217_516_2e-2, 1.144_662_695_149_825_3e-2, 1.051_200_491_552_474_5e-2, 9.566_904_411_326_136e-3, 8.612_267_615_478_888e-3, 7.649_046_279_335_258e-3, 6.678_200_860_575_098e-3, 5.700_699_773_395_927e-3, 4.717_519_037_520_83e-3, 3.729_643_487_243_035e-3, 2.738_075_873_626_878_3e-3, 1.743_906_958_219_245e-3, 7.494_736_467_374_053e-4];
-
-    /// The weights of the nodes for odd order
-    pub static ODD_WEIGHTS: [&[f64]; 49] = [&ODD_W_1, &ODD_W_2, &ODD_W_3, &ODD_W_4, &ODD_W_5, &ODD_W_6, &ODD_W_7, &ODD_W_8, &ODD_W_9, &ODD_W_10, &ODD_W_11, &ODD_W_12, &ODD_W_13, &ODD_W_14, &ODD_W_15, &ODD_W_16, &ODD_W_17, &ODD_W_18, &ODD_W_19, &ODD_W_20, &ODD_W_21, &ODD_W_22, &ODD_W_23, &ODD_W_24, &ODD_W_25, &ODD_W_26, &ODD_W_27, &ODD_W_28, &ODD_W_29, &ODD_W_30, &ODD_W_31, &ODD_W_32, &ODD_W_33, &ODD_W_34, &ODD_W_35, &ODD_W_36, &ODD_W_37, &ODD_W_38, &ODD_W_39, &ODD_W_40, &ODD_W_41, &ODD_W_42, &ODD_W_43, &ODD_W_44, &ODD_W_45, &ODD_W_46, &ODD_W_47, &ODD_W_48, &ODD_W_49];
-    /// The weights of the nodes for even order
-    pub static EVEN_WEIGHTS: [&[f64]; 50] = [&EVEN_W_1, &EVEN_W_2, &EVEN_W_3, &EVEN_W_4, &EVEN_W_5, &EVEN_W_6, &EVEN_W_7, &EVEN_W_8, &EVEN_W_9, &EVEN_W_10, &EVEN_W_11, &EVEN_W_12, &EVEN_W_13, &EVEN_W_14, &EVEN_W_15, &EVEN_W_16, &EVEN_W_17, &EVEN_W_18, &EVEN_W_19, &EVEN_W_20, &EVEN_W_21, &EVEN_W_22, &EVEN_W_23, &EVEN_W_24, &EVEN_W_25, &EVEN_W_26, &EVEN_W_27, &EVEN_W_28, &EVEN_W_29, &EVEN_W_30, &EVEN_W_31, &EVEN_W_32, &EVEN_W_33, &EVEN_W_34, &EVEN_W_35, &EVEN_W_36, &EVEN_W_37, &EVEN_W_38, &EVEN_W_39, &EVEN_W_40, &EVEN_W_41, &EVEN_W_42, &EVEN_W_43, &EVEN_W_44, &EVEN_W_45, &EVEN_W_46, &EVEN_W_47, &EVEN_W_48, &EVEN_W_49, &EVEN_W_50];
-
-    /// The factor in front of the direct product representation
-    pub static CL: [f64; 101] = [1.0, 1.0, -5e-1, -1.5, 3.75e-1, 1.875, -3.125e-1, -2.187_5, 2.734_375e-1, 2.460_937_5, -2.460_937_5e-1, -2.707_031_25, 2.255_859_375e-1, 2.932_617_187_5, -2.094_726_562_5e-1, -3.142_089_843_75, 1.963_806_152_343_75e-1, 3.338_470_458_984_375, -1.854_705_810_546_875e-1, -3.523_941_040_039_062_5, 1.761_970_520_019_531_3e-1, 3.700_138_092_041_015_6, -1.681_880_950_927_734_4e-1, -3.868_326_187_133_789, 1.611_802_577_972_412e-1, 4.029_506_444_931_03, -1.549_810_171_127_319_3e-1, -4.184_487_462_043_762, 1.494_459_807_872_772_2e-1, 4.333_933_442_831_039, -1.444_644_480_943_679_8e-1, -4.478_397_890_925_407, 1.399_499_340_914_189_8e-1, 4.618_347_825_016_826, -1.358_337_595_593_184_2e-1, -4.754_181_584_576_145, 1.320_605_995_715_595_8e-1, 4.886_242_184_147_704, -1.285_853_206_354_659e-1, -5.014_827_504_783_17, 1.253_706_876_195_792_6e-1, 5.140_198_192_402_749_6, -1.223_856_712_476_845_1e-1, -5.262_583_863_650_434, 1.196_041_787_193_280_5e-1, 5.382_188_042_369_762, -1.170_040_878_776_035_2e-1, -5.499_192_130_247_366, 1.145_665_027_134_867_8e-1, 5.613_758_632_960_852, -1.122_751_726_592_170_5e-1, -5.726_033_805_620_069_5, 1.101_160_347_234_628_7e-1, 5.836_149_840_343_532, -1.080_768_488_952_506e-1, -5.944_226_689_238_783, 1.061_469_051_649_782_7e-1, 6.050_373_594_403_761, -1.043_167_861_104_096_8e-1, -6.154_690_380_514_171, 1.025_781_730_085_695_1e-1, 6.257_268_553_522_74, -1.009_236_863_471_409_8e-1, -6.358_192_239_869_881, 9.934_675_374_796_689e-2, 6.457_538_993_617_848, -9.784_149_990_330_074e-2, -6.555_380_493_521_149, 9.640_265_431_648_748e-2, 6.651_783_147_837_636, -9.502_547_354_053_767e-2, -6.746_808_621_378_174, 9.370_567_529_691_909e-2, 6.840_514_296_675_093, -9.243_938_238_750_125e-2, -6.932_953_679_062_594, 9.122_307_472_450_782e-2, 7.024_176_753_787_103, -9.005_354_812_547_567e-2, -7.114_230_301_912_578, 8.892_787_877_390_723e-2, 7.203_158_180_686_485_5, -8.784_339_244_739_617e-2, -7.291_001_573_133_881, 8.679_763_777_540_335e-2, 7.377_799_210_909_285, -8.578_836_291_754_982e-2, -7.463_587_573_826_834_5, 8.481_349_515_712_312e-2, 7.548_401_068_983_957, -8.387_112_298_871_065e-2, -7.632_272_191_972_668, 8.295_948_034_752_901e-2, 7.715_231_672_320_197, -8.207_693_268_425_742e-2, -7.797_308_605_004_455, 8.122_196_463_546_306e-2, 7.878_530_569_639_918, -8.039_316_907_795_835e-2, -7.958_923_738_717_876, 7.958_923_738_717_877e-2];
 }
